### 2022-12-24 #5029 @editorialbot
[PRE REVIEW]: Sciris: Simplifying scientific software in Python
**Submitting author:** @cliffckerr (<a href="http://orcid.org/0000-0003-2517-2354">Cliff Kerr</a>)
**Repository:** https://github.com/sciris/sciris
**Branch with paper.md** (empty if default branch): joss
**Version:** 2.1.0
