### 2022-12-24 #5029 @danielskatz
👋 @cliffckerr - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and note that you can click on the error above to find out more about it. Please use the command `@editorialbot generate pdf` after making changes to the .md file to make a new PDF. editorialbot commands need to be the first entry in a new comment.

The problem here is
> Problem with ORCID (Sherrie Kelly) for Sherrie Kelly. ORCID looks malformed

which is occurring because the ORCID line has her name and not her ORCID. Please either change this to her ORCID, or if she does not have one, remove this line.
