### 2022-12-24 #5029 @cliffckerr
Thanks @danielskatz -- I see the compile error and will address it shortly. I'm surprised though because the GitHub Action ran fine and produced the `paper.zip` artifact:
https://github.com/sciris/sciris/actions/runs/3772305917

I've also been compiling locally without problem. I guess the build system here differs slightly from the GHA (or perhaps it's been updated since we cloned).
