+++ paper.bib
@@ -6,7 +6,8 @@
   number={7},
   pages={e1009149},
   year={2021},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pcbi.1009149}
 }
 
 
@@ -14,7 +15,8 @@
   title={Python vs. the pandemic: a case study in high-stakes software development},
   author={Kerr, Cliff C and Stuart, Robyn M and Mistry, Dina and Abeysuriya, Romesh G and Cohen, Jamie A and George, Lauren and Jastrzebski, Micha{\l} and Famulare, Michael and Wenger, Edward and Klein, Daniel J},
   year={2022},
-  journal={Proceedings of the 21st Python in Science Conference (SciPy 2022)}
+  journal={Proceedings of the 21st Python in Science Conference (SciPy 2022)},
+  doi={10.25080/majora-212e5952-00e}
 }
 
 @article{kerr2015optima,
@@ -38,7 +40,8 @@
   year =   {2020},
   editor =   {Zhao, Feng and Benedikt, C and Wilson, David},
   address =  {Washington, DC},
-  pages =        {291}
+  pages =        {291},
+  doi = {10.1596/978-1-4648-1523-2_ch13}
 }
 
 
@@ -48,7 +51,8 @@
   journal={Gates Open Research},
   volume={3},
   year={2019},
-  publisher={Gates Foundation-Open Access}
+  publisher={Gates Foundation-Open Access},
+  doi={10.12688/gatesopenres.13031.2}
 }
 
 
@@ -70,7 +74,8 @@
   number={1},
   pages={1--12},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/s12889-018-5294-z}
 }
 
 
@@ -82,7 +87,8 @@
   number={9},
   pages={e1009255},
   year={2021},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pcbi.1009255}
 }
 
 
@@ -103,7 +109,8 @@
   number={11},
   pages={e0260247},
   year={2021},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pone.0260247}
 }
 
 @article{nust2020ten,
@@ -114,7 +121,8 @@
   number={11},
   pages={e1008316},
   year={2020},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pcbi.1008316}
 }
 
 
@@ -126,7 +134,8 @@
   number={1},
   pages={ftz006},
   year={2019},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/femspd/ftz006}
 }
 
 
@@ -138,7 +147,8 @@
   number={1},
   pages={111--116},
   year={2019},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41592-018-0235-4}
 }
 
 
@@ -160,7 +170,8 @@
   number={23},
   pages={4121--4123},
   year={2018},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/bty407}
 }
 
 
@@ -171,7 +182,8 @@
   volume={6},
   number={60},
[…]
