### 2023-01-13 #5029 @aflaxman
Are you still looking for additional reviewers?  I can take this on.

From: Daniel S. Katz ***@***.***>
Sent: Friday, January 6, 2023 11:05 AM
Cc: Abraham Flaxman ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: Sciris: Simplifying scientific software in Python (Issue #5029)

👋 @aflaxman<https://urldefense.com/v3/__https:/github.com/aflaxman__;!!K-Hz7m0Vt54!knzz7eFBU3FRVdkQtMuYyK4LZv51KK0uYayhRhRBiP6gMEfz3EoSJNopn6bcLe37J6XRoplObJgpUdoRjyVw$> and @jnothman<https://urldefense.com/v3/__https:/github.com/jnothman__;!!K-Hz7m0Vt54!knzz7eFBU3FRVdkQtMuYyK4LZv51KK0uYayhRhRBiP6gMEfz3EoSJNopn6bcLe37J6XRoplObJgpUdDhJS0g$> - would one of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html<https://urldefense.com/v3/__https:/joss.readthedocs.io/en/latest/review_criteria.html__;!!K-Hz7m0Vt54!knzz7eFBU3FRVdkQtMuYyK4LZv51KK0uYayhRhRBiP6gMEfz3EoSJNopn6bcLe37J6XRoplObJgpUbjzq8V2$>
