### 2023-01-13 #5076 @editorialbot
[REVIEW]: Sciris: Simplifying scientific software in Python
**Submitting author:** @cliffckerr (<a href="http://orcid.org/0000-0003-2517-2354">Cliff Kerr</a>)
**Repository:** https://github.com/sciris/sciris
**Branch with paper.md** (empty if default branch): joss
**Version:** 3.1.0
