### 2023-01-25 #5076 @danielskatz
👋 @aflaxman - can you please generate your review checklist, and at least check off the COI item, just to ensure that the system is working for you?

To do this, you should use the command `@editorialbot generate my checklist`. @editorialbot commands need to be the first thing in a new comment.
