### 2023-01-25 #5076 @aflaxman
Based on [your COI document](https://github.com/openjournals/joss/blob/main/COI.md), I think I should disclose that I receive research funding from the Bill and Melinda Gates Foundation, which is the current employer of some of the authors, including submitting author, Cliff Kerr.  Although Cliff and I have met periodically through our connections with this institution over the last five years, we do not work together directly on anything, and none of the authors are in a role where they make decisions about my funding.  Therefore, I think it would be fine for me to be a reviewer for this paper, but it seems worth having the editors confirm this.
