### 2023-03-02 #5076 @aflaxman
This paper reports on a package called `sciris` that is dedicated to streamlining a certain type of scientific inquiry in Python.  It strikes me as a valuable contribution and eventually worthy of publication in JOSS. Before I can confirm all of the items on the checklist, however, I think that the authors should address a number of points in their documentation and software paper.

**Documentation** No statement of need, or statement of target audience. No list of dependencies (although `pip install sciris` handles them automatically for me).  No example usage for the sciris package (and possibly confusing examples of a related scirisweb package.  Extremely minimal documentation of core functionality, basically just formatted docstrings with no guidance about what is important or how it fits together.  Lack of documentation on testing, although it seems like there are automated tests.

**Software Paper**  Statement of need does not identify who the target audience is, or its relation to other work. No State-of-the-field section to describe how this software compares to other commonly-used packages.

I will now attempt to create issues in the target repository with more detailed feedback on the software paper.
