### 2023-03-02 sciris/sciris#411 @aflaxman
Unclear who precisely is the target audience
https://github.com/sciris/sciris/blob/cf6ca0c6805c32595fed3140411391a780380bd4/paper.md?plain=1#L92

I would like a clear statement of who the target audience of this package is by this point.  See openjournals/joss-reviews#5076 for more detail.
