### 2023-03-02 sciris/sciris#413 @aflaxman
Reduction is not 5-fold
https://github.com/sciris/sciris/blob/cf6ca0c6805c32595fed3140411391a780380bd4/paper.md?plain=1#L103

Going from 55 lines to 29 lines is not a five-fold reduction... it's not even twofold.

/cc https://github.com/openjournals/joss-reviews/issues/5076
