### 2023-03-02 sciris/sciris#414 @aflaxman
use the same colormap?
https://github.com/sciris/sciris/blob/cf6ca0c6805c32595fed3140411391a780380bd4/paper.md?plain=1#L105

I suggest you use the same colormap.  You don't want to distract people with this additional affordance.

/cc https://github.com/openjournals/joss-reviews/issues/5076
