### 2023-03-02 sciris/sciris#415 @aflaxman
Use >>> instead of > for code blocks
https://github.com/sciris/sciris/blob/cf6ca0c6805c32595fed3140411391a780380bd4/paper.md?plain=1#L131-L135

If you use `>>>` instead of `>` to start your code examples, the reader will be able to copy and paste directly into a Jupyter Notebook or python REPL to try out your code.

/cc https://github.com/openjournals/joss-reviews/issues/5076
