### 2023-03-27 sciris/sciris#411 @tacaswell
I very much agree with this concern.  "Simple" is very much in the eye of the beholder: the more the API encodes your personal assumptions the simpler it seems.  However baking more assumptions into an API is not a cost-free activity as now someone who has different assumptions will find the tool much harder to use.

In addition to a discussion of who the intended audience is (and the assumptions they want to make), I would also like to see a discussion of the path for users who grow needs beyond the reduced interface that sciris provides.
