### 2023-03-27 sciris/sciris#414 @tacaswell
Particularly as if you register the color map with Matplotlib you can use it by name.
