### 2023-03-28 sciris/sciris#426 @tacaswell
broken links to API docs
All of the API links in https://sciris.readthedocs.io/en/latest/overview.html#features are broken.

xref https://github.com/openjournals/joss-reviews/issues/5076
