### 2023-03-28 #5076 @tacaswell
The helloworlod webapp does not run

```
(clean) ~/s/p/s/s/e/helloworld develop|✓
✔ 17:38:46 $ python app.py
Traceback (most recent call last):
  File "/home/tcaswell/source/p/sciris/scirisweb/examples/helloworld/app.py", line 16, in <module>
    import scirisweb as sw
  File "/home/tcaswell/source/p/sciris/scirisweb/scirisweb/__init__.py", line 17, in <module>
    from .sw_app       import * # analysis:ignore
  File "/home/tcaswell/source/p/sciris/scirisweb/scirisweb/sw_app.py", line 27, in <module>
    from werkzeug.serving import run_with_reloader
ImportError: cannot import name 'run_with_reloader' from 'werkzeug.serving' (/tmp/clean/lib/python3.10/site-packages/werkzeug/serving.py)
```

I get this error both with scirisweb installed from pip and from the develop branch of scirisweb (following instructions from https://github.com/sciris/sciris#20-second-quick-start-guide-for-scirisweb).  I guess this is out of scope for the paper, but reading through the README it was one of the first things I tried.

-----

Installation

- The linux install section assumes a debian (apt -based) derived OS (https://sciris.readthedocs.io/en/latest/overview.html#installing-on-linux steps 1 and 2)
- `python setup.py develop` is discouraged (https://sciris.readthedocs.io/en/latest/overview.html#installing-on-linux step 6), earlier in the docs `pip install -e .` is used for editable installs
- Agree `pip install` works, but could not find a list of dependencies in the docs.

-----

Docs

- all of the API links in https://sciris.readthedocs.io/en/latest/overview.html#features are broken. [made issue]
- I could not find copy-paste-runable examples (but there is a decent amount of exapmles in the docstrings)
- the scirisweb docs do not use http (not really in the review but clicking on the link in the readme gets the "NOT SECURE" warning which is less than great)
- the documentation of `sc_fileio` does not include any mention of the dangers of using pickle as an on-disk format (https://docs.python.org/3/library/pickle.html)

On the plus side, I think the the API docs are acceptable complete

----

Performance:

 - What aspect of performance are `odict`'s optimizing?  I would expect some of the divergences from a normal `dict` to hurt performance (e.g. making `items` return a list and then `iteritems` is documented as a alias for `items`, I would expect both of those to be iterators and not make a copy).
 - The paper claims that sciris makes writing scientific Python faster, is there anything to quantify that?

-----

State-of-field comparison

- I would like to see a comparison between `odict` and `pandas.Series`.  They look very similar to me (ordered, key (aka) index based lookup, positional based lookup, index based slicing, and (implicit in odict) homogeneous types)
- How does the datetime handling in sciris compare to the datetime support from pandas?
- I would like to see a comparison with joblib and dask for `sc_parallel`
- maybe also `tqdm` and `rich` for the progress bar and formatting tools?

-----

General concerns

I am not convinced by the generality or ease of use claims that the paper makes.  First, I have issue with the implied generality of sciris.  There is a significant amount of scientific software that is not covered by the helper functions here.  I agree with @aflaxman that the audience for this work needs to be better specified.  Second, I the claims of "simple", "easy", "pleasent", and "dejargonifitation" are highly subjective (and I disagree with them in most cases!).  Software seems simple when the assumptions in the code matches the assumptions in the user head, connecting back to the first concern I think identifying what domain (I assume epidemiology based on the examples of usage) this is aimed at would go a long way to ameliorating my concern.  I also am not convinced by the claims that this will improve the long term maintainability of code.  There are many more people who can read "plain" numpy / pandas code and you can google for numpy support.  There is a trade off between short-term brevity and the long-term pool of people who can easily read and adapt the code.

I think the paper needs a discussion of the learning curve when users need to move beyond `sciris`, either because they need to do something it does not support yet or because they begin to need to make different assumption.
