### 2023-03-29 sciris/sciris#429 @cliffckerr
Clarify installation
cc https://github.com/openjournals/joss-reviews/issues/5076
