### 2023-03-29 sciris/sciris#430 @cliffckerr
Clarify performance
cc https://github.com/openjournals/joss-reviews/issues/5076
