### 2023-03-29 sciris/sciris#431 @cliffckerr
Clarify statement of need
cc https://github.com/openjournals/joss-reviews/issues/5076
