### 2023-03-29 sciris/sciris#432 @cliffckerr
Clarify example usage
cc https://github.com/openjournals/joss-reviews/issues/5076
