### 2023-03-29 sciris/sciris#433 @cliffckerr
Improve summary
cc https://github.com/openjournals/joss-reviews/issues/5076
