### 2023-03-29 sciris/sciris#434 @cliffckerr
Clarify state of field
cc https://github.com/openjournals/joss-reviews/issues/5076
