### 2023-03-29 sciris/sciris#435 @cliffckerr
Clarify functionality documentation
cc https://github.com/openjournals/joss-reviews/issues/5076
