### 2023-03-29 sciris/sciris#436 @cliffckerr
Clarify automated tests
cc https://github.com/openjournals/joss-reviews/issues/5076
