### 2023-03-29 sciris/sciris#413 @cliffckerr
Clarified in the caption: "Comparison of functionally identical scripts without Sciris (left) and with Sciris (right), showing a nearly five-fold reduction in lines of code required **(excluding whitespace, comments, and the shared "wave generator" code), from 33 to 7**".

This could instead be illustrated by removing comments and whitespace and putting the `randwave` function in a separate script, but I feel the version here is clearer.
