### 2023-03-29 48ff2e3
fix parallelize docstring
