### 2023-03-29 54221e1
update showcase script and readme
