### 2023-03-29 sciris/sciris#414 @cliffckerr
I felt there wasn't value in having two identical panels side-by-side. Happy to remove the color map and the RHS panel if you feel strongly, but I personally like it as a (fairly subtle) way of illustrating a Sciris feature (namely, the new color maps). I have however fixed the Sciris call -- the `plotkwargs` dictionary call was unnecessary, so the new line is just
`sc.surf3d(data[:], cmap='orangeblue')`.

See https://github.com/sciris/sciris/pull/446
