### 2023-03-29 sciris/sciris#437 @cliffckerr
cc https://github.com/openjournals/joss-reviews/issues/5076
