### 2023-03-29 sciris/sciris#415 @cliffckerr
Perhaps Daniel can clarify, but something about the Markdown format and JOSS template doesn't seem to render `>>>` correctly:
![image](https://user-images.githubusercontent.com/3239256/228424364-36410294-6b1b-457d-b055-c5cd922f8a61.png)

Looking through other papers though, it looks like they don't use anything -- so I've just removed them entirely, and then it should still work to copy-paste the code.
