### 2023-03-29 sciris/sciris#436 @cliffckerr
There is a tests folder with instructions for running the tests:
https://github.com/sciris/sciris/tree/main/tests
