### 2023-03-30 sciris/sciris#431 @cliffckerr
Added to readme: "It was originally written to help epidemiologists and neuroscientists focus on doing science, rather than on writing coding, but Sciris is applicable across scientific domains." PR:
https://github.com/sciris/sciris/pull/448
