### 2023-03-30 sciris/sciris#411 @cliffckerr
Added to the first paragraph of the "Statement of need":
> For some scientists, this increasing reliance on software has created a barrier between themselves and the science they want to do. It is these people – people who want things to "just work" rather than worry about the implementation details – who are the primary audience for Sciris. (In contrast, people who always use type hints in their code will likely *not* find Sciris to be as helpful.)

Working on part 2...
