+++ README.rst
@@ -7,7 +7,7 @@ Welcome to Sciris
 .. image:: https://static.pepy.tech/personalized-badge/sciris?period=total&units=international_system&left_color=black&right_color=blue&left_text=Downloads
  :target: https://pepy.tech/project/sciris
 
-.. image:: https://github.com/sciris/sciris/actions/workflows/tests.yml/badge.svg
+.. image:: https://github.com/sciris/sciris/actions/workflows/tests.yaml/badge.svg
  :target: https://github.com/sciris/sciris
 
 
