+++ README.rst
@@ -4,11 +4,11 @@ Welcome to Sciris
 .. image:: https://badgen.net/pypi/v/sciris/
  :target: https://pypi.com/project/sciris
 
-.. image:: https://static.pepy.tech/personalized-badge/sciris?period=total&units=international_system&left_color=black&right_color=blue&left_text=Downloads
- :target: https://pepy.tech/project/sciris
-
 .. image:: https://github.com/sciris/sciris/actions/workflows/tests.yaml/badge.svg
- :target: https://github.com/sciris/sciris
+ :target: https://github.com/sciris/sciris/actions/workflows/tests.yaml?query=workflow
+
+ .. image:: https://static.pepy.tech/personalized-badge/sciris?period=total&units=international_system&left_color=black&right_color=blue&left_text=Downloads
+ :target: https://pepy.tech/project/sciris
 
 
 What is Sciris?
