### 2023-04-03 sciris/sciris#411 @cliffckerr
Added a "Beyond Sciris" section
