### 2023-04-03 sciris/sciris#433 @cliffckerr
I've made various additions to the paper that hopefully clarify this -- but happy to revisit if still unclear.
