+++ README.rst
@@ -7,7 +7,7 @@ Welcome to Sciris
 .. image:: https://static.pepy.tech/personalized-badge/sciris?period=total&units=international_system&left_color=grey&right_color=yellow&left_text=Downloads
  :target: https://pepy.tech/project/sciris
 
- .. image:: https://img.shields.io/pypi/l/sciris.svg
+.. image:: https://img.shields.io/pypi/l/sciris.svg
  :target: https://github.com/sciris/sciris/blob/main/LICENSE
 
 .. image:: https://github.com/sciris/sciris/actions/workflows/tests.yaml/badge.svg
