### 2023-04-03 sciris/sciris#426 @cliffckerr
I think the issue here is that `README.rst` is doing dual-service as the docs page and, well, as the readme. Here, they weren't supposed to be API links, but I agree it's confusing! In this context, I couldn't get the Sphinx autosummary commands (e.g. `:func:findnearest`) to work (especially not in the `sc.name_of_function()` style that I think is clearest), and I'd rather not hard-code links from the rST file to the built docs pages. I've created an issue here: https://github.com/sciris/sciris/issues/455, but it will be quite involved to rebuild all the docs.
