+++ sciris/sc_parallel.py
@@ -28,7 +28,7 @@ __all__ = ['parallelize', 'parallelcmd', 'parallel_progress']
 
 
 def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncpus=None, maxcpu=None, maxmem=None,
-                interval=None, parallelizer=None, serial=False, returnpool=False, die=False, **func_kwargs):
+                interval=None, parallelizer=None, serial=False, returnpool=False, die=True, **func_kwargs):
     '''
     Execute a function in parallel.
 
@@ -66,10 +66,10 @@ def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncp
         maxcpu       (float)     : maximum CPU load; otherwise, delay the start of the next process (not used if ``ncpus`` is specified)
         maxmem       (float)     : maximum fraction of virtual memory (RAM); otherwise, delay the start of the next process
         interval     (float)     : number of seconds to pause between starting processes for checking load
-        parallelizer (str/func)  : parallelization function; default 'multiprocess' (other choices are 'concurrent.futures', 'multiprocessing', or user-supplied; see example below)
-        serial       (bool)      : whether to skip parallelization run in serial (useful for debugging)
+        parallelizer (str/func)  : parallelization function; default 'multiprocess' (see below for details)
+        serial       (bool)      : whether to skip parallelization and run in serial (useful for debugging; equivalent to ``parallelizer='serial'``)
         returnpool   (bool)      : whether to return the process pool as well as the results
-        die          (bool)      : whether to stop immediately if an exception is encountered (otherwise, try 'multiprocess')
+        die          (bool)      : whether to stop immediately if an exception is encountered (otherwise, store the exception as the result)
         func_kwargs  (dict)      : merged with kwargs (see above)
 
     Returns:
@@ -129,6 +129,32 @@ def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncp
         results = sc.parallelize(f, iterkwargs=dict(x=[1,2,3], y=[4,5,6]), parallelizer=pool.map) # Note: parallelizer is pool.map, not pool
 
 
+    **Example 6 -- using Sciris as an interface to Dask**::
+    
+        def f(x,y):
+            return [x]*y
+    
+        def dask_map(task, argslist):
+            import dask
+            queued = [dask.delayed(task)(args) for args in argslist]
+            return list(dask.compute(*queued))
+    
+        results = sc.parallelize(f, iterkwargs=dict(x=[1,2,3], y=[4,5,6]), parallelizer=dask_map)
+
+
+    The ``parallelizer`` argument allows a wide range of different parallelizers
+    (including different aliases for each), and also supports user-supplied ones.
+    Note that in most cases, the default parallelizer will suffice. However, the
+    full list of options is:
+        
+        - ``None``, ``'default'``, ``'robust'``, ``'multiprocess'``: the slow but robust dill-based parallelizer ``multiprocess``
+        - ``'fast'``, ``'concurrent'``, ``'concurrent.futures'``: the faster but more fragile pickle-based Python-default parallelizer ``concurrent.futures``
+        - ``'multiprocessing'``: the previous pickle-based Python default parallelizer, ``multiprocessing``
+        - ``'serial'``, ``'serial-nocopy'``: no parallelization (single-threaded); with "-nocopy", do not force pickling
+        - ``'thread'``', ``'threadpool'``', ``'thread-nocopy'``': thread- rather than process-based parallelization ("-nocopy" as above)
[…]
