+++ sciris/sc_parallel.py
@@ -27,8 +27,9 @@ from . import sc_profiling as scp
 __all__ = ['parallelize', 'parallelcmd', 'parallel_progress']
 
 
-def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncpus=None, maxcpu=None, maxmem=None,
-                interval=None, parallelizer=None, serial=False, returnpool=False, die=True, **func_kwargs):
+def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncpus=None, 
+                maxcpu=None, maxmem=None, interval=None, parallelizer=None, serial=False, 
+                returnpool=False, progress=False, callback=None, die=True, **func_kwargs):
     '''
     Execute a function in parallel.
 
@@ -69,6 +70,8 @@ def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncp
         parallelizer (str/func)  : parallelization function; default 'multiprocess' (see below for details)
         serial       (bool)      : whether to skip parallelization and run in serial (useful for debugging; equivalent to ``parallelizer='serial'``)
         returnpool   (bool)      : whether to return the process pool as well as the results
+        progress     (bool)      : whether to show a progress bar
+        callback     (func)      : an optional function to call from each worker
         die          (bool)      : whether to stop immediately if an exception is encountered (otherwise, store the exception as the result)
         func_kwargs  (dict)      : merged with kwargs (see above)
 
@@ -247,9 +250,10 @@ def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncp
             else:  # pragma: no cover # Should be caught by previous checking, so shouldn't happen
                 errormsg = f'iterkwargs type not understood ({type(iterkwargs)})'
                 raise TypeError(errormsg)
-        taskargs = TaskArgs(func=func, index=index, iterval=iterval, iterdict=iterdict,
+        taskargs = TaskArgs(func=func, index=index, niters=niters, iterval=iterval, iterdict=iterdict,
                             args=args, kwargs=kwargs, maxcpu=maxcpu, maxmem=maxmem,
-                            interval=interval, embarrassing=embarrassing, die=die)
+                            interval=interval, embarrassing=embarrassing, callback=callback,
+                            die=die)
         argslist.append(taskargs)
     
     # Set up the run
@@ -300,7 +304,7 @@ def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncp
         
         elif pname == 'multiprocess': # Main use case
             with mp.Pool(processes=ncpus) as pool:
-                outputlist = pool.map(_parallel_task, argslist)
+                outputlist = list(pool.imap(_parallel_task, argslist))
         
         elif pname == 'multiprocessing':
             with mpi.Pool(processes=ncpus) as pool:
@@ -446,7 +450,7 @@ def parallel_progress(fcn, inputs, num_workers=None, show_progress=True, initial
         errormsg = 'Module tqdm not found; please install with "pip install tqdm"'
         raise ModuleNotFoundError(errormsg) from E
 
-    pool = mp.pool.Pool(num_workers, initializer=initializer)
+    pool = mp.Pool(num_workers, initializer=initializer)
 
     results = [None]
     if scu.isnumber(inputs):
@@ -487,9 +491,11 @@ class TaskArgs(scu.prettyobj):
         A class to hold the arguments for the parallel task -- not to be invoked by the user.
 
         Arguments and ordering must match both ``sc.parallelize()`` and ``sc._parallel_task()`` '''
-        def __init__(self, func, index, iterval, iterdict, args, kwargs, maxcpu, maxmem, interval, embarrassing, die=True):
[…]
