+++ tests/test_parallel.py
@@ -112,9 +112,9 @@ def test_components():
 
     args = [0]*11
     args[0] = lambda: None
-    args[3] = None # Set iterdict to None
-    args[4] = None # Set args to empty list
-    args[5] = None # Set kwargs to empty dict
+    args[4] = None # Set iterdict to None
+    args[5] = None # Set args to empty list
+    args[6] = None # Set kwargs to empty dict
     args[10] = True # Set embarrassing
     taskargs = sc.sc_parallel.TaskArgs(*args)
     task = sc.sc_parallel._parallel_task(taskargs)
