### 2023-04-06 sciris/sciris#437 @cliffckerr
- I've tried to clarify the statement of need ("Sciris provides functions covering a wide range of common math, file I/O, and plotting operations. This means you can get more done with less code, and spend less time looking things up on StackOverflow.") and the target audience ("It was originally written to help epidemiologists and neuroscientists focus on doing science, rather than on writing coding, but Sciris is applicable across scientific domains."), but if either of these are unclear still, happy to edit.
- The dependencies are listed in `setup.py`, which I think is a sufficiently "discoverable" place for them, but again happy to amend (e.g. move to a separate `requirements.txt` file) if desired.
- It's a good point about minimal usage examples beyond docstrings. I had long thought that having tutorials was excessive since most of the functions are quite atomic and relatively simple, but I no longer think that :) Have made an issue here: https://github.com/sciris/sciris/issues/465. This will hopefully address the example usage and functionality points.
- For tests, I'm not sure what other documentation could be provided other than the tests folder, which includes a readme? I could mention this in the main readme ("The tests are in the tests folder"), but this seems redundant...
- I don't know of any *directly* comparable libraries to Sciris, but I do mention other libraries that have similar motivations (e.g., seaborn to make Matplotlib more accessible). I also added a reference to R's `tidyverse`, which I think is similar in intent (if very different in implementation) to Sciris.
