### 2023-04-06 #5076 @cliffckerr
Hi @danielskatz -- I think I've addressed all the review comments. I've found the "branch of main software repo" structure challenging to work with, so have been working on the paper in its own repo:
https://github.com/sciris/sciris_paper

The latest version of the paper is available at paper.sciris.org (direct link [here](https://github.com/sciris/sciris_paper/blob/main/paper_current.pdf)).

All of the reviewer comments that (I believe) have been addressed have been moved to the "Paper revision review" milestone:
https://github.com/sciris/sciris/milestone/19

I was thinking that once issues are resolved to the reviewers' satisfaction, they could close them. (Or, of course, add comments if they still have concerns.)

In addition, as part of the review process, a large number of needed improvements were suggested (which was extremely helpful!) or arose, which will comprise Sciris v2.2.0. These are listed here:
https://github.com/sciris/sciris/milestone/17

I'll work through addressing them, but they might take me a while to get through. I'm hoping the JOSS review process can continue while I work on these in parallel.
