### 2023-04-07 d873e79
working on refactoring functions
