+++ sciris/sc_fileio.py
@@ -49,7 +49,76 @@ zstd = scu.importbyname('zstandard', die=False, verbose=False) # Optional import
 #%% Pickling functions
 ##############################################################################
 
-__all__ = ['load', 'save', 'loadobj', 'saveobj', 'zsave', 'loadstr', 'dumpstr', 'rmpath']
+__all__ = ['load', 'save', 'loadobj', 'saveobj', 'zsave', 'loadstr', 'dumpstr']
+
+
+# Define common error messages
+def _gziperror(filename):
+    ''' Base error message when a file couldn't be opened '''
+    return f'''
+Unable to load
+    {filename}
+as either a gzipped, zstandard or regular pickle file. Ensure that it is actually a pickle file.
+'''
+
+def _unpicklingerror(filename):
+    ''' Base error message when a file couldn't be loaded '''
+    return f'''
+Unable to load file: "{filename}"
+as a gzipped pickle file. Loading pickles can fail if Python modules have changed
+since the object was saved. If you are loading a custom class that has failed,
+you can use the remapping argument; see the sc.load() docstring for details.
+Otherwise, you might need to revert to the previous version of that module (e.g.
+in a virtual environment), save in a format other than pickle, reload in a different
+environment with the new version of that module, and then re-save as a pickle.
+
+For general information on unpickling errors, see e.g.:
+https://wiki.python.org/moin/UsingPickle
+https://stackoverflow.com/questions/41554738/how-to-load-an-old-pickle-file
+'''
+
+
+def _load_filestr(filename, folder):
+    ''' Try different options for loading a file on disk into a string -- not for external use '''
+    
+    # Handle loading of either filename or file object
+    if isinstance(filename, Path):
+        filename = str(filename)
+    if scu.isstring(filename):
+        argtype = 'filename'
+        filename = makefilepath(filename=filename, folder=folder, makedirs=False) # If it is a file, validate the folder (but don't create one if it's missing)
+    elif isinstance(filename, io.BytesIO):
+        argtype = 'fileobj'
+    else: # pragma: no cover
+        errormsg = f'First argument to sc.load() must be a string or file object, not {type(filename)}; see also sc.loadstr()'
+        raise TypeError(errormsg)
+    fileargs = {'mode': 'rb', argtype: filename}
+    
+    try:
+        with gz.GzipFile(**fileargs) as fileobj:
+            filestr = fileobj.read() # Convert it to a string
+    except Exception as E: # pragma: no cover
+        exc = type(E) # Figure out what kind of error it is
+        if exc == FileNotFoundError: # This is simple, just raise directly
+            raise E
+        elif exc == gz.BadGzipFile:
+            try: # If the gzip file failed, first try as a zstd compressed object
+                with open(filename, 'rb') as fh:
+                    zdcompressor = zstd.ZstdDecompressor()
+                    with zdcompressor.stream_reader(fh) as fileobj:
+                        filestr = fileobj.read()
+            except Exception as E2: # If that fails...
+                try: # Try as a regular binary object
+                    with open(filename, 'rb') as fileobj:
+                        filestr = fileobj.read() # Convert it to a string
+                except Exception as E3:
+                    try: # And finally as a regular object
+                        with open(filename, 'r') as fileobj:
+                            filestr = fileobj.read() # Convert it to a string
+                    except Exception as E4:
+                        gziperror = _gziperror(filename) + f'\nAdditional errors encountered:\n{str(E2)}\n{str(E3)}\n{str(E4)}'
+                        raise exc(gziperror) from E
+    return filestr
 
 
[…]
