### 2023-04-07 4cf6774
improved sort docstring
