+++ sciris/sc_profiling.py
@@ -26,6 +26,7 @@ from . import sc_datetime as scd
 from . import sc_odict as sco
 from . import sc_fileio as scf
 from . import sc_nested as scn
+from . import sc_dataframe as scdf
 
 
 ##############################################################################
@@ -70,7 +71,8 @@ def memload():
 
 
 
-def checkmem(var, descend=True, alphabetical=False, compresslevel=0, plot=False, verbose=False, **kwargs):
+def checkmem(var, descend=1, alphabetical=False, compresslevel=0, maxitems=1000, 
+             plot=False, verbose=False, **kwargs):
     '''
     Checks how much memory the variable or variables in question use by dumping
     them to file.
@@ -86,6 +88,7 @@ def checkmem(var, descend=True, alphabetical=False, compresslevel=0, plot=False,
         descend (bool): whether or not to descend one level into the object
         alphabetical (bool): if descending into a dict or object, whether to list items by name rather than size
         compresslevel (int): level of compression to use when saving to file (typically 0)
+        maxitems (int): the maximum number of separate entries to check the size of
         plot (bool): if descending, show the results as a pie chart
         verbose (bool or int): detail to print, if >1, print repr of objects along the way
         **kwargs (dict): passed to :func:`load`
@@ -95,7 +98,14 @@ def checkmem(var, descend=True, alphabetical=False, compresslevel=0, plot=False,
         import numpy as np
         import sciris as sc
         sc.checkmem(['spiffy', np.random.rand(2483,589)])
+    
+    New in version 2.2.0: descend multiple levels; dataframe output
     '''
+    
+    # Handle input arguments -- used for recursion
+    _depth  = kwargs.pop('_depth', 0)
+    _prefix = kwargs.pop('_prefix', '')
+    _join   = kwargs.pop('_join', '→')
 
     def check_one_object(variable):
         ''' Check the size of one variable '''
@@ -105,7 +115,7 @@ def checkmem(var, descend=True, alphabetical=False, compresslevel=0, plot=False,
 
         # Create a temporary file, save the object, check the size, remove it
         filename = tempfile.mktemp()
-        scf.save(filename, variable, die='never', compresslevel=compresslevel)
+        scf.save(filename, variable, allow_empty=True, compresslevel=compresslevel)
         filesize = os.path.getsize(filename)
         os.remove(filename)
 
@@ -125,49 +135,56 @@ def checkmem(var, descend=True, alphabetical=False, compresslevel=0, plot=False,
     varnames  = []
     variables = []
     sizes     = []
-    sizestrs  = []
+    df = scdf.dataframe(columns=['variable', 'bytesize', 'human_readable'])
 
     # Create the object(s) to check the size(s) of
-    varnames = ['Variable'] # Set defaults
-    variables = [var]
-    if descend or descend is None:
-        if hasattr(var, '__dict__'): # It's an object
-            if verbose>1: print('Iterating over object')
+    if not descend:
+        varname = _prefix if _prefix else 'Variable'
+        bytesize, sizestr = check_one_object(var)
+        df.appendrow(dict(variable=varname, bytesize=bytesize, human_readable=sizestr))
+    else:
+        if isinstance(var, dict): # Handle dicts
+            if verbose>1: print('Iterating over dict')
+            varnames = list(var.keys())
+            variables = var.values()
+        elif np.iterable(var): # Handle dicts and lists
+            if verbose>1: print('Iterating over list-like object')
+            varnames = [f'item {i}' for i in range(len(var))]
+            variables = var
+        elif hasattr(var, '__dict__'): # It's an object
+            if verbose>1: print('Iterating over class-like object')
             varnames = sorted(list(var.__dict__.keys()))
[…]
