### 2023-04-07 7a559fc
recursive checkmem is working
