+++ sciris/sc_dataframe.py
@@ -72,6 +72,13 @@ class dataframe(pd.DataFrame):
                 raise ValueError(errormsg)
             dtypes = columns # Already in the right format
             columns = list(columns.keys())
+        if isinstance(data, dict):
+            if columns is not None:
+                colset = sorted(set(columns))
+                dataset = sorted(set(data.keys()))
+                if colset != dataset:
+                    errormsg = f'Incompatible column names provided:\nColumns: {colset}\n   Data: {dataset}'
+                    raise ValueError(errormsg)
 
         # Create the dataframe
         super().__init__(data=data, columns=columns, **kwargs)
@@ -90,12 +97,15 @@ class dataframe(pd.DataFrame):
 
 
     def set_dtypes(self, dtypes):
-        ''' Set dtypes in-place (see pandas ``df.astype()`` for the user-facing version) '''
+        '''
+        Set dtypes in-place (see pandas ``df.astype()`` for the user-facing version)
+        
+        New in version 2.2.0.
+        '''
         if not isinstance(dtypes, dict):
             dtypes = {col:dtype for col,dtype in zip(self.columns, dtypes)}
         for col,dtype in dtypes.items(): # NB: "self.astype(dtypes, copy=False)" does not modify in place
             self[col] = self[col].astype(dtype)
-        self.astype(dtypes, copy=False)
         return
 
 
@@ -146,7 +156,7 @@ class dataframe(pd.DataFrame):
     def _sanitizecol(self, col, die=True):
         ''' Take None or a string and return the index of the column '''
         if col is None:
-            output = 0 # If not supplied, assume first column is control
+            output = 0 # If not supplied, assume first column is intended
         elif scu.isstring(col):
             try:
                 output = self.cols.index(col) # Convert to index
@@ -320,7 +330,11 @@ class dataframe(pd.DataFrame):
 
 
     def poprow(self, key, returnval=True):
-        ''' Remove a row from the data frame '''
+        '''
+        Remove a row from the data frame.
+        
+        To pop a column, see ``df.pop()``.        
+        '''
         rowindex = int(key)
         thisrow = self.iloc[rowindex,:]
         self.drop(rowindex, inplace=True)
@@ -328,7 +342,7 @@ class dataframe(pd.DataFrame):
         else:         return
 
 
-    def replacedata(self, newdata=None, newdf=None, reset_index=True, inplace=True):
+    def replacedata(self, newdata=None, newdf=None, reset_index=True, inplace=True, keep_dtypes=True):
         '''
         Replace data in the dataframe with other data
 
@@ -337,41 +351,87 @@ class dataframe(pd.DataFrame):
             newdf (dataframe): substitute the current dataframe with this one
             reset_index (bool): update the index
             inplace (bool): whether to modify in-place
+            keep_dtypes (bool): whether to keep original dtypes
+        
+        New in version 2.2.0: "keep_dtypes" argument
         '''
+        dtypes = self.dtypes if keep_dtypes else None
         if newdf is None:
-            newdf = dataframe(data=newdata, columns=self.columns)
+            newdf = dataframe(data=newdata, columns=self.columns, dtypes=dtypes)
         if reset_index:
             newdf.reset_index(drop=True, inplace=True)
         if inplace:
             self.__dict__ = newdf.__dict__ # Hack to copy in-place
+            if keep_dtypes:
+                self.set_dtypes(dtypes)
             return self
         else:
             return newdf
 
 
-    def appendrow(self, value, reset_index=True, inplace=True):
+    def appendrow(self, row, reset_index=True, inplace=True):
         '''
-        Add a row to the end of the dataframe. See also ``concat()`` and ``insertrow()``.
+        Add row(s) to the end of the dataframe. 
+        
+        See also ``concat()`` and ``insertrow()``. Similar to the pandas operation
[…]
