+++ sciris/sc_dataframe.py
@@ -20,7 +20,7 @@ __all__ = ['dataframe']
 class dataframe(pd.DataFrame):
     '''
     An extension of the pandas dataframe with additional convenience methods for
-    accessing rows and columns and performing other operations.
+    accessing rows and columns and performing other operations, such as adding rows.
     
     Args:
         data (dict/array/dataframe): the data to use
@@ -31,7 +31,7 @@ class dataframe(pd.DataFrame):
 
     **Examples**::
 
-        a = sc.dataframe(cols=['x','y'],data=[[1238,2],[384,5],[666,7]]) # Create data frame
+        a = sc.dataframe(cols=['x','y'], data=[[1238,2],[384,5],[666,7]]) # Create data frame
         a['x'] # Print out a column
         a[0] # Print out a row
         a['x',0] # Print out an element
@@ -109,70 +109,84 @@ class dataframe(pd.DataFrame):
         return
 
 
-    def _to_array(self, arr):
-        ''' Try to conver to the current data type, or else use an object '''
-        try: # Try to use current type
-            output = np.array(arr, dtype=self.values.dtype)
-        except: # This is to e.g. not force conversion to strings # pragma: no cover
-            output = np.array(arr, dtype=object)
-        return output
+    # def _to_array(self, arr):
+    #     ''' Try to conver to the current data type, or else use an object '''
+    #     try: # Try to use current type
+    #         output = np.array(arr, dtype=self.values.dtype)
+    #     except: # This is to e.g. not force conversion to strings # pragma: no cover
+    #         output = np.array(arr, dtype=object)
+    #     return output
 
 
-    def _val2row(self, value=None, to2d=True):
-        ''' Convert a list, array, or dictionary to the right format for appending to a dataframe '''
-        if isinstance(value, dict):
-            output = np.zeros(self.ncols, dtype=object)
-            for c,col in enumerate(self.cols):
-                try:
-                    output[c] = value[col]
-                except:
-                    errormsg = f'Entry for column {col} not found; keys you supplied are: {scu.strjoin(value.keys())}'
-                    raise Exception(errormsg)
-        elif value is None:
-            output = np.empty(self.ncols)
-        else: # Not sure what it is, just make it an array
-            output = self._to_array(value)
+    # def _val2row(self, value=None, to2d=True):
+    #     ''' Convert a list, array, or dictionary to the right format for appending to a dataframe '''
+    #     if isinstance(value, dict):
+    #         output = np.zeros(self.ncols, dtype=object)
+    #         for c,col in enumerate(self.cols):
+    #             try:
+    #                 output[c] = value[col]
+    #             except:
+    #                 errormsg = f'Entry for column {col} not found; keys you supplied are: {scu.strjoin(value.keys())}'
+    #                 raise Exception(errormsg)
+    #     elif value is None:
+    #         output = np.empty(self.ncols)
+    #     else: # Not sure what it is, just make it an array
+    #         output = self._to_array(value)
 
-        # Validation
-        if output.ndim == 1: # Convert from 1D to 2D
-            shape = output.shape[0]
-            if to2d:
-                output = np.array([output])
-        else:
-            shape = output.shape[1]
-        if shape != self.ncols:
-            errormsg = f'Row has wrong length ({shape} supplied, {self.ncols} expected)'
-            raise IndexError(errormsg)
+    #     # Validation
+    #     if output.ndim == 1: # Convert from 1D to 2D
+    #         shape = output.shape[0]
+    #         if to2d:
+    #             output = np.array([output])
+    #     else:
+    #         shape = output.shape[1]
+    #     if shape != self.ncols:
+    #         errormsg = f'Row has wrong length ({shape} supplied, {self.ncols} expected)'
[…]
