### 2023-04-08 67ad12b
initial pass of refactor
