### 2023-04-08 5953201
fixed pandas methods
