### 2023-04-08 37e547b
simplify dtype checking
