### 2023-04-08 8281dbc
simplify dataframe constructor
