### 2023-04-08 38f1ec9
working on printmean
