+++ sciris/sc_printing.py
@@ -322,7 +322,7 @@ def indent(prefix=None, text=None, suffix='\n', n=0, pretty=False, width=70, **k
 
 #%% Data representation functions
 
-__all__ += ['sigfig', 'humanize_bytes', 'printarr', 'printdata', 'printvars']
+__all__ += ['sigfig', 'printmean', 'humanize_bytes', 'printarr', 'printdata', 'printvars']
 
 
 def sigfig(x, sigfigs=4, SI=False, sep=False, keepints=False):
@@ -393,6 +393,73 @@ def sigfig(x, sigfigs=4, SI=False, sep=False, keepints=False):
         return output
     else:
         return output[0]
+
+
+def _printstats(which, arr, stds=2, quantiles=None, val_sf=None, err_sf=None, doprint=True, **kwargs):
+    ''' Helper function -- see ``sc.printmean()`` and ``sc.printstats()`` '''
+    vsf = val_sf
+    esf = err_sf if err_sf is not None else 2
+    arr = scu.toarray(arr)
+    ismean = (which == 'mean')
+    if ismean:
+        val = arr.mean()
+        err = arr.std()*stds
+    else:
+        if quantiles is None:
+            quantiles = 95
+        elif quantiles == 'iqr':
+            quantiles = 50
+        if isinstance(quantiles, int):
+            x = quantiles/100/2
+            quantiles = [0.5-x, 0.5+x]
+        else:
+            errormsg = f'Could not understand quantiles "{quantiles}"'
+            raise ValueError(errormsg)
+                
+        val = np.quantile(arr, 0.5)
+        err = np.qunatile(arr, quantiles)
+    
+    relsize = np.floor(np.log10(abs(mean))) - np.floor(np.log10(abs(std)))
+    if vsf is None:
+        vsf = esf + relsize
+    elif msf is not None and err_sigfigs is None:
+        esf = min(msf, msf - relsize)
+        
+    meanstr = sigfig(mean, msf)
+    stdstr  = sigfig(std, esf)
+    string = f'{meanstr} ± {stdstr}'
+    
+    if doprint:
+        print(string)
+    else:
+        return string
+    
+    
+
+
+def printmean(arr, stds=2, mean_sf=None, err_sf=None, doprint=True, **kwargs):
+    '''
+    Quickly print the mean and standard deviation of an array.
+    
+    By default, will calculate the correct number of significant figures based on
+    the deviation.
+    
+    Args:
+        arr (array): the data to summarize
+        stds (int): the number of multiples of the standard deviation to show (default 2; can also use 1)
+        mean_sf (int): if provided, use this rather than the auto-calculated number of significant figures
+        err_sf (int): ditto, but for the error (standard deviation)
+        doprint (bool): whether to print (else, return the string)
+        kwargs (dict): passed to ``sc.sigfig()``
+    
+    **Example**::
+        
+        data = [1210, 1072, 1722, 1229, 1902]
+        sc.printmean(data) # Returns 1430 ± 320
+    
+    New in version 2.2.0.
+    '''
+    return _printstats(which='mean', stds=stds, mean_sf=mean_sf, err_sf=err_sf, doprint=doprint, **kwargs)
 
 
 def humanize_bytes(bytesize, decimals=3):
