+++ sciris/sc_datetime.py
@@ -413,7 +413,7 @@ def daydiff(*args):
     return output
 
 
-def daterange(start_date=None, end_date=None, interval=None, inclusive=True, as_date=False,
+def daterange(start_date=None, end_date=None, interval=None, inclusive=True, as_date=None,
               readformat=None, outformat=None, **kwargs):
     '''
     Return a list of dates from the start date to the end date. To convert a list
@@ -424,13 +424,13 @@ def daterange(start_date=None, end_date=None, interval=None, inclusive=True, as_
 
     Args:
         start_date (int/str/date) : the starting date, in any format
-        end_date   (int/str/date) : the end date, in any format
+        end_date   (int/str/date) : the end date, in any format (see also kwargs below)
         interval   (int/str/dict) : if an int, the number of days; if 'week', 'month', or 'year', one of those; if a dict, passed to ``dt.relativedelta()``
         inclusive  (bool)         : if True (default), return to end_date inclusive; otherwise, stop the day before
-        as_date    (bool)         : if True, return a list of datetime.date objects instead of strings (note: you can also use "asdate" instead of "as_date")
-        readformat (str)          : passed to date()
-        outformat  (str)          : passed to date()
-        days       (int)          : optional
+        as_date    (bool)         : if True, return a list of ``datetime.date`` objects; else, as input type (e.g. strings; note: you can also use "asdate" instead of "as_date")
+        readformat (str)          : passed to ``sc.date()``
+        outformat  (str)          : passed to ``sc.date()``
+        kwargs     (dict)         : optionally, use any valid argument to ``sc.datedelta()`` to create the end_date
 
     **Examples**::
 
@@ -441,12 +441,15 @@ def daterange(start_date=None, end_date=None, interval=None, inclusive=True, as_
     | New in version 1.0.0.
     | New in version 1.3.0: "interval" argument
     | New in version 2.0.0: ``sc.datedelta()`` arguments
+    | New in version 2.2.0: preserve input type
     '''
 
     # Handle inputs
     start_date = kwargs.pop('startdate', start_date) # Handle with or without underscore
     end_date   = kwargs.pop('enddate',   end_date) # Handle with or without underscore
     as_date    = kwargs.pop('asdate', as_date) # Handle with or without underscore
+    if as_date is None: # Typical case, return the same format as the input
+        as_date = False if isinstance(start_date, str) else True
     if len(kwargs):
         end_date = datedelta(start_date, **kwargs)
     start_date = date(start_date, readformat=readformat)
@@ -466,7 +469,7 @@ def daterange(start_date=None, end_date=None, interval=None, inclusive=True, as_
     while curr_date < end_date:
         dates.append(curr_date)
         curr_date += delta
-
+    
     # Convert to final format
     dates = date(dates, start_date=start_date, as_date=as_date, outformat=outformat)
     return dates
