+++ sciris/sc_fileio.py
@@ -31,6 +31,7 @@ import pandas as pd
 import datetime as dt
 import gzip as gz
 import pickle as pkl
+import zstandard as zstd
 from zipfile import ZipFile
 from contextlib import closing
 from pathlib import Path
@@ -42,7 +43,6 @@ from . import sc_printing as scp
 from . import sc_datetime as scd
 from . import sc_odict as sco
 from . import sc_dataframe as scdf
-zstd = scu.importbyname('zstandard', die=False, verbose=False) # Optional import
 
 
 ##############################################################################
@@ -118,6 +118,8 @@ def _load_filestr(filename, folder):
                     except Exception as E4:
                         gziperror = _gziperror(filename) + f'\nAdditional errors encountered:\n{str(E2)}\n{str(E3)}\n{str(E4)}'
                         raise exc(gziperror) from E
+        else:
+            raise E # TODO: combine with FileNotFoundError?
     return filestr
 
 
@@ -339,6 +341,89 @@ def dumpstr(obj=None):
     return result
 
 
+##############################################################################
+#%% Versioned pickling functions
+##############################################################################
+
+__all__ += ['getmetadata', 'saveversioned', 'loadversioned']
+
+
+
+def getmetadata(paths=True, caller=True, git=True, pip=True, frame=2, comments=None):
+    
+    # Additional imports
+    import sys
+    from .sc_version import __version__
+    import matplotlib as mpl
+    
+    # Store key metadata
+    md = dict()
+    md['date']       = scd.now()
+    md['platform']   = scu.getplatform()
+    md['executable'] = sys.executable if paths else None # NB, will likely include username info
+    md['comments']   = comments
+    
+    # Store key version info
+    md['python_version']     = sys.version
+    md['sciris_version']     = __version__
+    md['numpy_version']      = np.__version__
+    md['pandas_version']     = pd.__version__
+    md['matplotlib_version'] = mpl.__version__
+    
+    # Store optional metadata
+    if caller and paths:
+        md['called_by'] = scu.getcaller(frame=frame, tostring=False)
+    if git:
+        md['git'] = scu.gitinfo()
+    if pip:
+        md['pip'] = scu.freeze()
+    
+    return md
+
+
+# Default filenames for 
+_mdata_fn = 'metadata.json'
+_data_fn  = 'data.obj'
+
+def saveversioned(filename, data, paths=True, caller=True, git=True, pip=True, 
+                  mdata_fn=_mdata_fn,  data_fn=_data_fn, comments=None, **kwargs):
+
+    # Get the metadata
+    metadata = getmetadata(paths=paths, caller=caller, git=git, pip=pip, comments=comments, frame=3)
+    
+    # Convert both to strings
+    metadatastr = jsonify(metadata, tostring=True, indent=2)
+    datastr     = dumpstr(data)
+    
+    # Construct output
+    datadict = dict(mdata_fn=metadatastr, data_fn=datastr)
+    
+    return savezip(filename=filename, data=datadict, tobytes=False, **kwargs)
+    
+
+def loadversioned(filename, folder=None, return_metadata=False, mdata_fn=_mdata_fn, data_fn=_data_fn, **kwargs):
+    filename = makefilepath(filename=filename, folder=folder)
+   
+    with ZipFile(filename, 'r') as zf: # Create the zip file
+        
+        # Read in the strings
+        try:
+            metadatastr = zf.read(mdata_fn)
+            datastr     = zf.read(data_fn)
+        except Exception as E:
+            errormsg = f'Could not load metadata file "{mdata_fn}" and/or data file "{data_fn}": are you sure this is a Sciris-versioned data zipfile?'
+            raise FileNotFoundError(errormsg) from E
+        
+        # Convert into Python objects
[…]
