+++ sciris/sc_utils.py
@@ -21,14 +21,10 @@ Highlights:
 #%% Imports
 ##############################################################################
 
-import os
 import re
 import sys
 import copy
-import time
 import json
-import zlib
-import types
 import string
 import numbers
 import pprint
@@ -41,7 +37,6 @@ import numpy as np
 import pandas as pd
 import random as rnd
 import uuid as py_uuid
-import packaging.version
 import contextlib as cl
 import traceback as py_traceback
 
@@ -56,8 +51,8 @@ _booltypes   = (bool, np.bool_)
 ##############################################################################
 
 # Define the modules being loaded
-__all__ = ['fast_uuid', 'uuid', 'dcp', 'cp', 'pp', 'sha', 'freeze', 'require',
-           'traceback', 'getplatform', 'iswindows', 'islinux', 'ismac', 'asciify']
+__all__ = ['fast_uuid', 'uuid', 'dcp', 'cp', 'pp', 'sha', 'traceback', 
+           'getplatform', 'iswindows', 'islinux', 'ismac', 'asciify']
 
 
 def fast_uuid(which=None, length=None, n=1, secure=False, forcelist=False, safety=1000, recursion=0, recursion_limit=10, verbose=True):
@@ -356,110 +351,6 @@ def sha(obj, encoding='utf-8', digest=False):
     if digest:
         output = output.hexdigest()
     return output
-
-
-def freeze(lower=False):
-    '''
-    Alias for pip freeze.
-
-    Args:
-        lower (bool): convert all keys to lowercase
-
-    **Example**::
-
-        assert 'numpy' in sc.freeze() # One way to check for versions
-
-    New in version 1.2.2.
-    '''
-    import pkg_resources as pkgr # Imported here since slow (>0.1 s)
-    raw = dict(tuple(str(ws).split()) for ws in pkgr.working_set)
-    keys = sorted(raw.keys())
-    if lower:
-        labels = {k:k.lower() for k in keys}
-    else:
-        labels = {k:k for k in keys}
-    data = {labels[k]:raw[k] for k in keys} # Sort alphabetically
-    return data
-
-
-def require(reqs=None, *args, exact=False, detailed=False, die=True, verbose=True, **kwargs):
-    '''
-    Check whether environment requirements are met. Alias to pkg_resources.require().
-
-    Args:
-        reqs (list/dict): a list of strings, or a dict of package names and versions
-        args (list): additional requirements
-        kwargs (dict): additional requirements
-        exact (bool): use '==' instead of '>=' as the default comparison operator if not specified
-        detailed (bool): return a dict of which requirements are/aren't met
-        die (bool): whether to raise an exception if requirements aren't met
-        verbose (bool): print out the exception if it's not being raised
-
-    **Examples**::
-
-        sc.require('numpy')
-        sc.require(numpy='')
-        sc.require(reqs={'numpy':'1.19.1', 'matplotlib':'3.2.2'})
-        sc.require('numpy>=1.19.1', 'matplotlib==3.2.2', die=False)
-        sc.require(numpy='1.19.1', matplotlib='==4.2.2', die=False, detailed=True)
-
-    New in version 1.2.2.
-    '''
-    import pkg_resources as pkgr # Imported here since slow (>0.1 s)
-
-    # Handle inputs
-    reqlist = list(args)
-    reqdict = kwargs
-    if isinstance(reqs, dict):
-        reqdict.update(reqs)
-    else:
-        reqlist = mergelists(reqs, reqlist)
-
-    # Turn into a list of strings
-    comparechars = '<>=!~'
-    for k,v in reqdict.items():
-        if not v:
-            entry = k # If no version is provided, entry is just the module name
-        else:
[…]
