+++ sciris/__init__.py
@@ -23,20 +23,21 @@ _lazy = _os.getenv('SCIRIS_LAZY', False)
 
 # Otherwise, import everything
 if not _lazy:
-    from .sc_version   import *
-    from .sc_utils     import *
-    from .sc_printing  import *
-    from .sc_nested    import *
-    from .sc_odict     import *
-    from .sc_settings  import *
-    from .sc_datetime  import *
-    from .sc_math      import *
-    from .sc_dataframe import *
-    from .sc_fileio    import *
-    from .sc_profiling import *
-    from .sc_parallel  import *
-    from .sc_asd       import *
-    from .sc_plotting  import *
-    from .sc_colors    import *
+    from .sc_version    import *
+    from .sc_utils      import *
+    from .sc_printing   import *
+    from .sc_nested     import *
+    from .sc_odict      import *
+    from .sc_settings   import *
+    from .sc_datetime   import *
+    from .sc_math       import *
+    from .sc_versioning import *
+    from .sc_dataframe  import *
+    from .sc_fileio     import *
+    from .sc_profiling  import *
+    from .sc_parallel   import *
+    from .sc_asd        import *
+    from .sc_plotting   import *
+    from .sc_colors     import *
 
 del _os, _lazy, _threads
