+++ sciris/sc_fileio.py
@@ -43,6 +43,7 @@ from . import sc_printing as scp
 from . import sc_datetime as scd
 from . import sc_odict as sco
 from . import sc_dataframe as scdf
+from . import sc_versioning as scv
 
 
 ##############################################################################
@@ -157,10 +158,8 @@ def load(filename=None, folder=None, verbose=False, die=None, remapping=None, me
     | New in version 1.1.0: "remapping" argument
     | New in version 1.2.2: ability to load non-gzipped pickles; support for dill; arguments passed to loader
     '''
-
     # Load the file
     filestr = _load_filestr(filename, folder)
-    
 
     # Unpickle it
     try:
@@ -235,9 +234,11 @@ def save(filename='default.obj', obj=None, folder=None, compression='gzip', comp
     
     # Handle path
     if filename is None: # If the user explicitly specifies None as the file, create a byte stream instead
-        bytesobj = io.BytesIO()
+        tobytes = True
+        bytestream = io.BytesIO()
     else:
-        bytesobj = None
+        tobytes = False
+        bytestream = None
 
         # Process and sanitize the filename passed in by the user
         filetypes = (str, type(Path()), type(None))
@@ -265,22 +266,31 @@ def save(filename='default.obj', obj=None, folder=None, compression='gzip', comp
 
     # Compress and actually save
     success = False
-    if compression in ['gz', 'gzip']:
+    
+    if compression in ['gz', 'gzip']:  # Main use case
         # File extension is .gz
-        with gz.GzipFile(filename=filename, fileobj=bytesobj, mode='wb', compresslevel=compresslevel) as fileobj:
+        with gz.GzipFile(filename=filename, fileobj=bytestream, mode='wb', compresslevel=compresslevel) as fileobj:
             serialize(fileobj, obj, success, **kwargs)
-    elif compression in ['zst', 'zstd', 'zstandard']:
-        # File extension is .zst
-        with open(filename, 'wb') as fh:
-            zcompressor = zstd.ZstdCompressor(level=compresslevel)
-            with zcompressor.stream_writer(fh) as fileobj:
-                serialize(fileobj, obj, success, **kwargs)
-    elif compression in ['none']:
-        with open(filename, 'wb') as fileobj:
-            serialize(fileobj, obj, success, **kwargs)
+    
     else:
-        errormsg = f"Invalid compression format '{compression}': must be 'gzip', 'zstd', or 'none'"
-        raise ValueError(errormsg)
+        if tobytes:
+            filecontext = closing(bytestream)
+        else:
+            filecontext = open(filename, 'wb')
+        
+        if compression in ['zst', 'zstd', 'zstandard']:
+            # File extension is .zst
+            with filecontext as fh:
+                zcompressor = zstd.ZstdCompressor(level=compresslevel)
+                with zcompressor.stream_writer(fh) as fileobj:
+                    serialize(fileobj, obj, success, **kwargs)
+        elif compression in ['none']:
+            # File extension can be anything
+            with filecontext as fileobj:
+                serialize(fileobj, obj, success, **kwargs)
+        else:
+            errormsg = f"Invalid compression format '{compression}': must be 'gzip', 'zstd', or 'none'"
+            raise ValueError(errormsg)
 
     if verbose and filename:
         print(f'Object saved to "{filename}"')
@@ -288,8 +298,8 @@ def save(filename='default.obj', obj=None, folder=None, compression='gzip', comp
     if filename:
         return filename
     else: # pragma: no cover
-        bytesobj.seek(0)
[…]

[…]
