+++ sciris/sc_fileio.py
@@ -131,12 +131,15 @@ def load(filename=None, folder=None, verbose=False, die=None, remapping=None, me
     Accepts either a filename (standard usage) or a file object as the first argument.
     Note that `:func:`load`/:func:`loadobj` are aliases of each other.
 
-    Note: be careful when loading pickle files, since a malicious pickle can be
-    used to execute arbitrary code.
+    **Note 1**: Since this function relies on pickle, it can potentially execute arbitrary
+    code, so you should only use it with sources you trust. For more information, see:
+        https://docs.python.org/3/library/pickle.html
 
-    When a pickle file is loaded, Python imports any modules that are referenced
+    **Note 2**: When a pickle file is loaded, Python imports any modules that are referenced
     in it. This is a problem if module has been renamed. In this case, you can
-    use the ``remapping`` argument to point to the new modules or classes.
+    use the ``remapping`` argument to point to the new modules or classes. For
+    more robustness, use the :func:`savewithmetadata`/:func:`loadwithmetadata`
+    functions.
 
     Args:
         filename  (str/Path): the filename (or full path) to load
@@ -183,6 +186,15 @@ def save(filename='default.obj', obj=None, folder=None, compression='gzip', comp
     Save an object to file as a gzipped pickle -- use compression 5 by default,
     since more is much slower but not much smaller. Once saved, can be loaded
     with :func:`load`. Note that :func:`save`/:func:`saveobj` are identical.
+    
+    **Note 1**: Since this function relies on pickle, it can potentially execute arbitrary
+    code, so you should only use it with sources you trust. For more information, see:
+        https://docs.python.org/3/library/pickle.html
+    
+    **Note 2**: When a pickle file is loaded, Python imports any modules that are referenced
+    in it. This is a problem if module has been renamed (in which case the pickle
+    usually can't be opened). For more robustness, use the :func:`savewithmetadata`/
+    :func:`loadwithmetadata` functions.
 
     Args:
         filename      (str/Path) : the filename or path to save to; if None, return an io.BytesIO filestream instead of saving to disk
@@ -255,7 +267,7 @@ def save(filename='default.obj', obj=None, folder=None, compression='gzip', comp
                 errormsg = f'Filename type {type(filename)} is not valid: must be one of {filetypes}'
                 raise Exception(errormsg)
 
-        filename = makefilepath(filename=filename, folder=folder, sanitize=sanitizepath)
+        filename = makefilepath(filename=filename, folder=folder, sanitize=sanitizepath, makedirs=True)
 
     # Handle object
     if obj is None: # pragma: no cover
@@ -402,7 +414,7 @@ def savetext(filename=None, string=None, **kwargs):
         string = '\n'.join(string) # Convert from list to string)
     elif not is_array and not scu.isstring(string):
         string = str(string)
-    filename = makefilepath(filename=filename)
+    filename = makefilepath(filename=filename, makedirs=True)
     if is_array: # Shortcut to Numpy for saving arrays -- basic CSV
         kw = scu.mergedicts(dict(fmt='%s', delimiter=', '), kwargs)
         np.savetxt(filename, string, **kw)
@@ -469,7 +481,7 @@ def unzip(filename=None, outfolder='.', folder=None, members=None):
         names = zf.namelist()
         zf.extractall(outfolder, members=members)
     
-    output = [makefilepath(filename=name, folder=outfolder) for name in names]
+    output = [makefilepath(filename=name, folder=outfolder, makedirs=True) for name in names]
     return output
 
 
@@ -501,7 +513,7 @@ def savezip(filename=None, filelist=None, data=None, folder=None, basename=True,
     '''
 
[…]
