+++ sciris/sc_versioning.py
@@ -28,9 +28,9 @@ __all__ = ['freeze', 'require', 'gitinfo', 'compareversions', 'getcaller',
 
 
 # Define shared variables
-_metadataflag = 'sciris_metadata' # Key name used to identify metadata saved in a figure
-_mdata_fn = 'metadata.json' # Default filenames for metadata and data
-_data_fn  = 'data.obj'
+_metadataflag      = 'sciris_metadata' # Key name used to identify metadata saved in a figure
+_metadata_filename = 'metadata.json' # Default filenames for metadata and data
+_obj_filename      = 'sciris_pickle.obj'
 
 
 def freeze(lower=False):
@@ -408,7 +408,7 @@ def metadata(outfile=None, comments=None, pipfreeze=True, user=True, caller=True
     calling_info = dict_fn(getcaller(relframe=relframe+1, tostring=False))
     
     # Store metadata
-    metadata = dict_fn(
+    md = dict_fn(
         timestamp = scd.getdate(),
         user      = scu.getuser() if user else None,
         system = dict_fn(
@@ -430,16 +430,16 @@ def metadata(outfile=None, comments=None, pipfreeze=True, user=True, caller=True
     )
     
     # Store any additional data
-    metadata.update(kwargs)
+    md.update(kwargs)
     
     if outfile is not None:
-        outfile = scf.makefilepath(outfile)
-        scf.savejson(outfile, metadata)
+        outfile = scf.makefilepath(outfile, makedirs=True)
+        scf.savejson(outfile, md)
     
     if tostring:
-        metadata = scf.jsonify(metadata, tostring=True, indent=2)
+        md = scf.jsonify(md, tostring=True, indent=2)
     
-    return metadata
+    return md
 
 
 def loadmetadata(filename, load_all=False, die=True):
@@ -447,8 +447,9 @@ def loadmetadata(filename, load_all=False, die=True):
     Read metadata from a saved image; currently only PNG and SVG are supported.
 
     Only for use with images saved with ``sc.savefig()``. Metadata retrieval for PDF
-    is not currently supported. To load metadata saved with ``sc.getmetadata()``, 
-    use ``sc.loadjson()`` instead.
+    is not currently supported. To load metadata saved with ``sc.metadata()``, 
+    you can also use ``sc.loadjson()`` instead. To load metadata saved with ``sc.savewithmetadata()``,
+    use ``sc.loadwithmetadata()`` instead.
 
     Args:
         filename (str): the name of the file to load the data from
@@ -464,7 +465,7 @@ def loadmetadata(filename, load_all=False, die=True):
     from . import sc_fileio as scf # To avoid circular import
 
     # Initialize
-    metadata = {}
+    md = {}
     lcfn = str(filename).lower() # Lowercase filename
 
     # Handle bitmaps
@@ -482,10 +483,10 @@ def loadmetadata(filename, load_all=False, die=True):
         # Usual case, can find metadata and is PNG
         if is_png and (load_all or _metadataflag in keys):
             if load_all:
-                metadata = im.info
+                md = im.info
             else:
                 jsonstr = im.info[_metadataflag]
-                metadata = scf.loadjson(string=jsonstr)
+                md = scf.loadjson(string=jsonstr)
 
         # JPG -- from https://www.thepythoncode.com/article/extracting-image-metadata-in-python
         elif is_jpg: # pragma: no cover
@@ -496,7 +497,7 @@ def loadmetadata(filename, load_all=False, die=True):
                 data = exifdata.get(tag_id)
                 if isinstance(data, bytes):
                     data = data.decode()
-                metadata[tag] = data
+                md[tag] = data
 
         # Can't find metadata
         else: # pragma: no cover
[…]
