+++ sciris/sc_fileio.py
@@ -125,7 +125,8 @@ def _load_filestr(filename, folder):
     return filestr
 
 
-def load(filename=None, folder=None, verbose=False, die=None, remapping=None, method='pickle', **kwargs):
+def load(filename=None, folder=None, verbose=False, die=None, remapping=None, 
+         method='pickle', auto_remap=True, **kwargs):
     '''
     Load a file that has been saved as a gzipped pickle file, e.g. by :func:`save`.
     Accepts either a filename (standard usage) or a file object as the first argument.
@@ -142,13 +143,14 @@ def load(filename=None, folder=None, verbose=False, die=None, remapping=None, me
     functions.
 
     Args:
-        filename  (str/Path): the filename (or full path) to load
-        folder    (str/Path): the folder
-        verbose   (bool):     print details
-        die       (bool):     whether to raise an exception if errors are encountered (otherwise, load as much as possible)
-        remapping (dict):     way of mapping old/unavailable module names to new
-        method    (str):      method for loading (usually pickle or dill)
-        kwargs    (dict):     passed to ``pickle.loads()``/``dill.loads()``
+        filename   (str/Path): the filename (or full path) to load
+        folder     (str/Path): the folder
+        verbose    (bool):     print details
+        die        (bool):     whether to raise an exception if errors are encountered (otherwise, load as much as possible)
+        remapping  (dict):     way of mapping old/unavailable module names to new
+        method     (str):      method for loading (usually pickle or dill)
+        auto_remap (bool):     whether to use known deprecations to load failed pickles
+        kwargs     (dict):     passed to ``pickle.loads()``/``dill.loads()``
 
     **Examples**::
 
@@ -165,7 +167,8 @@ def load(filename=None, folder=None, verbose=False, die=None, remapping=None, me
 
     # Unpickle it
     try:
-        obj = _unpickler(filestr, filename=filename, verbose=verbose, die=die, remapping=remapping, method=method, **kwargs) # Actually load it
+        kw = dict(filename=filename, verbose=verbose, die=die, remapping=remapping, method=method, auto_remap=auto_remap)
+        obj = _unpickler(filestr, **kw, **kwargs) # Actually load it
     except Exception as E: # pragma: no cover
         exc = type(E) # Figure out what kind of error it is
         errormsg = _unpicklingerror(filename) + '\n\nSee the stack trace above for more information on this specific error.'
@@ -1928,7 +1931,7 @@ class Failed(object):
             output += f'Error: {failure["error"]}\n'
             if verbose:
                 output += '\nTraceback:\n'
-                output += failure['exception']
+                output += f"\n{failure['exception']}"
                 output += '\n\n'
         if tostring:
             return output
@@ -1985,7 +1988,7 @@ class UniversalFailed(Failed): # pragma: no cover
         return
 
     def disp(self, *args, **kwargs):
-        return scu.pr(self, *args, **kwargs)
+        return scp.pr(self, *args, **kwargs)
 
 
 def makefailed(module_name=None, name=None, error=None, exception=None, universal=False):
@@ -2012,19 +2015,19 @@ class _RobustUnpickler(pkl.Unpickler):
         key = f'{module_name}.{name}'
         obj = self.remapping.get(key) # If the user has supplied the module directly
         if obj is None or (isinstance(obj, tuple) and len(obj)==2): # Either it's not in the remapping, or it's a tuple
-            try:
-                if obj is None: # Key not in remapping
-                    load_module = module_name
-                    load_name = name
-                else: # It's a tuple of names, process
-                    load_module = obj[0]
-                    load_name = obj[1]
-                module = importlib.import_module(load_module)
-                obj = getattr(module, load_name)
-            except Exception as E:
-                if verbose: print(f'Unpickling warning: could not import {load_module}.{load_name}: {str(E)}')
[…]
