+++ sciris/sc_version.py
@@ -4,6 +4,6 @@ Version and license information.
 
 __all__ = ['__version__', '__versiondate__', '__license__']
 
-__version__      = '2.1.1'
-__versiondate__  = '2023-03-30'
+__version__      = '2.2.0'
+__versiondate__  = '2023-04-10'
 __license__      = f'Sciris {__version__} ({__versiondate__}) – © 2014-2023 by the Sciris Development Team'
