+++ sciris/sc_printing.py
@@ -1044,16 +1044,16 @@ def percentcomplete(step=None, maxsteps=None, stepsize=1, prefix=None):
     return
 
 
-def progressbar(i, maxiters=None, label='', every=1, length=30, empty='—', full='•', newline=False, **kwargs):
+def progressbar(i=None, maxiters=None, label='', every=1, length=30, empty='—', full='•', newline=False, **kwargs):
     '''
     Show a progress bar for a for loop.
     
     It can be called manually inside each iteration of the loop, or it can be used 
     to wrap the object being iterated. In the latter case, it acts as an alias 
-    for ``tqdm.tqdm()``.
+    for the ``tqdm.tqdm()`` progress bar.
 
     Args:
-        i        (int/iterable): current iteration, or iterable object
+        i        (int/iterable): current iteration (for text output), or iterable object (for tqdm)
         maxiters (int): maximum number of iterations (can also use an object with length)
         label    (str): initial label to print
         every    (int/float): if int, print every "every"th iteration (if 1, print all); if float and <1, print every maxiters*every iteration
@@ -1071,11 +1071,11 @@ def progressbar(i, maxiters=None, label='', every=1, length=30, empty='—', ful
             sc.timedsleep(0.05)
         
         # Direct usage inside a loop with custom formatting
-        for i in range(2000):
-            sc.progressbar(i+1, 2000)
+        for i in range(1000):
+            sc.progressbar(i+1, 1000, every=100, length=10, empty=' ', full='✓', newline=True)
             sc.timedsleep(0.001)
 
-        # Used to wrap an iterable
+        # Used to wrap an iterable, using tqdm
         x = np.arange(100)
         for i in sc.progressbar(x):
             pl.pause(0.01)
@@ -1085,8 +1085,9 @@ def progressbar(i, maxiters=None, label='', every=1, length=30, empty='—', ful
     | New in version 1.3.3: "every" argument
     | New in version 2.2.0: wrapper for tqdm
     '''
-    if scu.isiterable(i):
-        return tqdm(i, **kwargs)
+    if i is None or scu.isiterable(i):
+        desc = kwargs.pop('desc', label)
+        return tqdm.tqdm(i, desc=desc, **kwargs)
     
     # Handle inputs
     if hasattr(maxiters, '__len__'):
