### 2023-04-10 8f9486e
add benchmark, update save docstring, etc
