+++ sciris/sc_odict.py
@@ -87,6 +87,12 @@ class odict(OD):
         return
 
 
+    @classmethod
+    def _new(cls, *args, **kwargs):
+        ''' Constructor for a new instance -- used in place of self.__class__() '''
+        return cls(*args, **kwargs)
+
+
     def _cache_keys(self):
         ''' Store a copy of the keys as a list so integer lookup doesn't have to regenerate it each time '''
         self._setattr('_cached_keys', self.keys())
@@ -121,7 +127,7 @@ class odict(OD):
                         _defaultdict = None
                     if _defaultdict is not None and allow_default: # If it does, use it, then get the key again
                         if _defaultdict == 'nested':
-                            _defaultdict = lambda: self.__class__(defaultdict=_defaultdict) # Make recursive
+                            _defaultdict = lambda: self._new(defaultdict=_defaultdict) # Make recursive
                         dd = _defaultdict() # Create the new object
                         self._setitem(key, dd) # Add it to the dictionary
                         self._setattr('_stale', True) # Flag to refresh the cached keys
@@ -217,7 +223,9 @@ class odict(OD):
         return
 
 
-    def __repr__(self, maxlen=None, showmultilines=True, divider=False, dividerthresh=10, numindents=0, recursionlevel=0, sigfigs=None, numformat=None, maxitems=200, classname='odict()', quote='', numleft='#', numsep=':', keysep=':', dividerchar='—'):
+    def __repr__(self, maxlen=None, showmultilines=True, divider=False, dividerthresh=10, 
+                 numindents=0, recursionlevel=0, sigfigs=None, numformat=None, maxitems=200, 
+                 classname='odict', quote='', numleft='#', numsep=':', keysep=':', dividerchar='—'):
         ''' Print a meaningful representation of the odict '''
 
         # Set non-customizable primitives for display
@@ -238,7 +246,7 @@ class odict(OD):
         halfmax = int(maxitems/2)
         extraitems = 0
         if nkeys == 0:
-            output = classname
+            output = classname + '()'
         else:
             output = ''
             keystrs = [] # Start with an empty list which we'll save key strings in.
@@ -464,9 +472,9 @@ class odict(OD):
         return output
 
 
-    def export(self, doprint=True):
+    def export(self, doprint=True, classname='odict'):
         ''' Export the odict in a form that is valid Python code '''
-        start = 'odict(['
+        start = classname + '(['
         end = '])'
         output = start
 
@@ -619,7 +627,7 @@ class odict(OD):
         if scu.isstring(keys) and pattern is None: # Assume first argument, transfer
             pattern = keys
             keys = None
-        filtered = odict()
+        filtered = self._new() # Create a new instance of the same class
         if keys is None:
             keys = self.findkeys(pattern=pattern, method=method, first=False)
         if not exclude:
@@ -658,11 +666,11 @@ class odict(OD):
 
         **Example**::
 
-            z = odict()
+            z = sc.odict()
             z['foo'] = 1492
             z.insert(1604)
             z.insert(0, 'ganges', 1444)
-            z.insert(2, 'meikang', 1234)
+            z.insert(2, 'mekong', 1234)
         '''
 
         # Handle inputs
@@ -682,7 +690,7 @@ class odict(OD):
             raise ValueError(errormsg)
 
         # Create a temporary dictionary to hold all of the items after the insertion point
-        tmpdict = odict()
+        tmpdict = self._new()
         origkeys = self.keys()
         originds = range(len(origkeys))
[…]
