+++ tests/README.rst
@@ -2,4 +2,14 @@
 Tests
 =====
 
-This folder contains the core tests for Sciris. Recommended usage is ``./check_coverage``. You can also use ``pytest`` to run all the tests in the folder. To install test dependencies, use ``pip install -r requirements_test.txt``.
+This folder contains the core tests for Sciris.
+
+Installation
+------------
+
+To install test dependencies, use ``pip install -r requirements_test.txt``.
+
+Usage
+-----
+
+Recommended usage is ``./check_coverage`` or ``./run_tests``. You can also use ``pytest`` to run all the tests in the folder.
