+++ sciris/sc_legacy.py
@@ -1,20 +1,25 @@
 '''
-Legacy methods for handling old pickles (eg Python 2 pickles). Included for backwards
-compatibility, but not imported into Sciris by default.
+This is the graveyard for old Sciris functions that are no longer used. It is preserved
+for backwards compatibility, but is not imported into Sciris by default.
 '''
 
 import types
+import warnings
+import textwrap
 import traceback
 import gzip as gz
 import numpy as np
 import pickle as pkl
 import datetime as dt
 import copyreg as cpreg
+import multiprocess as mp
 from io import BytesIO as IO
+from functools import partial
 from contextlib import closing
-from . import sc_fileio as scf
 from . import sc_utils as scu
 from . import sc_odict as sco
+from . import sc_fileio as scf
+from . import sc_profiling as scp
 
 
 ##############################################################################
@@ -392,3 +397,169 @@ class legacy_dataframe(object): # pragma: no cover
         self.cols = list(cols)
         self.data = data
         return
+    
+
+
+##############################################################################
+#%% Parallelization
+##############################################################################
+
+
+def parallelcmd(cmd=None, parfor=None, returnval=None, maxcpu=None, maxmem=None, interval=None, die=True, **kwargs):
+    '''
+    A function to parallelize any block of code. Note: this is intended for quick
+    prototyping only; since it uses exec(), it is not recommended for use in production
+    code.
+
+    Args:
+        cmd       (str):   a string representation of the code to be run in parallel
+        parfor    (dict):  a dictionary of lists of the variables to loop over
+        returnval (str):   the name of the output variable
+        maxcpu    (float): maximum CPU load; used by ``sc.loadbalancer()``
+        maxmem    (float): maximum fraction of virtual memory (RAM); used by ``sc.loadbalancer()``
+        interval  (float): the time delay to poll to see if load is OK,  used in ``sc.loadbalancer()``
+        die       (bool):  whether to stop immediately if an exception is encountered (otherwise, store the exception as the result)
+        kwargs    (dict):  variables to pass into the code
+
+    **Example**::
+
+        const = 4
+        parfor = {'val':[3,5,9]}
+        returnval = 'result'
+        cmd = """
+        newval = val+const
+        result = newval**2
+        """
+        results = sc.parallelcmd(cmd=cmd, parfor=parfor, returnval=returnval, const=const)
+
+    | New in version 2.0.0: replaced ``maxload`` with ``maxcpu``/``maxmem``; automatically de-indent the command
+    | Migrated to ``sc_legacy`` in version 2.2.0.
+    '''
+
+    # Handle maxload
+    maxload = kwargs.pop('maxload', None)
+    if maxload is not None: # pragma: no cover
+        maxcpu = maxload
+        warnmsg = 'sc.loadbalancer() argument "maxload" has been renamed "maxcpu" as of v2.0.0'
+        warnings.warn(warnmsg, category=FutureWarning, stacklevel=2)
+
+    # Deindent the command
+    cmd = textwrap.dedent(cmd)
+
+    # Create queue
+    nfor = len(list(parfor.values())[0])
+    outputqueue = mp.Queue()
+    outputlist = np.empty(nfor, dtype=object)
+    processes = []
+    for i in range(nfor):
+        args = (cmd, parfor, returnval, i, outputqueue, maxcpu, maxmem, interval, die, kwargs)
+        prc = mp.Process(target=_parallelcmd_task, args=args)
+        prc.start()
+        processes.append(prc)
+    for i in range(nfor):
+        _i,returnval = outputqueue.get()
+        outputlist[_i] = returnval
+    for prc in processes:
[…]
