+++ sciris/sc_parallel.py
@@ -8,14 +8,11 @@ Highlights:
     - :func:`parallelize`: as-easy-as-possible parallelization
 '''
 
-
+import warnings
 import numpy as np
 import multiprocess as mp
 import multiprocessing as mpi
 import concurrent.futures as cf
-from functools import partial
-import textwrap
-import warnings
 from . import sc_utils as scu
 from . import sc_profiling as scp
 
@@ -24,7 +21,7 @@ from . import sc_profiling as scp
 #%% Parallelization functions
 ##############################################################################
 
-__all__ = ['parallelize', 'parallelcmd', 'parallel_progress']
+__all__ = ['parallelize']
 
 
 def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncpus=None, 
@@ -325,11 +322,7 @@ def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncp
                 argslist = scu.dcp(argslist) # Also need to deepcopy here
             with cf.ThreadPoolExecutor(max_workers=ncpus) as pool:
                 outputlist = list(pool.map(_parallel_task, argslist))
-        
-        elif pname == 'process':
-            with mpi.Pool(processes=ncpus) as pool:
-                outputlist = pool.map(_parallel_task, argslist)
-        
+
         elif pname == 'custom':
             outputlist = parallelizer(_parallel_task, argslist)
             
@@ -368,128 +361,6 @@ def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncp
         return pool, outputlist
     else:
         return outputlist
-
-
-def parallelcmd(cmd=None, parfor=None, returnval=None, maxcpu=None, maxmem=None, interval=None, die=True, **kwargs):
-    '''
-    A function to parallelize any block of code. Note: this is intended for quick
-    prototyping only; since it uses exec(), it is not recommended for use in production
-    code.
-
-    Args:
-        cmd       (str):   a string representation of the code to be run in parallel
-        parfor    (dict):  a dictionary of lists of the variables to loop over
-        returnval (str):   the name of the output variable
-        maxcpu    (float): maximum CPU load; used by ``sc.loadbalancer()``
-        maxmem    (float): maximum fraction of virtual memory (RAM); used by ``sc.loadbalancer()``
-        interval  (float): the time delay to poll to see if load is OK,  used in ``sc.loadbalancer()``
-        die       (bool):  whether to stop immediately if an exception is encountered (otherwise, store the exception as the result)
-        kwargs    (dict):  variables to pass into the code
-
-    **Example**::
-
-        const = 4
-        parfor = {'val':[3,5,9]}
-        returnval = 'result'
-        cmd = """
-        newval = val+const
-        result = newval**2
-        """
-        results = sc.parallelcmd(cmd=cmd, parfor=parfor, returnval=returnval, const=const)
-
-    | New in version 2.0.0: replaced ``maxload`` with ``maxcpu``/``maxmem``; automatically de-indent the command
-    | New in version 2.2.0: added "die" argument
-    '''
-
-    # Handle maxload
-    maxload = kwargs.pop('maxload', None)
-    if maxload is not None: # pragma: no cover
-        maxcpu = maxload
-        warnmsg = 'sc.loadbalancer() argument "maxload" has been renamed "maxcpu" as of v2.0.0'
-        warnings.warn(warnmsg, category=FutureWarning, stacklevel=2)
-
-    # Deindent the command
-    cmd = textwrap.dedent(cmd)
-
-    # Create queue
-    nfor = len(list(parfor.values())[0])
-    outputqueue = mp.Queue()
-    outputlist = np.empty(nfor, dtype=object)
-    processes = []
-    for i in range(nfor):
[…]
