### 2023-04-11 05cc48a
remove deprecated test
