+++ sciris/sc_parallel.py
@@ -18,10 +18,242 @@ from . import sc_profiling as scp
 
 
 ##############################################################################
-#%% Parallelization functions
+#%% Parallelization class
 ##############################################################################
 
-__all__ = ['parallelize']
+__all__ = ['Parallel', 'parallelize']
+
+# Define a common error message
+freeze_support_error = '''
+Uh oh! It appears you are trying to run with multiprocessing on Windows outside
+of the __main__ block; please see https://docs.python.org/3/library/multiprocessing.html
+for more information. The correct syntax to use is e.g.
+
+import sciris as sc
+
+def my_func(x):
+    return
+
+if __name__ == '__main__':
+    sc.parallelize(my_func)
+'''
+
+
+class Parallel:
+    '''
+    Parallelization manager
+    
+    # TODO!!!!!
+    
+    '''
+    def __init__(self, func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncpus=None, 
+                 maxcpu=None, maxmem=None, interval=None, parallelizer=None, serial=False, 
+                 returnpool=False, progress=False, callback=None, die=True, **func_kwargs):
+        
+        # Store input arguments
+        self.func         = func
+        self.iterarg      = iterarg
+        self.iterkwargs   = iterkwargs
+        self.args         = args
+        self.kwargs       = scu.mergedicts(kwargs, func_kwargs)
+        self.ncpus        = ncpus
+        self.maxcpu       = maxcpu
+        self.maxmem       = maxmem
+        self.interval     = interval
+        self.parallelizer = parallelizer
+        self.serial       = serial
+        self.returnpool   = returnpool
+        self.progress     = progress
+        self.callback     = callback
+        self.die          = die
+        
+        # Additional setup
+        self.set_balancer()
+        self.set_iterargs()
+        
+        return
+        
+    
+    def set_balancer(self):
+        ''' Configure load balancer settings '''
+        maxload = self.kwargs.pop('maxload', None)
+        if maxload is not None: # pragma: no cover
+            self.maxcpu = maxload
+            warnmsg = 'sc.loadbalancer() argument "maxload" has been renamed "maxcpu" as of v2.0.0'
+            warnings.warn(warnmsg, category=FutureWarning, stacklevel=2)
+        if self.ncpus is None and self.maxcpu is None:
+            self.ncpus = scp.cpu_count()
+        if self.ncpus is not None and self.ncpus < 1: # Less than one, treat as a fraction of total
+            self.ncpus = int(scp.cpu_count()*self.ncpus)
+        return
+    
+    
+    def set_iterargs(self):
+        ''' Handle iterarg and iterkwargs '''
+        niters = 0
+        embarrassing = False # Whether or not it's an embarrassingly parallel optimization
+        if iterarg is not None and iterkwargs is not None: # pragma: no cover
+            errormsg = 'You can only use one of iterarg or iterkwargs as your iterable, not both'
+            raise ValueError(errormsg)
+        if iterarg is not None:
+            if not(scu.isiterable(iterarg)):
+                try:
+                    iterarg = np.arange(iterarg)
+                    embarrassing = True
+                except Exception as E: # pragma: no cover
+                    errormsg = f'Could not understand iterarg "{iterarg}": not iterable and not an integer: {str(E)}'
+                    raise TypeError(errormsg)
+            niters = len(iterarg)
+        if iterkwargs is not None: # Check that iterkwargs has the right format
+            if isinstance(iterkwargs, dict): # It's a dict of lists, e.g. {'x':[1,2,3], 'y':[2,3,4]}
+                for key,val in iterkwargs.items():
+                    if not scu.isiterable(val): # pragma: no cover
+                        errormsg = f'iterkwargs entries must be iterable, not {type(val)}'
[…]
