+++ sciris/sc_parallel.py
@@ -14,7 +14,8 @@ import multiprocess as mp
 import multiprocessing as mpi
 import concurrent.futures as cf
 from . import sc_utils as scu
-from . import sc_profiling as scp
+from . import sc_printing as scp
+from . import sc_profiling as scpro
 
 
 ##############################################################################
@@ -67,12 +68,64 @@ class Parallel:
         self.callback     = callback
         self.die          = die
         
+        # Pre-allocate other results
+        self.niters       = None
+        self.embarrassing = None
+        self.argslist     = None
+        self.outputlist   = None
+        self.already_run  = False
+        
         # Additional setup
+        self.configure()
+        return
+    
+    
+    def __repr__(self):
+        return 'TODO'
+    
+    
+    def disp(self):
+        return scp.prepr(self)
+        
+    
+    def configure(self):
+        '''
+        Perform all configuration steps; this can safely be called after initialization
+        '''
+        self.set_defaults()
         self.set_balancer()
-        self.set_iterargs()
+        self.validate_iterable()
+        self.make_argslist()
+        return
+    
+    
+    def set_defaults(self):
+        ''' Define defaults for parallelization '''
+        self.defaults = scu.objdict()
+        
+        # Define default parallelizers
+        self.defaults.fast   = 'concurrent.futures'
+        self.defaults.robust = 'multiprocess'
+        
+        # Map parallelizer to consistent choices
+        self.defaults.mapping = {
+            None                 : self.defaults.robust,
+            'default'            : self.defaults.robust,
+            'robust'             : self.defaults.robust,
+            'fast'               : self.defaults.fast,
+            'serial'             : 'serial',
+            'serial-copy'        : 'serial',
+            'concurrent.futures' : 'concurrent.futures',
+            'concurrent'         : 'concurrent.futures',
+            'multiprocess'       : 'multiprocess',
+            'multiprocessing'    : 'multiprocessing',
+            'thread'             : 'thread',
+            'threadpool'         : 'thread',
+            'thread-copy'        : 'thread',
+        }
         
         return
-        
+    
     
     def set_balancer(self):
         ''' Configure load balancer settings '''
@@ -82,19 +135,25 @@ class Parallel:
             warnmsg = 'sc.loadbalancer() argument "maxload" has been renamed "maxcpu" as of v2.0.0'
             warnings.warn(warnmsg, category=FutureWarning, stacklevel=2)
         if self.ncpus is None and self.maxcpu is None:
-            self.ncpus = scp.cpu_count()
+            self.ncpus = scpro.cpu_count()
         if self.ncpus is not None and self.ncpus < 1: # Less than one, treat as a fraction of total
-            self.ncpus = int(scp.cpu_count()*self.ncpus)
+            self.ncpus = int(scpro.cpu_count()*self.ncpus)
         return
     
     
-    def set_iterargs(self):
+    def validate_iterable(self):
         ''' Handle iterarg and iterkwargs '''
+        iterarg = self.iterarg
+        iterkwargs = self.iterkwargs
         niters = 0
         embarrassing = False # Whether or not it's an embarrassingly parallel optimization
+        
+        # Check that only one was provided
         if iterarg is not None and iterkwargs is not None: # pragma: no cover
             errormsg = 'You can only use one of iterarg or iterkwargs as your iterable, not both'
             raise ValueError(errormsg)
+        
+        # Validate iterarg
         if iterarg is not None:
             if not(scu.isiterable(iterarg)):
                 try:
@@ -104,7 +163,10 @@ class Parallel:
                     errormsg = f'Could not understand iterarg "{iterarg}": not iterable and not an integer: {str(E)}'
                     raise TypeError(errormsg)
[…]
