+++ sciris/sc_parallel.py
@@ -50,7 +50,8 @@ class Parallel:
     '''
     def __init__(self, func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncpus=None, 
                  maxcpu=None, maxmem=None, interval=None, parallelizer=None, serial=False, 
-                 returnpool=False, progress=False, callback=None, die=True, **func_kwargs):
+                 returnpool=False, progress=False, callback=None, label=None, use_async=True,
+                 die=True, **func_kwargs):
         
         # Store input arguments
         self.func         = func
@@ -67,10 +68,12 @@ class Parallel:
         self.returnpool   = returnpool
         self.progress     = progress
         self.callback     = callback
+        self.label        = label
+        self.use_async    = use_async
         self.die          = die
         
         # Pre-allocate other results
-        self.niters       = None
+        self.njobs       = None
         self.embarrassing = None
         self.argslist     = None
         self.outputlist   = None
@@ -82,11 +85,16 @@ class Parallel:
     
     
     def __repr__(self):
+        # Set label string
+        labelstr = f'"{self.label}"' if self.label else '<no label>'
+        string   = f'Parallel({labelstr}; {start} to {end}; pop: {pop_size:n} {pop_type}; epi: {results})'
+
         return 'TODO'
     
     
     def disp(self):
-        return scp.prepr(self)
+        ''' Display the full representation of the object '''
+        return scp.pr(self)
         
     
     def configure(self):
@@ -146,7 +154,7 @@ class Parallel:
         ''' Handle iterarg and iterkwargs '''
         iterarg = self.iterarg
         iterkwargs = self.iterkwargs
-        niters = 0
+        njobs = 0
         embarrassing = False # Whether or not it's an embarrassingly parallel optimization
         
         # Check that only one was provided
@@ -163,7 +171,7 @@ class Parallel:
                 except Exception as E: # pragma: no cover
                     errormsg = f'Could not understand iterarg "{iterarg}": not iterable and not an integer: {str(E)}'
                     raise TypeError(errormsg)
-            niters = len(iterarg)
+            njobs = len(iterarg)
             
         # Validate iterkwargs
         if iterkwargs is not None: 
@@ -173,15 +181,15 @@ class Parallel:
                     if not scu.isiterable(val): # pragma: no cover
                         errormsg = f'iterkwargs entries must be iterable, not {type(val)}'
                         raise TypeError(errormsg)
-                    if not niters:
-                        niters = len(val)
+                    if not njobs:
+                        njobs = len(val)
                     else:
-                        if len(val) != niters: # pragma: no cover
-                            errormsg = f'All iterkwargs iterables must be the same length, not {niters} vs. {len(val)}'
+                        if len(val) != njobs: # pragma: no cover
+                            errormsg = f'All iterkwargs iterables must be the same length, not {njobs} vs. {len(val)}'
                             raise ValueError(errormsg)
             
             elif isinstance(iterkwargs, list): # It's a list of dicts, e.g. [{'x':1, 'y':2}, {'x':2, 'y':3}, {'x':3, 'y':4}]
-                niters = len(iterkwargs)
+                njobs = len(iterkwargs)
                 for item in iterkwargs:
                     if not isinstance(item, dict): # pragma: no cover
                         errormsg = f'If iterkwargs is a list, each entry must be a dict, not {type(item)}'
@@ -192,11 +200,11 @@ class Parallel:
                 raise TypeError(errormsg)
     
         # Final error checking
-        if niters == 0:
+        if njobs == 0:
             errormsg = 'Nothing found to parallelize: please supply an iterarg, iterkwargs, or both'
             raise ValueError(errormsg)
         else:
[…]
