+++ sciris/sc_parallel.py
@@ -59,7 +59,7 @@ class Parallel:
         self.iterkwargs   = iterkwargs
         self.args         = args
         self.kwargs       = scu.mergedicts(kwargs, func_kwargs)
-        self.ncpus        = ncpus
+        self._ncpus       = ncpus # With a prefix since dynamically calculated later
         self.maxcpu       = maxcpu
         self.maxmem       = maxmem
         self.interval     = interval
@@ -72,40 +72,48 @@ class Parallel:
         self.use_async    = use_async
         self.die          = die
         
-        # Pre-allocate other results
-        self.njobs       = None
+        # Additional initialization
+        self.init()
+        return
+    
+    
+    def init(self):
+        '''
+        Perform all remaining initialization steps; this can safely be called after object creation
+        '''
+        self.reset()
+        self.set_defaults()
+        self.validate_args()
+        self.make_argslist()
+        self.set_ncpus()
+        self.set_method()
+        return
+    
+    
+    def reset(self):
+        ''' Reset to the pre-run state '''
+        self.ncpus        = None
+        self.njobs        = None
         self.embarrassing = None
         self.argslist     = None
+        self.method       = None
+        self.pool         = None
         self.outputlist   = None
         self.already_run  = False
-        
-        # Additional setup
-        self.configure()
         return
     
     
     def __repr__(self):
-        # Set label string
+        ''' Brief representation of the object '''
         labelstr = f'"{self.label}"' if self.label else '<no label>'
-        string   = f'Parallel({labelstr}; {start} to {end}; pop: {pop_size:n} {pop_type}; epi: {results})'
-
-        return 'TODO'
+        string   = f'Parallel({labelstr}; njobs={self.njobs}; ncpus={self.ncpus}; method={self.method}; run={self.already_run})'
+        return string
     
     
     def disp(self):
         ''' Display the full representation of the object '''
         return scp.pr(self)
         
-    
-    def configure(self):
-        '''
-        Perform all configuration steps; this can safely be called after initialization
-        '''
-        self.set_defaults()
-        self.set_balancer()
-        self.validate_iterable()
-        self.make_argslist()
-        return
     
     
     def set_defaults(self):
@@ -136,22 +144,8 @@ class Parallel:
         return
     
     
-    def set_balancer(self):
-        ''' Configure load balancer settings '''
-        maxload = self.kwargs.pop('maxload', None)
-        if maxload is not None: # pragma: no cover
-            self.maxcpu = maxload
-            warnmsg = 'sc.loadbalancer() argument "maxload" has been renamed "maxcpu" as of v2.0.0'
-            warnings.warn(warnmsg, category=FutureWarning, stacklevel=2)
-        if self.ncpus is None and self.maxcpu is None:
-            self.ncpus = scpro.cpu_count()
-        if self.ncpus is not None and self.ncpus < 1: # Less than one, treat as a fraction of total
-            self.ncpus = int(scpro.cpu_count()*self.ncpus)
-        return
-    
-    
-    def validate_iterable(self):
-        ''' Handle iterarg and iterkwargs '''
+    def validate_args(self):
+        ''' Validate iterarg and iterkwargs '''
         iterarg = self.iterarg
         iterkwargs = self.iterkwargs
         njobs = 0
@@ -241,57 +235,89 @@ class Parallel:
         self.argslist = argslist
         return
     
+    
+    def set_ncpus(self):
+        ''' Configure number of CPUs '''
+        
+        # Handle maxload deprecation
+        maxload = self.kwargs.pop('maxload', None)
+        if maxload is not None: # pragma: no cover
+            self.maxcpu = maxload
[…]
