+++ sciris/sc_parallel.py
@@ -50,7 +50,7 @@ class Parallel:
     '''
     def __init__(self, func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncpus=None, 
                  maxcpu=None, maxmem=None, interval=None, parallelizer=None, serial=False, 
-                 returnpool=False, progress=False, callback=None, label=None, use_async=True,
+                 returnpool=False, progress=False, callback=None, label=None, use_async=False,
                  die=True, **func_kwargs):
         
         # Store input arguments
@@ -98,7 +98,8 @@ class Parallel:
         self.argslist     = None
         self.method       = None
         self.pool         = None
-        self.outputlist   = None
+        self.jobs         = None
+        self.results      = None
         self.already_run  = False
         return
     
@@ -266,11 +267,12 @@ class Parallel:
     def set_method(self):
         ''' Choose which method to use for parallelization '''
         
-        # Handle the choice of parallelizer
-        parallelizer = self.parallelizer
-        if self.serial: # This is a separate keyword argument, but make it consistent
+        # This is a separate keyword argument for convenience, but is equivalent to paralellizer='serial'
+        if self.serial: 
             parallelizer = 'serial'
         
+        # Handle the choice of parallelizer
+        parallelizer = self.parallelizer
         if parallelizer is None or scu.isstring(parallelizer):
             try:
                 self.method = self.defaults.mapping[parallelizer]
@@ -282,59 +284,82 @@ class Parallel:
         
         return
     
-        
-    def run_parallel(self):
-        ''' Choose how to run in parallel, and do it '''
+    
+    def make_pool(self):
+        ''' Make the pool '''
         
         # Shorten variables
         ncpus = self.ncpus
         method = self.method
-        parallelizer = self.parallelizer
-        argslist = self.argslist
-        
-        # Choose which parallelizer to use
         pool = None
-        if method == 'serial':
-            if 'copy' in parallelizer: # Niche use case of running without deepcopying
-                argslist = scu.dcp(argslist) # Need to deepcopy here, since effectively deecopied by other parallelization methods
-            outputlist = list(map(_task, argslist))
         
-        elif method == 'multiprocess': # Main use case
-            with mp.Pool(processes=ncpus) as pool:
-                outputlist = list(pool.map(_task, argslist))
+        if method == 'multiprocess': # Main use case
+            pool = mp.Pool(processes=ncpus)
         
         elif method == 'multiprocessing':
-            with mpi.Pool(processes=ncpus) as pool:
-                outputlist = pool.map(_task, argslist)
+            pool = mpi.Pool(processes=ncpus)
         
         elif method == 'concurrent.futures':
-            with cf.ProcessPoolExecutor(max_workers=ncpus) as pool:
-                outputlist = list(pool.map(_task, argslist))
+            pool = cf.ProcessPoolExecutor(max_workers=ncpus)
         
         elif method == 'thread':
-            if 'copy' in parallelizer: # Niche use case of running without deepcopying
-                argslist = scu.dcp(argslist) # Also need to deepcopy here
-            with cf.ThreadPoolExecutor(max_workers=ncpus) as pool:
-                outputlist = list(pool.map(_task, argslist))
-
-        elif method == 'custom':
-            outputlist = parallelizer(_task, argslist)
-            
-        else: # Should be unreachable; exception should have already been caught
-            errormsg = f'Invalid parallelizer "{parallelizer}"'
-            raise ValueError(errormsg)
+            pool = cf.ThreadPoolExecutor(max_workers=ncpus)
+        
+        # Reset
+        self.pool    = pool
+        self.jobs    = None
+        self.results = None
+        
+        return
+        
+    
+    
+    def run_async(self):
[…]
