+++ sciris/sc_utils.py
@@ -1629,11 +1629,11 @@ class prettyobj(object):
     | New in version 2.0.0: allow positional arguments
     '''
 
-    def __init__(self, *args, **kwargs):
-        kwargs = mergedicts(*args, kwargs)
-        for k,v in kwargs.items():
-            self.__dict__[k] = v
-        return
+    # def __init__(self, *args, **kwargs):
+    #     kwargs = mergedicts(*args, kwargs)
+    #     for k,v in kwargs.items():
+    #         self.__dict__[k] = v
+    #     return
 
 
     def __repr__(self):
