### 2023-04-11 bf2e23e
fails from command line with NotImplementedError: pool objects cannot be passed between processes or pickled, going back to earlier version
