+++ sciris/sc_parallel.py
@@ -117,7 +117,6 @@ class Parallel:
     def disp(self):
         ''' Display the full representation of the object '''
         return scp.pr(self)
-        
     
     
     def set_defaults(self):
@@ -135,14 +134,13 @@ class Parallel:
             'robust'             : self.defaults.robust,
             'fast'               : self.defaults.fast,
             'serial'             : 'serial',
-            'serial-copy'        : 'serial',
             'concurrent.futures' : 'concurrent.futures',
             'concurrent'         : 'concurrent.futures',
+            'futures'            : 'concurrent.futures',
             'multiprocess'       : 'multiprocess',
             'multiprocessing'    : 'multiprocessing',
             'thread'             : 'thread',
             'threadpool'         : 'thread',
-            'thread-copy'        : 'thread',
         }
         
         return
@@ -277,6 +275,7 @@ class Parallel:
         # Handle the choice of parallelizer
         parallelizer = self.parallelizer
         if parallelizer is None or scu.isstring(parallelizer):
+            parallelizer = parallelizer.replace('-copy', '').replace('-async', '')
             try:
                 self.method = self.defaults.mapping[parallelizer]
             except:
@@ -288,82 +287,83 @@ class Parallel:
         return
     
     
-    def make_pool(self):
-        ''' Make the pool and map function '''
+    # def make_pool(self):
+    #     ''' Make the pool and map function '''
         
-        # Shorten variables
-        ncpus = self.ncpus
-        method = self.method
+    #     # Reset
+    #     self.pool     = pool
+    #     self.map_func = map_func
+    #     self.is_async = is_async
+    #     self.jobs     = None
+    #     self.results  = None
         
-        if method == 'serial':
-            pool = None
-            map_func = lambda task,argslist: list(map(task, argslist))
-            is_async = False
-        
-        elif method == 'multiprocess': # Main use case
-            pool = mp.Pool(processes=ncpus)
-            map_func = pool.map_async
-            is_async = True
-        
-        elif method == 'multiprocessing':
-            pool = mpi.Pool(processes=ncpus)
-            map_func = pool.map_async
-            is_async = True
-        
-        elif method == 'concurrent.futures':
-            pool = cf.ProcessPoolExecutor(max_workers=ncpus)
-            map_func = pool.map
-            is_async = False
-        
-        elif method == 'thread':
-            pool = cf.ThreadPoolExecutor(max_workers=ncpus)
-            map_func = pool.map_async
-            is_async = True
-        
-        elif method == 'custom':
-            pool = None
-            map_func = self.parallelizer
-            is_async = False
-        
-        else: # Should be unreachable; exception should have already been caught
-            errormsg = f'Invalid parallelizer "{self.parallelizer}"'
-            raise ValueError(errormsg)
-            
-        # Reset
-        self.pool     = pool
-        self.map_func = map_func
-        self.is_async = is_async
-        self.jobs     = None
-        self.results  = None
-        
-        return
+    #     return
         
     
     def run_async(self):
         ''' Choose how to run in parallel, and do it '''
         
-        # Force deepcopying for serial or thread arguments
-        if scu.isstring(self.parallelizer) and 'copy' in self.parallelizer: 
-            argslist = scu.dcp(self.argslist, die=self.die)
-        else:
-            argslist = self.argslist
+        # Shorten variables
+        ncpus = self.ncpus
+        method = self.method
+        is_async = False
+        needs_copy = ['serial', 'thread', 'custom']
+        supports_async = ['multiprocess', 'multiprocessing', 'thread']
         
-        # Create the pool and initialize results
-        self.make_pool()
+        # Handle additional options
+        if scu.isstring(self.parallelizer):
+            
+            # Handle deepcopying
[…]
