+++ sciris/sc_datetime.py
@@ -8,7 +8,7 @@ Highlights:
     - :func:`datedelta`: perform calculations on date strings
 '''
 
-import time
+import time as pytime
 import warnings
 import numpy as np
 import pandas as pd
@@ -24,8 +24,20 @@ from . import sc_printing as scp
 #%% Date functions
 ###############################################################################
 
-__all__ = ['now', 'getdate', 'readdate', 'date', 'day', 'daydiff', 'daterange', 'datedelta', 'datetoyear']
+__all__ = ['time', 'now', 'getdate', 'readdate', 'date', 'day', 'daydiff', 'daterange', 'datedelta', 'datetoyear']
 
+
+def time():
+    '''
+    Get current time in seconds -- alias to time.time()
+    
+    See also :func:`now()` to return a datetime object, and :func:`getdate()` to 
+    return a string.
+    
+    New in version 2.2.0.
+    '''
+    return pytime.time()
+    
 
 def now(astype='dateobj', timezone=None, utc=False, tostring=False, dateformat=None):
     '''
@@ -233,7 +245,7 @@ def readdate(datestr=None, *args, dateformat=None, return_defaults=False, verbos
     return output
 
 
-def date(obj, *args, start_date=None, readformat=None, to='date', as_date=None, outformat=None, **kwargs):
+def date(obj=None, *args, start_date=None, readformat=None, to='date', as_date=None, outformat=None, **kwargs):
     '''
     Convert any reasonable object -- a string, integer, or datetime object, or
     list/array of any of those -- to a date object (or string, pandas, or numpy
@@ -248,7 +260,7 @@ def date(obj, *args, start_date=None, readformat=None, to='date', as_date=None, 
     underscores (e.g. ``start_date`` or ``startdate``)
 
     Args:
-        obj (str/int/date/datetime/list/array): the object to convert
+        obj (str/int/date/datetime/list/array): the object to convert; if None, return current date
         args (str/int/date/datetime): additional objects to convert
         start_date (str/date/datetime): the starting date, if an integer is supplied
         readformat (str/list): the format to read the date in; passed to ``sc.readdate()``
@@ -269,7 +281,7 @@ def date(obj, *args, start_date=None, readformat=None, to='date', as_date=None, 
     | New in version 1.0.0.
     | New in version 1.2.2: "readformat" argument; renamed "dateformat" to "outformat"
     | New in version 2.0.0: support for ``np.datetime64`` objects
-    | New in version 2.2.0: added "to" argument, and support for ``pd.Timestamp`` and ``np.datetime64`` output
+    | New in version 2.2.0: added "to" argument, and support for ``pd.Timestamp`` and ``np.datetime64`` output; allow None
     '''
 
     # Handle deprecation
@@ -285,7 +297,7 @@ def date(obj, *args, start_date=None, readformat=None, to='date', as_date=None, 
 
     # Convert to list and handle other inputs
     if obj is None:
-        return
+        obj = dt.datetime.now().date()
     if outformat is None:
         outformat = '%Y-%m-%d'
     obj, is_list, is_array = scu._sanitize_iterables(obj, *args)
@@ -580,7 +592,7 @@ def tic():
     See also ``sc.timer()``.
     '''
     global _tictime  # The saved time is stored in this global
-    _tictime = time.time()  # Store the present time in the global
+    _tictime = pytime.time()  # Store the present time in the global
     return _tictime    # Return the same stored number
 
 
@@ -612,7 +624,7 @@ def toc(start=None, label=None, baselabel=None, sigfigs=None, reset=False, outpu
 
     New in version 1.3.0: new arguments.
     '''
-    now = time.time() # Get the time as quickly as possible
+    now = pytime.time() # Get the time as quickly as possible
 
     from . import sc_printing as scp # To avoid circular import
[…]
