### 2023-04-12 93fe22c
working, renamed to globaldict
