### 2023-04-12 7dbf2f0
fix asd coverage and change how coverage is run
