+++ sciris/sc_nested.py
@@ -3,7 +3,7 @@ Nested dictionary functions.
 '''
 
 import itertools
-from functools import reduce
+from functools import reduce, partial
 from . import sc_utils as scu
 
 
@@ -71,13 +71,43 @@ def makenested(nesteddict, keylist=None, value=None, overwrite=False, generator=
     if isinstance(currentlevel, dict):
         if overwrite or lastkey not in currentlevel:
             currentlevel[lastkey] = value
-        elif not overwrite and value is not None:
+        elif not overwrite and value is not None: # pragma: no cover
             errormsg = f'Not overwriting entry {keylist} since overwrite=False'
             raise ValueError(errormsg)
-    elif value is not None:
+    elif value is not None: # pragma: no cover
         errormsg = f'Cannot set value {value} since entry {keylist} is a {type(currentlevel)}, not a dict'
         raise TypeError(errormsg)
     return
+
+
+def check_iter_type(obj):
+    ''' Helper function to determine if an object is a dict, list, or neither '''
+    if isinstance(obj, dict):
+        out = 'dict'
+    elif isinstance(obj, list):
+        out = 'list'
+    elif hasattr(obj, '__dict__'):
+        out = 'object'
+    else:
+        out = '' # Evaluates to false
+    return out
+    
+
+def get_from_obj(ndict, key, safe=False):
+    ''' Get an item from a dict, list, or object by key '''
+    itertype = check_iter_type(ndict)
+    if itertype == 'dict':
+        if safe:
+            out = ndict.get(key)
+        else:
+            out = ndict[key]
+    elif itertype == 'list':
+        out = ndict[key]
+    elif itertype == 'object':
+        out = getattr(ndict, key)
+    else:
+        out = None
+    return out
 
 
 def getnested(nesteddict, keylist, safe=False):
@@ -88,7 +118,8 @@ def getnested(nesteddict, keylist, safe=False):
 
     See sc.makenested() for full documentation.
     '''
-    output = reduce(lambda d, k: d.get(k) if d else None if safe else d[k], keylist, nesteddict)
+    get = partial(get_from_obj, safe=safe)
+    output = reduce(get, keylist, nesteddict)
     return output
 
 
@@ -103,7 +134,7 @@ def setnested(nesteddict, keylist, value, force=True):
     if force:
         makenested(nesteddict, keylist, overwrite=False)
     currentlevel = getnested(nesteddict, keylist[:-1])
-    if not isinstance(currentlevel, dict):
+    if not isinstance(currentlevel, dict): # pragma: no cover
         errormsg = f'Cannot set {keylist} since parent is a {type(currentlevel)}, not a dict'
         raise TypeError(errormsg)
     else:
@@ -211,7 +242,7 @@ def flattendict(nesteddict, sep=None, _prefix=None):
 
 # Define a custom "None" value to allow searching for actual None values
 _None = 'sc.search() placeholder'
-def search(obj, key=_None, value=_None, _trace=''):
+def search(obj, key=_None, value=_None, aslist=False, _trace=None):
     """
     Find a key/attribute or value within a list, dictionary or object.
 
@@ -219,51 +250,71 @@ def search(obj, key=_None, value=_None, _trace=''):
     by searching recursively through keys or attributes.
 
     Args:
-        obj: A dict or class with __dict__ attribute
-        attribute: The substring to search for
+        obj (any): A dict, list, or object
+        key (any): The key to search for
+        value (any): The value to search for
+        aslist (bool): return entries as a list (else, return as a string)
         _trace: Not for user input - internal variable used for recursion
 
     Returns:
         A list of matching attributes. The items in the list are the Python
[…]
