+++ sciris/sc_colors.py
@@ -503,7 +503,7 @@ def colormapdemo(cmap=None, n=None, smoothing=None, randseed=None, doshow=True):
     cb2.set_label('Height (km)',horizontalalignment='right', labelpad=50)
     pl.xlabel('Position (km)')
     pl.ylabel('Position (km)')
-    if doshow:
+    if doshow: # pragma: no cover
         pl.show()
 
     # Plot in 3D
@@ -516,7 +516,7 @@ def colormapdemo(cmap=None, n=None, smoothing=None, randseed=None, doshow=True):
     cb.set_label('Height (km)', horizontalalignment='right', labelpad=50)
     pl.xlabel('Position (km)')
     pl.ylabel('Position (km)')
-    if doshow:
+    if doshow: # pragma: no cover
         pl.show()
 
     return {'2d':fig1, '3d':fig2}
@@ -580,7 +580,7 @@ def alpinecolormap(apply=False):
 
     # Make map
     cmap = mplc.LinearSegmentedColormap('alpine', cdict, 256)
-    if apply:
+    if apply: # pragma: no cover
         pl.set_cmap(cmap)
     return cmap
 
@@ -632,7 +632,7 @@ def bicolormap(gap=0.1, mingreen=0.2, redbluemix=0.5, epsilon=0.01, demo=False, 
                      (1.00000, 0.0, 0.0))}
 
     cmap = mplc.LinearSegmentedColormap('bi', cdict, 256)
-    if apply:
+    if apply: # pragma: no cover
         pl.set_cmap(cmap)
 
     def demoplot(): # pragma: no cover
@@ -651,7 +651,7 @@ def bicolormap(gap=0.1, mingreen=0.2, redbluemix=0.5, epsilon=0.01, demo=False, 
             pl.colorbar()
         pl.show()
 
-    if demo:
+    if demo: # pragma: no cover
         demoplot()
 
     return cmap
@@ -703,7 +703,7 @@ def parulacolormap(apply=False):
             [0.9642,0.9437,0.1216], [0.9657,0.9494,0.1168], [0.9674,0.9552,0.1116], [0.9692,0.9609,0.1061], [0.9711,0.9667,0.1001], [0.9730,0.9724,0.0938], [0.9749,0.9782,0.0872], [0.9769,0.9839,0.0805]]
 
     cmap = mplc.LinearSegmentedColormap.from_list('parula', data)
-    if apply:
+    if apply: # pragma: no cover
         pl.set_cmap(cmap)
     return cmap
 
@@ -765,7 +765,7 @@ def turbocolormap(apply=False):
             [0.57103,0.04474,0.00529],[0.55852,0.04028,0.00579],[0.54583,0.03593,0.00638],[0.53295,0.03169,0.00705],[0.51989,0.02756,0.00780],[0.50664,0.02354,0.00863],[0.49321,0.01963,0.00955],[0.47960,0.01583,0.01055]]
 
     cmap = mplc.LinearSegmentedColormap.from_list('turbo', data)
-    if apply:
+    if apply: # pragma: no cover
         pl.set_cmap(cmap)
     return cmap
 
@@ -802,7 +802,7 @@ def bandedcolormap(minvalue=None, minsaturation=None, hueshift=None, saturations
 
     # Create and use
     cmap = mplc.LinearSegmentedColormap.from_list('banded', data)
-    if apply:
+    if apply: # pragma: no cover
         pl.set_cmap(cmap)
     return cmap
 
@@ -820,13 +820,13 @@ def orangebluecolormap(apply=False):
     New in version 1.0.0.
     '''
     bottom = pl.get_cmap('Oranges', 128)
-    top = pl.get_cmap('Blues_r', 128)
[…]
