### 2023-04-12 62abbfd
add coverage to settings
