+++ tests/test_dataframe.py
@@ -12,6 +12,9 @@ def dfprint(label, val=None):
         print(val)
     return None
 
+def subheading(label):
+    return sc.printgreen('\n\n' + label)
+
 def test_dataframe():
     sc.heading('Testing dataframe')
 
@@ -35,22 +38,31 @@ def test_dataframe():
     a.rmrow(value=666); dfprint('Remove the row starting with element "666"', a)
     p = a.to_pandas(); dfprint('Convert to pandas', p)
     b = a.filtercols(['m','x']); dfprint('Filter to columns m and x', b)
+    
+    # Do tests on the final dataframe
+    assert a.x.sum() == 789
+    assert a.y.sum() == 8
+    assert a.m.sum() == 30
+    assert a.shape == (2,3)
+    
     return a
 
 
 def test_methods():
     sc.heading('Testing dataframe methods')
+    
+    subheading('Initialization')
     df = sc.dataframe(cols=['a', 'b'], nrows=3)
     assert df.shape == (3,2)
     df += np.random.random(df.shape)
     dfprint('To start', df)
 
-    # Append row
+    subheading('Append row')
     df.appendrow(dict(a=4, b=4)); dfprint('Append row as dict', df)
     with pytest.raises(ValueError):
         df.appendrow([1,2,3])
 
-    # Get
+    subheading('Get')
     dfprint('Key get', df['a'])
     dfprint('Array get', df[[0,2]])
     dfprint('Tuple get 1', df[0,'a'])
@@ -63,14 +75,14 @@ def test_methods():
     with pytest.raises(sc.KeyNotFoundError):
         df[sc.prettyobj({'wrong':'type'})]
 
-    # Set
+    subheading('Set')
     df['c'] = np.random.randn(df.nrows); dfprint('Set column', df)
     df[2] = [17,15,13]; dfprint('Insert row', df)
     df[0,'a'] = 300; dfprint('Tuple set 1', df)
     df[1,1]   = 400; dfprint('Tuple set 2', df)
     out = df.flexget(cols=['a','c'], rows=[0,2]); dfprint('Flexget', out)
 
-    # Other
+    subheading('Other')
     df.rmrows([1,3]); dfprint('Removing rows', df)
     df.replacecol('a', 300, 333); dfprint('Replacing 300→333', df)
     od = df.to_odict(); dfprint('To dict', od)
@@ -85,13 +97,47 @@ def test_methods():
     df.sortcols(reverse=True); dfprint('Sorting columns', df)
     assert df.cols[-1] == 'a'
     
-    # Printing
+    dfnew = sc.dataframe(cols=['x','y'], data=[['a',2],['b',5],['c',7]])
+    
+    print('df.col_index()')
+    assert dfnew.col_index('y') == dfnew.col_index(1)
+    
+    print('df.set()')
+    dfnew.set('x', ['d','e','f'])
+    assert dfnew.x[2] == 'f'
+    
+    subheading('Printing')
     dfprint('Custom display')
     df2 = sc.dataframe(data=np.random.rand(100,10))
     df2.disp(precision=2, ncols=5, nrows=5, options={'display.colheader_justify': 'left'})
 
     return df
 
+
+def test_errors():
+    sc.heading('Testing dataframe error handling')
+    
+    df = sc.dataframe(cols=['x','y'], data=[['a',2],['b',5],['c',7]])
+    
+    print('Duplicate dtype definitions')
+    with pytest.raises(ValueError):
+        data = [['a','b'],[1,2]]
+        columns = {'str':str,'int':int}
+        dtypes = [str, int]
+        sc.dataframe(data=data, columns=columns, dtypes=dtypes)
+        
+    print('Incompatible columns')
[…]
