+++ sciris/sc_odict.py
@@ -319,10 +319,10 @@ class odict(OD):
                 else:
                     rawoutput = f'{numleft}{ind:d}{numsep} {quote}{keystr}{quote}{keysep}{spacer} \n{valstr}\n'
 
-                # Perform the indentation.
+                # Perform the indentation
                 newoutput = scp.indent(prefix=theprefix, text=rawoutput, width=80)
 
-                # Strip ot any terminal newline.
+                # Strip out any terminal newline
                 if newoutput[-1] == '\n':
                     newoutput = newoutput[:-1]
 
@@ -372,7 +372,7 @@ class odict(OD):
         self._setattr('_stale', True) # Flag to refresh the cached keys
         try:
             return OD.__delitem__(self, key)
-        except Exception as E:
+        except Exception as E: # pragma: no cover
             if isinstance(key, scu._numtype): # If it's a number, use that
                 thiskey = self._ikey(key)
                 return OD.__delitem__(self, thiskey) # Note that defaultdict behavior isn't supported for non-string lookup
@@ -401,7 +401,7 @@ class odict(OD):
             self._cache_keys()
         try:
             return self._cached_keys[key]
-        except IndexError:
+        except IndexError: # pragma: no cover
             errormsg = f'index {key} out of range for dict of length {len(self)}'
             raise IndexError(errormsg) from None # Don't show the traceback
 
@@ -439,7 +439,7 @@ class odict(OD):
                 if odict._matchkey(item, pattern, method):
                     return True
         else: # For everything except a tuple, treat it as a string
-            if not scu.isstring(key):
+            if not scu.isstring(key): # pragma: no cover
                 try:
                     key = str(key) # Try to cast it to a string
                 except Exception as E: # pragma: no cover
@@ -590,9 +590,9 @@ class odict(OD):
 
         **Example**::
 
-            z = odict({'dog':[2,3], 'cat':[4,6], 'mongoose':[4,6]})
-            z.findvals([4,6]) # returns 'cat'
-            z.findvals([4,6], first=False) # returns ['cat', 'mongoose']
+            z = sc.odict({'dog':[2,3], 'cat':[4,6], 'mongoose':[4,6]})
+            z.findbyval([4,6]) # returns 'cat'
+            z.findbyval([4,6], first=False) # returns ['cat', 'mongoose']
         '''
         keys = []
         for key,val in self.items():
@@ -654,7 +654,7 @@ class odict(OD):
             needkey = True
         if key is None or needkey:
             keyname = 'key'+scu.flexstr(len(self))  # Define the key just to be the current index
-        else:
+        else: # pragma: no cover
             keyname = key
         self.__setitem__(keyname, value)
         return
@@ -694,7 +694,7 @@ class odict(OD):
         origkeys = self.keys()
         originds = range(len(origkeys))
         if not len(originds) or realpos==len(originds): # It's empty or in the final position, just append
-            self.__setitem__(realkey, realvalue)
+            self.__setitem__(realkey, realvalue) # pragma: no cover
         else: # Main usage case, it's not empty
             try:
                 insertind = originds.index(realpos) # Figure out which index we're inseting at
@@ -721,7 +721,7 @@ class odict(OD):
     def rename(self, oldkey, newkey):
         ''' Change a key name (note: not optimized for speed) '''
         nkeys = len(self)
-        if isinstance(oldkey, scu._numtype):
+        if isinstance(oldkey, scu._numtype): # pragma: no cover
             index = oldkey
             keystr = self.keys()[index]
         else: # Forge ahead for strings and anything else!
@@ -832,9 +832,9 @@ class odict(OD):
         '''
         # Handle keys
[…]
