+++ sciris/sc_math.py
@@ -114,6 +114,7 @@ def findinds(arr=None, val=None, *args, eps=1e-6, first=False, last=False, ind=N
 
     | New in version 1.2.3: "die" argument
     | New in version 2.0.0: fix string matching; allow multiple arguments
+    | New in version 2.2.0: 
     '''
 
     # Handle first or last
@@ -328,10 +329,11 @@ def sanitize(data=None, returninds=False, replacenans=None, defaultval=None, die
             sanitized4 = sc.sanitize(data, replacenans='linear') # Replace NaNs using linear interpolation
             sanitized5 = sc.sanitize(data, replacenans=0) # Replace NaNs with 0
 
-        New in version 2.0.0: handle multidimensional arrays
+        | New in version 2.0.0: handle multidimensional arrays
+        | New in version 2.2.0: return zero-length arrays if all NaN
         '''
         try:
-            data = np.array(data, dtype=float) # Make sure it's an array of float type
+            data = np.array(data, dtype=float) # Make sure it's an array of float type, otherwise nan operations fail
             is_multidim = data.ndim > 1
             if is_multidim:
                 if not replacenans:
@@ -352,7 +354,7 @@ def sanitize(data=None, returninds=False, replacenans=None, defaultval=None, die
                             raise NotImplementedError(errormsg)
                         newx = range(len(data)) # Create a new x array the size of the original array
                         sanitized = smoothinterp(newx, inds, sanitized, method=replacenans, smoothness=0) # Replace nans with interpolated values
-                    else:
+                    else: # pragma: no cover
                         errormsg = f'Interpolation method "{replacenans}" not found: must be "nearest" or "linear"'
                         raise ValueError(errormsg)
                 else:
@@ -364,7 +366,6 @@ def sanitize(data=None, returninds=False, replacenans=None, defaultval=None, die
                 if defaultval is not None:
                     sanitized = defaultval
                 else:
-                    sanitized = data
                     inds = []
 
                     if verbose: # pragma: no cover
@@ -374,7 +375,7 @@ def sanitize(data=None, returninds=False, replacenans=None, defaultval=None, die
             if die:
                 raise E
             else:
-                sanitized = data # Give up and just return an empty array
+                sanitized = data # Give up and just return original array
                 inds = []
         if returninds: return sanitized, inds
         else:          return sanitized
@@ -474,7 +475,7 @@ def numdigits(n, *args, count_minus=False, count_decimal=False):
         abs_n = abs(n)
         is_decimal = 0 < abs_n < 1
         n_digits = 1
-        if n < 0 and count_minus:
+        if n < 0 and count_minus: # pragma: no cover
             n_digits += 1
         if is_decimal:
             if count_decimal:
@@ -704,7 +705,7 @@ def linregress(x, y, full=False, **kwargs):
     x = scu.toarray(x)
     y = scu.toarray(y)
     fit = np.polyfit(x, y, deg=1, **kwargs) # Do the fit
-    if not full:
+    if not full: # pragma: no cover
         return fit
     else:
         out = sco.objdict()
@@ -728,7 +729,7 @@ def linregress(x, y, full=False, **kwargs):
 __all__ += ['rolling', 'convolve', 'smooth', 'smoothinterp', 'gauss1d', 'gauss2d']
 
 
-def rolling(data, window=7, operation='mean', **kwargs):
[…]
