+++ tests/test_parallel.py
@@ -6,9 +6,18 @@ built-in parallelization, and since functions-within-functions can't be pickled.
 import sciris as sc
 import numpy as np
 import pylab as pl
+import multiprocessing as mp
 import pytest
 
 if 'doplot' not in locals(): doplot = False
+
+
+def subheading(label):
+    return sc.printgreen('\n\n' + label)
+
+def f(i):
+    ''' Test function for parallelization -- here to avoid pickling errors '''
+    return i**2
 
 
 def test_simple():
@@ -96,21 +105,66 @@ def test_exceptions():
 def test_class():
     sc.heading('Testing sc.Parallel class')
     
-    def f(i):
-        print(f'Running {i}')
+    def slowfunc(i):
         np.random.seed(i)
-        sc.timedsleep(1*np.random.rand())
+        sc.timedsleep(0.2*np.random.rand())
         return i**2
-        
     
-    P = sc.Parallel(f, iterarg=range(100), parallelizer='multiprocessing-async', ncpus=2)
+    subheading('Creation and display')
+    P = sc.Parallel(slowfunc, iterarg=range(4), parallelizer='multiprocess-async', ncpus=2) # NB, multiprocessing-async fails with a pickle error
     print(P)
     P.disp()
     
-    sc.timedsleep(1)
+    subheading('Running asynchronously and monitoring')
     P.run_async()
-    P.monitor()
+    P.monitor(interval=0.05)
+    P.finalize()
     
+    
+    subheading('Checking parallelizers')
+    
+    iterarg = np.arange(4)
+    
+    print('Checking serial with copy')
+    r1 = sc.parallelize(f, iterarg, parallelizer='serial-copy')
+    
+    print('Checking multiprocessing')
+    r2 = sc.parallelize(f, iterarg, parallelizer='multiprocessing')
+    
+    print('Checking fast (concurrent.futures)')
+    r3 = sc.parallelize(f, iterarg, parallelizer='fast')
+    
+    print('Checking thread')
+    r4 = sc.parallelize(f, iterarg, parallelizer='thread')
+    
+    print('Checking custom')
+    with mp.Pool(processes=2) as pool:
+        r5 = sc.parallelize(f, iterarg, parallelizer=pool.imap)
+    
+    assert r1 == r2 == r3 == r4 == r5
+
+    
+    subheading('Other')
+    
+    print('Checking CPUs')
+    sc.Parallel(f, 10, ncpus=0.7)
+    
+    print('Validation: no jobs to run')
+    with pytest.raises(ValueError):
+        sc.Parallel(f, iterarg=[])
+        
+    print('Validation: no CPUs')
+    with pytest.raises(ValueError):
+        sc.Parallel(f, 10, ncpus=-3)
+    
+    print('Validation: invalid parallelizer')
+    with pytest.raises(sc.KeyNotFoundError):
+        sc.Parallel(f, 10, parallelizer='invalid-parallelizer')
+        
+    print('Validation: invalid async')
+    with pytest.raises(ValueError):
+        sc.Parallel(f, 10, parallelizer='serial-async')
+        
     return P
 
 
