+++ sciris/sc_versioning.py
@@ -50,7 +50,7 @@ def freeze(lower=False):
     import pkg_resources as pkgr # Imported here since slow (>0.1 s)
     raw = dict(tuple(str(ws).split()) for ws in pkgr.working_set)
     keys = sorted(raw.keys())
-    if lower:
+    if lower: # pragma: no cover
         labels = {k:k.lower() for k in keys}
     else:
         labels = {k:k for k in keys}
@@ -266,22 +266,27 @@ def compareversions(version1, version2):
     if isinstance(version1, types.ModuleType):
         try:
             version1 = version1.__version__
-        except Exception as E:
+        except Exception as E: # pragma: no cover
             errormsg = f'{version1} is a module, but does not have a __version__ attribute'
             raise AttributeError(errormsg) from E
     v1 = str(version1)
     v2 = str(version2)
 
-    # Process version2
+    # Process version2 -- note that order matters, and two-char prefixes have to be handled first
     valid = None
-    if   v2.startswith('>'):  valid = [1]
+    if   v2.startswith('<='): valid = [0,-1]
     elif v2.startswith('>='): valid = [0,1]
-    elif v2.startswith('='):  valid = [0]
     elif v2.startswith('=='): valid = [0]
     elif v2.startswith('~='): valid = [-1,1]
     elif v2.startswith('!='): valid = [-1,1]
-    elif v2.startswith('<='): valid = [0,-1]
     elif v2.startswith('<'):  valid = [-1]
+    elif v2.startswith('>'):  valid = [1]
+    elif v2.startswith('='):  valid = [0]
+    elif v2.startswith('!'):  valid = [-1,1]
+    elif v2.startswith('~'): # pragma: no cover
+        errormsg = 'Loose version pinning is not supported; for not, use "~="'
+        raise ValueError(errormsg)
+    
     v2 = v2.lstrip('<>=!~')
 
     # Do comparison
@@ -339,11 +344,11 @@ def getcaller(frame=2, tostring=True, includelineno=False, includeline=False, re
         lineno = result[frame][2]
         if tostring:
             output = f'{fname}'
-            if includelineno:
+            if includelineno: # pragma: no cover
                 output += f', line {lineno}'
         else:
             output = {'filename':fname, 'lineno':lineno}
-            if includeline:
+            if includeline: # pragma: no cover
                 try:
                     with open(fname) as f:
                         lines = f.read().splitlines()
@@ -434,7 +439,7 @@ def metadata(outfile=None, comments=None, pipfreeze=True, user=True, caller=True
     md.update(kwargs)
     
     if outfile is not None:
-        outfile = scf.makefilepath(outfile, makedirs=True)
+        outfile = scf.makepath(outfile, makedirs=True)
         scf.savejson(outfile, md)
     
     if tostring:
@@ -442,6 +447,15 @@ def metadata(outfile=None, comments=None, pipfreeze=True, user=True, caller=True
     
     return md
 
+
+def _metadata_to_objdict(md):
+    ''' Convert a metadata dictionary into an objdict -- descend two levels but no deeper '''
+    md = sco.objdict(md)
+    for k,v in md.items():
+        if isinstance(v, dict):
+            md[k] = sco.objdict(v)
+    return md
+        
 
 def loadmetadata(filename, load_all=False, die=True):
     '''
@@ -483,7 +497,7 @@ def loadmetadata(filename, load_all=False, die=True):
 
         # Usual case, can find metadata and is PNG
         if is_png and (load_all or _metadataflag in keys):
-            if load_all:
+            if load_all: # pragma: no cover
                 md = im.info
             else:
                 jsonstr = im.info[_metadataflag]
@@ -537,13 +551,13 @@ def loadmetadata(filename, load_all=False, die=True):
             else:
                 print(errormsg)
     
[…]
