+++ tests/test_utils.py
@@ -146,28 +146,6 @@ def test_tryexcept():
     return tryexc
 
 
-def test_versions():
-    sc.heading('Testing freeze, compareversions, and require')
-
-    # Test freeze
-    assert 'numpy' in sc.freeze()
-
-    # Test compareversions
-    assert sc.compareversions(np, '>1.0')
-
-    # Test require
-    sc.require('numpy')
-    sc.require(numpy='')
-    sc.require(reqs={'numpy':'1.19.1', 'matplotlib':'3.2.2'})
-    sc.require('numpy>=1.19.1', 'matplotlib==3.2.2', die=False)
-    data, _ = sc.require(numpy='1.19.1', matplotlib='==4.2.2', die=False, detailed=True)
-    with pytest.raises(ModuleNotFoundError): sc.require('matplotlib==99.23')
-    with pytest.raises(ModuleNotFoundError): sc.require('not_a_valid_module')
-
-    print('↑↑↑ Will print warnings')
-
-    return data
-
 
 #%% Type functions
 
@@ -379,7 +357,6 @@ if __name__ == '__main__':
     uid       = test_uuid()
     traceback = test_traceback()
     tryexc    = test_tryexcept()
-    versions  = test_versions()
 
     # Type
     plist     = test_promotetolist()
