+++ sciris/sc_printing.py
@@ -73,9 +73,9 @@ def objatt(obj, strlen=18, ncol=3):
     ''' Return a sorted string of object attributes for the Python __repr__ method '''
     if   hasattr(obj, '__dict__'):  oldkeys = sorted(obj.__dict__.keys())
     elif hasattr(obj, '__slots__'): oldkeys = sorted(obj.__slots__)
-    else:                           oldkeys = []
+    else:                           oldkeys = [] # pragma: no cover
     if len(oldkeys): output = createcollist(oldkeys, 'Attributes', strlen = 18, ncol = 3)
-    else:            output = ''
+    else:            output = '' # pragma: no cover
     return output
 
 
@@ -167,7 +167,7 @@ def prepr(obj, maxlen=None, maxitems=None, skip=None, dividerchar='—', divider
 
     # Wrap entire process in a try-except in case it fails
     try:
-        if not (hasattr(obj, '__dict__') or hasattr(obj, '__slots__')):
+        if not (hasattr(obj, '__dict__') or hasattr(obj, '__slots__')): # pragma: no cover
             # It's a plain object
             labels = [f'{type(obj)}']
             values = [repr_fn(obj)]
@@ -176,7 +176,7 @@ def prepr(obj, maxlen=None, maxitems=None, skip=None, dividerchar='—', divider
                 labels = sorted(set(obj.__dict__.keys()) - set(skip))  # Get the dict attribute keys
             else:
                 labels = sorted(set(obj.__slots__) - set(skip))  # Get the slots attribute keys
-            if debug:
+            if debug: # pragma: no cover
                 print(f'Working on {len(labels)} entries...')
 
             if len(labels):
@@ -185,11 +185,11 @@ def prepr(obj, maxlen=None, maxitems=None, skip=None, dividerchar='—', divider
                     labels = labels[:maxitems]
                 values = []
                 for a,attr in enumerate(labels):
-                    if debug:
+                    if debug: # pragma: no cover
                         print(f'  Working on attribute {a}: {attr}...')
                     if (time.time() - T) < maxtime:
                         try: value = repr_fn(getattr(obj, attr))
-                        except: value = 'N/A'
+                        except: value = 'N/A' # pragma: no cover
                         values.append(value)
                     else:
                         labels = labels[:a]
@@ -291,7 +291,7 @@ def indent(prefix=None, text=None, suffix='\n', n=0, pretty=False, width=70, **k
     New in version 1.3.1: more flexibility in arguments
     '''
     # If "prefix" is given but text isn't, swap them
-    if text is None and prefix is not None:
+    if text is None and prefix is not None: # pragma: no cover
         text, prefix = prefix, text
 
     # Handle prefix and width
@@ -351,6 +351,7 @@ def sigfig(x, sigfigs=4, SI=False, sep=False, keepints=False):
     output = []
 
     islist = scu.isiterable(x)
+    istuple = isinstance(x, tuple)
     xlist = x if islist else scu.tolist(x)
     for x in xlist:
         suffix = ''
@@ -389,7 +390,7 @@ def sigfig(x, sigfigs=4, SI=False, sep=False, keepints=False):
         except: # pragma: no cover
             output.append(scu.flexstr(x))
     if islist:
-        if isinstance(x, tuple):
+        if istuple:
             output = tuple(output)
         return output
     else:
@@ -406,7 +407,7 @@ def printmean(data, stds=2, mean_sf=None, err_sf=None, doprint=True, **kwargs):
     Args:
         data (array): the data to summarize
         stds (int): the number of multiples of the standard deviation to show (default 2; can also use 1)
[…]
