### 2023-04-13 sciris/sciris#415 @aflaxman
this is a helpful improvement, thanks.
