### 2023-04-13 sciris/sciris#437 @aflaxman
This addresses almost all of my concerns.  The one point I would prefer a little more change on is the tests --- for users who are not experienced in python testing, I think it would be helpful to have a section in the docs that says (1) automatic tests exist, (2) this can confirm that things are working on your installation, and (3) here is how to run them and see that they have all passed.
