+++ sciris/sc_fileio.py
@@ -120,7 +120,7 @@ def _load_filestr(filename, folder):
                         raise exc(gziperror) from E
         else:
             exc = type(E)
-            errormsg = 'sc.load(): Could not open the file string for an unknown reason; see error above for details'
+            errormsg = 'sc.load(): Could not open the # pragma: no cover file string for an unknown reason; see error above for details'
             raise exc(errormsg) from E
     return filestr
 
@@ -177,7 +177,7 @@ def load(filename=None, folder=None, verbose=False, die=None, remapping=None,
     # If it loaded but with errors, print them here
     if isinstance(obj, Failed):
         print(_unpicklingerror(filename))
-    elif verbose:
+    elif verbose: # pragma: no cover
         print(f'Object loaded from "{filename}"')
 
     return obj
@@ -316,7 +316,7 @@ def save(filename='default.obj', obj=None, folder=None, method='pickle', compres
             serialize(fileobj, obj, success, **kwargs)
     
     else:
-        if tobytes:
+        if tobytes: # pragma: no cover
             filecontext = closing(bytestream)
         else:
             filecontext = open(filename, 'wb')
@@ -331,11 +331,11 @@ def save(filename='default.obj', obj=None, folder=None, method='pickle', compres
             # File extension can be anything
             with filecontext as fileobj:
                 serialize(fileobj, obj, success, **kwargs)
-        else:
+        else: # pragma: no cover
             errormsg = f"Invalid compression format '{compression}': must be 'gzip', 'zstd', or 'none'"
             raise ValueError(errormsg)
 
-    if verbose and filename:
+    if verbose and filename: # pragma: no cover
         print(f'Object saved to "{filename}"')
 
     if filename:
@@ -443,7 +443,7 @@ def savetext(filename=None, string=None, **kwargs):
     '''
     is_array = scu.isarray(string)
     if isinstance(string, list):
-        string = '\n'.join(string) # Convert from list to string)
+        string = '\n'.join(str(string)) # Convert from list to string)
     elif not is_array and not scu.isstring(string):
         string = str(string)
     filename = makefilepath(filename=filename, makedirs=True)
@@ -527,7 +527,7 @@ def savezip(filename=None, filelist=None, data=None, folder=None, basename=True,
         filelist (list): the list of files to compress
         data (dict): if supplied, instead of files, write this data instead (must be a dictionary of filename keys and data values)
         folder (str): optional additional folder for the filename
-        basename (bool): whether to use only the file's basename as the name inside the zip file
+        basename (bool): whether to use only the file's basename as the name inside the zip file (otherwise, store folder info)
         sanitizepath (bool): whether to sanitize the path prior to saving
         tobytes (bool): if data is provided, convert it automatically to bytes (otherwise, up to the user)
         verbose (bool): whether to print progress
@@ -548,7 +548,7 @@ def savezip(filename=None, filelist=None, data=None, folder=None, basename=True,
     fullpath = makefilepath(filename=filename, folder=folder, sanitize=sanitizepath, makedirs=True)
     filelist = scu.tolist(filelist)
     if data is not None:
-        if not isinstance(data, dict):
+        if not isinstance(data, dict): # pragma: no cover
             errormsg = 'Data has invalid format: must be a dictionary of filename keys and data values'
             raise ValueError(errormsg)
 
@@ -558,7 +558,7 @@ def savezip(filename=None, filelist=None, data=None, folder=None, basename=True,
             for thisfile in filelist:
[…]
