+++ sciris/sc_plotting.py
@@ -48,7 +48,7 @@ def fig3d(num=None, returnax=False, figkwargs=None, axkwargs=None, **kwargs):
     axkwargs = scu.mergedicts(axkwargs)
 
     fig,ax = ax3d(returnfig=True, figkwargs=figkwargs, **axkwargs)
-    if returnax:
+    if returnax: # pragma: no cover
         return fig,ax
     else:
         return fig
@@ -72,7 +72,7 @@ def ax3d(fig=None, ax=None, returnfig=False, silent=False, elev=None, azim=None,
     if fig is None:
         if ax is None:
             fig = pl.figure(**figkwargs) # It's necessary to have an open figure or else the commands won't work
-        else:
+        else: # pragma: no cover
             fig = ax.figure
             silent = False
     else:
@@ -81,13 +81,13 @@ def ax3d(fig=None, ax=None, returnfig=False, silent=False, elev=None, azim=None,
     # Create and initialize the axis
     if ax is None:
         ax = fig.add_subplot(nrows, ncols, index, projection='3d', **axkwargs)
-    if elev is not None or azim is not None:
+    if elev is not None or azim is not None: # pragma: no cover
         ax.view_init(elev=elev, azim=azim)
-    if silent:
+    if silent: # pragma: no cover
         pl.close(fig)
     if returnfig:
         return fig,ax
-    else:
+    else: # pragma: no cover
         return ax
 
 
@@ -122,7 +122,7 @@ def plot3d(x, y, z, c=None, fig=None, ax=None, returnfig=False, figkwargs=None, 
 
     ax.plot(x, y, z, **plotkwargs)
 
-    if returnfig:
+    if returnfig: # pragma: no cover
         return fig,ax
     else:
         return ax
@@ -159,7 +159,7 @@ def scatter3d(x, y, z, c=None, fig=None, returnfig=False, figkwargs=None, axkwar
 
     ax.scatter(x, y, z, c=c, **plotkwargs)
 
-    if returnfig:
+    if returnfig: # pragma: no cover
         return fig,ax
     else:
         return ax
@@ -203,14 +203,14 @@ def surf3d(data, x=None, y=None, fig=None, returnfig=False, colorbar=True, figkw
 
     if x.ndim == 1 or y.ndim == 1:
         X,Y = np.meshgrid(x, y)
-    else:
+    else: # pragma: no cover
         X,Y = x,y
 
     surf = ax.plot_surface(X, Y, data, **plotkwargs)
     if colorbar:
         fig.colorbar(surf)
 
-    if returnfig:
+    if returnfig: # pragma: no cover
         return fig,ax
     else:
         return ax
@@ -262,7 +262,7 @@ def bar3d(data, fig=None, returnfig=False, cmap='viridis', figkwargs=None, axkwa
             dz.append(data[i,j])
     ax.bar3d(x=x, y=y, z=z, dz=dz, **plotkwargs)
 
-    if returnfig:
+    if returnfig: # pragma: no cover
         return fig,ax
     else:
         return ax
@@ -352,7 +352,7 @@ def stackedbar(x=None, values=None, colors=None, labels=None, transpose=False,
     for i in range(nstack):
         if labels is not None:
             label = labels[i]
-        else:
+        else: # pragma: no cover
             label = None
         
         h = values[i+1,:]
@@ -360,7 +360,7 @@ def stackedbar(x=None, values=None, colors=None, labels=None, transpose=False,
         kw = dict(facecolor=colors[i], label=label, **kwargs)
         if not barh:
             artist = pl.bar(x=x, height=h, bottom=b, **kw)
-        else:
+        else: # pragma: no cover
             artist = pl.barh(y=x, width=h, left=b, **kw)
[…]
