+++ sciris/sc_utils.py
@@ -115,7 +115,7 @@ def fast_uuid(which=None, length=None, n=1, secure=False, forcelist=False, safet
             raise ValueError(errormsg)
 
     # Secure uses system random which is secure, but >10x slower
-    if secure:
+    if secure: # pragma: no cover
         choices_func = rnd.SystemRandom().choices
     else:
         choices_func = rnd.choices
@@ -315,7 +315,7 @@ def pp(obj, jsonify=True, doprint=None, output=False, verbose=False, **kwargs):
         except Exception as E:
             if verbose: print(f'Could not jsonify object ("{str(E)}"), printing default...')
             toprint = obj # If problems are encountered, just return the object
-    else:
+    else: # pragma: no cover
         toprint = obj
 
     # Decide what to do with object
@@ -344,13 +344,13 @@ def sha(obj, encoding='utf-8', digest=False):
     '''
     if not isstring(obj): # Ensure it's actually a string
         string = repr(obj)
-    else:
+    else: # pragma: no cover
         string = obj
     needsencoding = isinstance(string, str)
     if needsencoding: # If it's unicode, encode it to bytes first
         string = string.encode(encoding)
     output = hashlib.sha224(string)
-    if digest:
+    if digest: # pragma: no cover
         output = output.hexdigest()
     return output
 
@@ -414,10 +414,10 @@ def getplatform(expected=None, platform=None, die=False):
     # Handle output
     if expected is not None:
         output = (expected.lower() in mapping[plat]) # Check if it's as expecte
-        if not output and die:
+        if not output and die: # pragma: no cover
             errormsg = f'System is "{plat}", not "{expected}"'
             raise EnvironmentError(errormsg)
-    else:
+    else: # pragma: no cover
         output = plat
     return output
 
@@ -511,9 +511,9 @@ def urlopen(url, filename=None, save=False, headers=None, params=None, data=None
     if prefix is not None:
         if not full_url.startswith(prefix):
             full_url = prefix + '://' + full_url # Leaves https alone, but adds http:// otherwise
-    if params is not None:
+    if params is not None: # pragma: no cover
         full_url = full_url + '?' + up.urlencode(params)
-    if data is not None:
+    if data is not None: # pragma: no cover
         data = up.urlencode(data).encode(encoding='utf-8', errors='ignore')
 
     if verbose: print(f'Downloading {url}...')
@@ -532,7 +532,7 @@ def urlopen(url, filename=None, save=False, headers=None, params=None, data=None
                 print(errormsg)
 
     # Set filename -- from https://stackoverflow.com/questions/31804799/how-to-get-pdf-filename-with-python-requests
-    if filename is None and save:
+    if filename is None and save: # pragma: no cover
         headers = dict(response.getheaders())
         string = "Content-Disposition"
         if string in headers.keys():
@@ -544,7 +544,7 @@ def urlopen(url, filename=None, save=False, headers=None, params=None, data=None
         if verbose: print(f'Saving to {filename}...')
         filename = scf.makefilepath(filename, makedirs=True)
         if isinstance(output, bytes):
-            with open(filename, 'wb') as f:
+            with open(filename, 'wb') as f: # pragma: no cover
                 f.write(output)
         else:
             with open(filename, 'w') as f:
@@ -553,7 +553,7 @@ def urlopen(url, filename=None, save=False, headers=None, params=None, data=None
 
     if verbose:
         T.toc(f'Time to download {url}')
-    if return_response:
+    if return_response: # pragma: no cover
         output = response
 
     return output
@@ -624,7 +624,7 @@ def download(url, *args, filename=None, save=True, parallel=True, die=True, verb
         for url,filename in zip(urls, filenames):
             try:
[…]

[…]
