### 2023-04-13 #5076 @cliffckerr
> I created this pull request to make it easier to focus on the changes in this revision: [aflaxman/sciris_paper#1](https://github.com/aflaxman/sciris_paper/pull/1)
>
> Was there an easier way for me to see the full diff?

Sorry, I should've done that myself as a pull request rather than just making the changes directly. I do think in future it would be nice if authors at JOSS had the option of keeping the manuscript files + manuscript issues in a separate repo, rather than mixed in with the software package itself, and I like your implication @aflaxman that second round review could be reviewing a pull request in that repo!
