### 2023-04-13 #5076 @danielskatz
You can put the paper in a branch, but we (JOSS) don't want to keep track of multiple repos
