+++ docs/requirements.txt
@@ -5,5 +5,5 @@ pandoc
 pypandoc
 sphinx>=1.8
 sphinx-autodoc-typehints
-sphinx-rtd-theme
+pydata-sphinx-theme
 recommonmark
