+++ docs/_templates/custom-class-template.rst
@@ -1,11 +1,10 @@
-{{ fullname | escape | underline}}
+{{ objname | escape | underline}}
 
 .. currentmodule:: {{ module }}
 
 .. autoclass:: {{ objname }}
    :members:
    :show-inheritance:
-   :inherited-members:
    :special-members: __call__, __add__, __mul__
 
    {% block attributes %}
@@ -23,13 +22,6 @@
    {% if methods %}
    .. rubric:: {{ _('Methods') }}
 
-   .. autosummary::
-      :nosignatures:
-   {% for item in methods %}
-      {%- if not item.startswith('_') %}
-      ~{{ name }}.{{ item }}
-      {%- endif -%}
-   {%- endfor %}
    {% endif %}
    {% endblock %}
 
