+++ docs/conf.py
@@ -50,7 +50,10 @@ extensions = [
 #     'members': None,
 # }
 
-autodoc_mock_imports = []
+# Actually use this to skip modules! Confusing...
+autodoc_mock_imports = [
+    'ansicolors',
+]
 napoleon_google_docstring = True
 
 # Configure autosummary
@@ -67,7 +70,7 @@ templates_path = ["_templates"]
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "**.ipynb_checkpoints"]
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "**.ipynb_checkpoints", "ansicolors.py"]
 
 
 # -- Options for HTML output -------------------------------------------------
