+++ docs/conf.py
@@ -72,9 +72,8 @@ html_theme_options = {
             "icon": "fab fa-github-square",
         },
     ],
-    # "external_links": [{"name": "Guides", "url": "https://networkx.org/nx-guides/"}],
-    # "navbar_end": ["theme-switcher", "navbar-icon-links", "version"],
     "navbar_end": ["theme-switcher", "navbar-icon-links", "search-field"],
+    "search_bar_text": "Search",
     "secondary_sidebar_items": ["page-toc", "edit-this-page"],
     "header_links_before_dropdown": 7,
 }
