+++ docs/overview.rst
@@ -1,1 +1,130 @@
-.. include:: ../README.rst
+Getting started
+===============
+
+.. image:: https://badgen.net/pypi/v/sciris/?color=blue
+ :target: https://pypi.com/project/sciris
+
+.. image:: https://static.pepy.tech/personalized-badge/sciris?period=total&units=international_system&left_color=grey&right_color=yellow&left_text=Downloads
+ :target: https://pepy.tech/project/sciris
+
+.. image:: https://img.shields.io/pypi/l/sciris.svg
+ :target: https://github.com/sciris/sciris/blob/main/LICENSE
+
+.. image:: https://github.com/sciris/sciris/actions/workflows/tests.yaml/badge.svg
+ :target: https://github.com/sciris/sciris/actions/workflows/tests.yaml?query=workflow
+
+
+What is Sciris?
+---------------
+
+Glad you asked! **Sciris** (http://sciris.org) is a library of tools that can help make writing scientific Python code easier and more pleasant. Built on top of `NumPy <https://numpy.org/>`__ and `Matplotlib <https://matplotlib.org/>`__, Sciris provides functions covering a wide range of common math, file I/O, and plotting operations. This means you can get more done with less code, and spend less time looking things up on StackOverflow. It was originally written to help epidemiologists and neuroscientists focus on doing science, rather than on writing coding, but Sciris is applicable across scientific domains.
+
+Sciris is available on `PyPi <https://pypi.org/project/sciris/>`__ (``pip install sciris``) and `GitHub <https://github.com/sciris/sciris>`__. Full documentation is available at http://docs.sciris.org. If you have questions, feature suggestions, or would like some help getting started, please reach out to us at info@sciris.org.
+
+
+Highlights
+~~~~~~~~~~
+Some highlights of Sciris (``import sciris as sc``):
+
+- **Powerful containers** – The ``sc.odict`` class is what ``OrderedDict`` (almost) could have been, allowing reference by position or key, casting to a NumPy array, sorting and enumeration functions, etc.
+- **Array operations** – Want to find the indices of an array that match a certain value or condition? ``sc.findinds()`` will do that. How about just the nearest value, regardless of exact match? ``sc.findnearest()``. What about the last matching value? ``sc.findlast()``. Yes, you could do ``np.nonzero()[0][-1]`` instead, but ``sc.findlast()`` is easier to read, type, and remember, and handles edge cases more elegantly.
+- **File I/O** – One-liner functions for saving and loading text, JSON, spreadsheets, or even arbitrary Python objects.
+- **Plotting recipes** – Simple functions for mapping sequential or qualitative data onto colors, manipulating color data, and updating axis limits and tick labels, plus several new colormaps.
+
+
+I'm not convinced.
+~~~~~~~~~~~~~~~~~~
+That's OK. Perhaps you'd be interested in seeing what a script that performs tasks like parallelization, saving and loading files, and 3D plotting looks like when written in "`vanilla Python <https://github.com/sciris/sciris/blob/main/tests/showcase_vanilla.py>`_" (left) compared to `using Sciris <https://github.com/sciris/sciris/blob/main/tests/showcase_sciris.py>`_ (right):
+
+|Sciris showcase|
+
+Both of these do the same thing, but the plain Python version (left) requires 50% more lines of code to produce the same graph as Sciris (right):
+
+|Sciris output|
+
+
+Where did Sciris come from?
+~~~~~~~~~~~~~~~~~~~~~~~~~~~
[…]
