+++ docs/api_reference.rst
@@ -1,0 +1,26 @@
+=============
+API reference
+=============
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-module-template.rst
+   :recursive:
+   :nosignatures:
+
+   sciris.sc_asd
+   sciris.sc_colors
+   sciris.sc_dataframe
+   sciris.sc_datetime
+   sciris.sc_fileio
+   sciris.sc_math
+   sciris.sc_nested
+   sciris.sc_odict
+   sciris.sc_parallel
+   sciris.sc_plotting
+   sciris.sc_printing
+   sciris.sc_profiling
+   sciris.sc_settings
+   sciris.sc_utils
+   sciris.sc_version
+   sciris.sc_versioning
