+++ docs/api/other.rst
@@ -1,0 +1,15 @@
+===============
+Other utilities
+===============
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-module-template.rst
+   :recursive:
+   :nosignatures:
+
+   sciris.sc_utils
+   sciris.sc_datetime
+   sciris.sc_nested
+   sciris.sc_settings
+   
