+++ docs/api/containers.rst
@@ -1,0 +1,12 @@
+=============
+Containers
+=============
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-module-template.rst
+   :recursive:
+   :nosignatures:
+
+    sciris.sc_odict
+    sciris.sc_dataframe
