+++ docs/api/printing.rst
@@ -1,0 +1,13 @@
+=====================
+Printing and plotting
+=====================
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-module-template.rst
+   :recursive:
+   :nosignatures:
+
+   sciris.sc_printing
+   sciris.sc_plotting
+   sciris.sc_colors
