+++ docs/api/math.rst
@@ -1,0 +1,12 @@
+====================
+Math and array tools
+====================
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-module-template.rst
+   :recursive:
+   :nosignatures:
+
+   sciris.sc_math
+   sciris.sc_asd
