+++ docs/api/parallel.rst
@@ -1,0 +1,12 @@
+=============================
+Parallelization and profiling
+=============================
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-module-template.rst
+   :recursive:
+   :nosignatures:
+
+   sciris.sc_parallel
+   sciris.sc_profiling
