+++ docs/api/files.rst
@@ -1,0 +1,12 @@
+=====================
+Files and versioning
+=====================
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-module-template.rst
+   :recursive:
+   :nosignatures:
+
+   sciris.sc_fileio
+   sciris.sc_versioning
