+++ sciris/sc_utils.py
@@ -243,7 +243,7 @@ def dcp(obj, die=True, verbose=True):
         die (bool): if False, fall back to copy()
         verbose (bool): if die is False, then print a warning if deepcopy() fails
 
-    New in version 2.0.0: default die=True instead of False
+    *New in version 2.0.0:* default die=True instead of False
     '''
     try:
         output = copy.deepcopy(obj)
@@ -261,7 +261,7 @@ def cp(obj, die=True, verbose=True):
 
     Almost identical to ``copy.copy()``, but optionally allow failures
 
-    New in version 2.0.0: default die=True instead of False
+    *New in version 2.0.0:* default die=True instead of False
     '''
     try:
         output = copy.copy(obj)
@@ -305,7 +305,7 @@ def pp(obj, jsonify=True, doprint=None, output=False, verbose=False, **kwargs):
         verbose (bool): whether to show warnings when jsonifying the object
         kwargs  (dict): passed to ``pprint.pprint()``
 
-    New in version 1.3.1: output argument
+    *New in version 1.3.1:* output argument
     '''
 
     # Get object
@@ -370,7 +370,7 @@ def getuser():
     
     Alias to ``getpass.getuser()`` -- see https://docs.python.org/3/library/getpass.html#getpass.getuser
     
-    New in version 2.2.0.
+    *New in version 2.2.0.*
     '''
     return getpass.getuser()
 
@@ -448,7 +448,7 @@ def asciify(string, form='NFKD', encoding='ascii', errors='ignore', **kwargs):
     **Example**:
         sc.asciify('föö→λ ∈ ℝ') # Returns 'foo  R'
     
-    New in version 2.0.1.
+    *New in version 2.0.1.*
     '''
     normalized = unicodedata.normalize(form, string) # First, normalize Unicode encoding
     encoded = normalized.encode(encoding, errors) # Then, convert to ASCII
@@ -488,9 +488,9 @@ def urlopen(url, filename=None, save=False, headers=None, params=None, data=None
         sc.urlopen('http://sciris.org', filename='sciris.html') # Save to file sciris.html
         sc.urlopen('http://sciris.org', save=True, headers={'User-Agent':'Custom agent'}) # Save to the default filename (here, sciris.org), with headers
 
-    | New in version 2.0.0: renamed from ``wget`` to ``urlopen``; new arguments
-    | New in version 2.0.1: creates folders by default if they do not exist
-    | New in version 2.0.4: "prefix" argument, e.g. prepend "http://" if not present
+    | *New in version 2.0.0:* renamed from ``wget`` to ``urlopen``; new arguments
+    | *New in version 2.0.1:* creates folders by default if they do not exist
+    | *New in version 2.0.4:* "prefix" argument, e.g. prepend "http://" if not present
     '''
     from urllib import request as ur # Need to import these directly, not via urllib
     from urllib import parse as up
@@ -585,8 +585,8 @@ def download(url, *args, filename=None, save=True, parallel=True, die=True, verb
         sc.download({'http://sciris.org':'sciris.html', 'http://covasim.org':'covasim.html'}) # Downlaod two and save to disk
         sc.download(['http://sciris.org', 'http://covasim.org'], filename=['sciris.html', 'covasim.html']) # Ditto
 
-    | New in version 2.0.0.
-    | New in version 2.2.0: "die" argument
+    | *New in version 2.0.0.*
+    | *New in version 2.2.0:* "die" argument
     '''
[…]
