+++ sciris/sc_printing.py
@@ -146,7 +146,7 @@ def prepr(obj, maxlen=None, maxitems=None, skip=None, dividerchar='—', divider
         die (bool): whether to raise an exception if an error is encountered
         debug (bool): print out detail during string construction
     
-    New in version 2.2.0: "debug" argument
+    *New in version 2.2.0:* "debug" argument
     '''
 
     # Decide how to handle representation function -- repr is dangerous since can lead to recursion
@@ -288,7 +288,7 @@ def indent(prefix=None, text=None, suffix='\n', n=0, pretty=False, width=70, **k
 
         print('my fave is: ' + sc.indent(text=rand(100), n=12))
 
-    New in version 1.3.1: more flexibility in arguments
+    *New in version 1.3.1:* more flexibility in arguments
     '''
     # If "prefix" is given but text isn't, swap them
     if text is None and prefix is not None: # pragma: no cover
@@ -346,7 +346,7 @@ def sigfig(x, sigfigs=4, SI=False, sep=False, keepints=False):
         vals = np.random.rand(5)
         sc.sigfig(vals, sigfigs=3)
     
-    New in version 2.2.0: changed default number of significant figures from 5 to 4; return list rather than tuple
+    *New in version 2.2.0:* changed default number of significant figures from 5 to 4; return list rather than tuple
     '''
     output = []
 
@@ -417,7 +417,7 @@ def printmean(data, stds=2, mean_sf=None, err_sf=None, doprint=True, **kwargs):
         data = [1210, 1072, 1722, 1229, 1902]
         sc.printmean(data) # Returns 1430 ± 320
     
-    New in version 2.2.0.
+    *New in version 2.2.0.*
     '''
     vsf = mean_sf # vsf = "value significant figures"
     esf = err_sf if err_sf is not None else 2
@@ -464,7 +464,7 @@ def printmedian(data, ci=95, sf=3, doprint=True, **kwargs):
         data = [1210, 1072, 1722, 1229, 1902]
         sc.printmedian(data, 80) # Returns '1230 (80.0% CI: 1130, 1830)'
     
-    New in version 2.2.0.
+    *New in version 2.2.0.*
     '''
     # Handle quantiles
     if ci is None: # pragma: no cover
@@ -532,7 +532,7 @@ def humanize_bytes(bytesize, decimals=3):
         
     See the humansize library for more flexibility.
     
-    New in version 2.2.0.
+    *New in version 2.2.0.*
     '''
     # Convert to string
     factor = 1
@@ -569,8 +569,8 @@ def printarr(arr, fmt=None, colsep='  ', vsep='—', decimals=2, doprint=True, d
         sc.printarr(numeric)
         sc.printarr(mixed)
 
-    New in version 2.0.3: "fmt", "colsep", "vsep", "decimals", and "dtype" arguments
-    New in version 2.2.0: "doprint" argument
+    *New in version 2.0.3:* "fmt", "colsep", "vsep", "decimals", and "dtype" arguments
+    *New in version 2.2.0:* "doprint" argument
     '''
     from . import sc_math as scm # To avoid circular import
     
@@ -743,7 +743,7 @@ def colorize(color=None, string=None, doprint=None, output=False, enable=True, s
 
     To get available colors, type ``sc.colorize(showhelp=True)``.
 
-    | New in version 1.3.1: "doprint" argument; ansicolors shortcut
+    | *New in version 1.3.1:* "doprint" argument; ansicolors shortcut
     '''
 
     # Handle short-circuit case
@@ -881,7 +881,7 @@ def heading(string=None, *args, color=None, divider=None, spaces=None, spacesaft
         sc.heading('This is a heading')
[…]
