### 2023-04-14 69a1610
done with plotting
