+++ sciris/sc_plotting.py
@@ -299,7 +299,7 @@ def stackedbar(x=None, values=None, colors=None, labels=None, transpose=False,
         sc.stackedbar(values, labels=['bottom','middle','top'])
         pl.legend()
     
-    New in version 2.0.4.
+    *New in version 2.0.4.*
     '''
     from . import sc_colors as scc # To avoid circular import
     
@@ -394,7 +394,7 @@ def boxoff(ax=None, which=None, removeticks=True):
         pl.scatter(np.arange(100), pl.rand(100))
         sc.boxoff('top, bottom')
 
-    New in version 1.3.3: ability to turn off multiple spines; removed "flipticks" arguments
+    *New in version 1.3.3:* ability to turn off multiple spines; removed "flipticks" arguments
     '''
     # Handle axes
     if isinstance(ax, (str, list)): # Swap input arguments # pragma: no cover
@@ -529,9 +529,9 @@ def commaticks(ax=None, axis='y', precision=2, cursor_precision=0):
 
     See http://stackoverflow.com/questions/25973581/how-to-format-axis-number-format-to-thousands-with-a-comma-in-matplotlib
 
-    | New in version 1.3.0: ability to use non-comma thousands separator
-    | New in version 1.3.1: added "precision" argument
-    | New in version 2.0.0: ability to set x and y axes simultaneously
+    | *New in version 1.3.0:* ability to use non-comma thousands separator
+    | *New in version 1.3.1:* added "precision" argument
+    | *New in version 2.0.0:* ability to set x and y axes simultaneously
     '''
     def commaformatter(x, pos=None): # pragma: no cover
         interval = thisaxis.get_view_interval()
@@ -630,9 +630,9 @@ def getrowscols(n, nrows=None, ncols=None, ratio=1, make=False, tight=True, remo
         nrows,ncols = sc.get_rows_cols(100, ratio=0.5) # Returns 8,13 since rows are prioritized
         fig,axs     = sc.getrowscols(37, make=True) # Create 7x6 subplots, using the alias
 
-    | New in version 1.0.0.
-    | New in version 1.2.0: "make", "tight", and "remove_extra" arguments
-    | New in version 1.3.0: alias without underscores
+    | *New in version 1.0.0.*
+    | *New in version 1.2.0:* "make", "tight", and "remove_extra" arguments
+    | *New in version 1.3.0:* alias without underscores
     '''
 
     # Simple cases -- calculate the one missing
@@ -677,7 +677,7 @@ def figlayout(fig=None, tight=True, keep=False, **kwargs):
         fig,axs = sc.get_rows_cols(37, make=True, tight=False) # Create 7x6 subplots, squished together
         sc.figlayout(bottom=0.3)
 
-    New in version 1.2.0.
+    *New in version 1.2.0.*
     '''
     if isinstance(fig, bool): # pragma: no cover
         fig = None
@@ -714,7 +714,7 @@ def maximize(fig=None, die=False):  # pragma: no cover
         pl.plot([2,3,5])
         sc.maximize()
 
-    New in version 1.0.0.
+    *New in version 1.0.0.*
     '''
     backend = pl.get_backend().lower()
     if fig is not None:
@@ -875,7 +875,7 @@ class ScirisDateFormatter(mpl.dates.ConciseDateFormatter):
 
     See :func:`dateformatter` for explanation of arguments.
 
-    New in version 1.3.0.
+    *New in version 1.3.0.*
     '''
 
     def __init__(self, locator, formats=None, zero_formats=None, show_offset=False, show_year=True, **kwargs):
[…]
