+++ sciris/sc_fileio.py
@@ -159,8 +159,8 @@ def load(filename=None, folder=None, verbose=False, die=None, remapping=None,
         old = sc.load('my-old-file.obj', remapping={'foo.Bar':cat.Mat}) # If loading a saved object containing a reference to foo.Bar that is now cat.Mat
         old = sc.load('my-old-file.obj', remapping={'foo.Bar':('cat', 'Mat')}) # Equivalent to the above
 
-    | New in version 1.1.0: "remapping" argument
-    | New in version 1.2.2: ability to load non-gzipped pickles; support for dill; arguments passed to loader
+    | *New in version 1.1.0:* "remapping" argument
+    | *New in version 1.2.2:* ability to load non-gzipped pickles; support for dill; arguments passed to loader
     '''
     # Load the file
     filestr = _load_filestr(filename, folder)
@@ -247,11 +247,11 @@ def save(filename='default.obj', obj=None, folder=None, method='pickle', compres
     
     See also :func:`zsave()` (identical except defaults to zstandard compression).
 
-    | New in version 1.1.1: removed Python 2 support.
-    | New in version 1.2.2: automatic swapping of arguments if order is incorrect; correct passing of arguments
-    | New in version 2.0.4: "die" argument for saving as dill
-    | New in version 2.1.0: "zstandard" compression method
-    | New in version 2.2.0: "allow_empty" argument; removed "args"
+    | *New in version 1.1.1:* removed Python 2 support.
+    | *New in version 1.2.2:* automatic swapping of arguments if order is incorrect; correct passing of arguments
+    | *New in version 2.0.4:* "die" argument for saving as dill
+    | *New in version 2.1.0:* "zstandard" compression method
+    | *New in version 2.2.0:* "allow_empty" argument; removed "args"
     '''
 
     def serialize(fileobj, obj, success, **kwargs):
@@ -358,7 +358,7 @@ def zsave(*args, compression='zstd', **kwargs):
     Note: there is no matching function ``sc.zload()`` since :func:`load()` will
     automatically try loading zstandard.
     
-    New in version 2.1.0.    
+    *New in version 2.1.0.*
     '''
     return save(*args, compression=compression, **kwargs)
 
@@ -378,7 +378,7 @@ def loadstr(string, **kwargs):
         obj2 = sc.loadstr(bytestring)
         assert obj == obj2
     
-    | New in version 2.2.0: uses ``sc.load()`` for more robustness
+    | *New in version 2.2.0:* uses ``sc.load()`` for more robustness
     '''
     with closing(io.BytesIO(string)) as bytestream: # Open a "fake file" with the Gzip string pickle in it
         obj = load(bytestream, **kwargs)
@@ -394,7 +394,7 @@ def dumpstr(obj=None, **kwargs):
         obj (any): the object to convert
         kwargs (dict): passed to :func:`save()`
     
-    New in version 2.2.0: uses ``sc.save()`` for more robustness
+    *New in version 2.2.0:* uses ``sc.save()`` for more robustness
     '''
     bytesobj = save(filename=None, obj=obj, **kwargs)
     result = bytesobj.read() # Read all of the content into result
@@ -472,8 +472,8 @@ def loadzip(filename=None, folder=None, **kwargs):
 
         data = sc.loadzip('my-files.zip')
 
-    | New in version 2.0.0.
-    | New in version 2.2.0: load into memory instead of extracting to disk; see :func:`unzip` for extracting
+    | *New in version 2.0.0.*
[…]
