+++ sciris/sc_datetime.py
@@ -34,7 +34,7 @@ def time():
     See also :func:`now()` to return a datetime object, and :func:`getdate()` to 
     return a string.
     
-    New in version 2.2.0.
+    *New in version 2.2.0.*
     '''
     return pytime.time()
     
@@ -61,7 +61,7 @@ def now(astype='dateobj', timezone=None, utc=False, tostring=False, dateformat=N
         sc.now(tostring=True) # Backwards-compatible alias for astype='str'
         sc.now(dateformat='%Y-%b-%d') # Return a different date format
 
-    New in version 1.3.0: made "astype" the first argument; removed "tostring" argument
+    *New in version 1.3.0:* made "astype" the first argument; removed "tostring" argument
     '''
     if isinstance(utc, str): timezone = utc # Assume it's a timezone
     if timezone is not None: tzinfo = du.tz.gettz(timezone) # Timezone is a string
@@ -278,10 +278,10 @@ def date(obj=None, *args, start_date=None, readformat=None, to='date', as_date=N
         sc.date([35,36,37], start_date='2020-01-01', to='str') # Returns ['2020-02-05', '2020-02-06', '2020-02-07']
         sc.date(1923288822, readformat='posix') # Interpret as a POSIX timestamp
 
-    | New in version 1.0.0.
-    | New in version 1.2.2: "readformat" argument; renamed "dateformat" to "outformat"
-    | New in version 2.0.0: support for ``np.datetime64`` objects
-    | New in version 2.2.0: added "to" argument, and support for ``pd.Timestamp`` and ``np.datetime64`` output; allow None
+    | *New in version 1.0.0.*
+    | *New in version 1.2.2:* "readformat" argument; renamed "dateformat" to "outformat"
+    | *New in version 2.0.0:* support for ``np.datetime64`` objects
+    | *New in version 2.2.0:* added "to" argument, and support for ``pd.Timestamp`` and ``np.datetime64`` output; allow None
     '''
 
     # Handle deprecation
@@ -369,8 +369,8 @@ def day(obj, *args, start_date=None, **kwargs):
         sc.day(sc.now()) # Returns how many days into the year we are
         sc.day(['2021-01-21', '2024-04-04'], start_date='2022-02-22') # Days can be positive or negative
 
-    | New in version 1.0.0.
-    | New in version 1.2.2: renamed "start_day" to "start_date"
+    | *New in version 1.0.0.*
+    | *New in version 1.2.2:* renamed "start_day" to "start_date"
     '''
 
     # Handle deprecation
@@ -425,8 +425,8 @@ def daydiff(*args):
         
         doy = sc.daydiff('2022-03-20') # Returns 79, the number of days since 2022-01-01
 
-    | New in version 1.0.0.
-    | New in version 2.2.0: Calculated relative days with one argument
+    | *New in version 1.0.0.*
+    | *New in version 2.2.0:* Calculated relative days with one argument
     '''
     days = [date(day) for day in args]
     if len(days) == 1:
@@ -468,10 +468,10 @@ def daterange(start_date=None, end_date=None, interval=None, inclusive=True, as_
         dates2 = sc.daterange('2020-03-01', '2022-05-01', interval=dict(months=2), asdate=True)
         dates3 = sc.daterange('2020-03-01', weeks=5)
 
-    | New in version 1.0.0.
-    | New in version 1.3.0: "interval" argument
-    | New in version 2.0.0: ``sc.datedelta()`` arguments
-    | New in version 2.2.0: preserve input type
+    | *New in version 1.0.0.*
+    | *New in version 1.3.0:* "interval" argument
+    | *New in version 2.0.0:* ``sc.datedelta()`` arguments
[…]
