+++ sciris/sc_dataframe.py
@@ -27,7 +27,7 @@ class dataframe(pd.DataFrame):
         columns (list): column labels (if a dict is supplied, the value sets the dtype)
         nrows (int): the number of arrows to preallocate (default 0)
         dtypes (list): optional list of data types to set each column to
-        kwargs (dict): passed to ``pd.DataFrame()``
+        kwargs (dict): passed to :class:`pd.DataFrame() <pandas.DataFrame>`
 
     **Examples**::
 
@@ -98,7 +98,7 @@ class dataframe(pd.DataFrame):
 
     def set_dtypes(self, dtypes):
         '''
-        Set dtypes in-place (see pandas ``df.astype()`` for the user-facing version)
+        Set dtypes in-place (see :meth:`df.astype() <pandas.DataFrame.astype>` for the user-facing version)
         
         *New in version 2.2.0.*
         '''
@@ -115,7 +115,7 @@ class dataframe(pd.DataFrame):
         
         Args:
             col (str): the column to get the index of (return 0 if None)
-            die (bool): whether to raise an exception if the column could not be found (else, return NOne)
+            die (bool): whether to raise an exception if the column could not be found (else, return None)
         
         *New in version 2.2.0:* renamed from "_sanitizecols"
         '''
@@ -264,7 +264,7 @@ class dataframe(pd.DataFrame):
             ncols (int): maximum number of columns to show (default: all)
             width (int): maximum screen width (default: 999)
             precision (int): number of decimal places to show (default: 4)
-            kwargs (dict): passed to ``pd.option_context()``
+            kwargs (dict): passed to :class:`pd.option_context() <pandas.option_context>`
         
         **Examples**::
             
@@ -291,7 +291,7 @@ class dataframe(pd.DataFrame):
         '''
         Remove a row from the data frame.
         
-        To pop a column, see ``df.pop()``.        
+        To pop a column, see :meth:`df.pop() <pandas.DataFrame.pop>`
         '''
         rowindex = int(key)
         thisrow = self.iloc[rowindex,:]
@@ -303,7 +303,7 @@ class dataframe(pd.DataFrame):
     def replacedata(self, newdata=None, newdf=None, reset_index=True, inplace=True):
         '''
         Replace data in the dataframe with other data; usually not used directly
-        by the user, but used as part of e.g. ``df.concat()``.
+        by the user, but used as part of e.g. :meth:`df.concat() <dataframe.concat>`.
 
         Args:
             newdata (array): replace the dataframe's data with these data
@@ -328,7 +328,7 @@ class dataframe(pd.DataFrame):
         '''
         Add row(s) to the end of the dataframe. 
         
-        See also ``concat()`` and ``insertrow()``. Similar to the pandas operation
+        See also :meth:`df.concat() <dataframe.concat>` and :meth:`df.insertrow() <dataframe.insertrow>`. Similar to the pandas operation
         ``df.iloc[-1] = ...``, but faster and provides additional type checking.
 
         Args:
@@ -363,14 +363,15 @@ class dataframe(pd.DataFrame):
 
     def insertrow(self, index=0, value=None, reset_index=True, inplace=True, **kwargs):
         '''
-        Insert row(s) at the specified location. See also ``concat()`` and ``appendrow()``.
+        Insert row(s) at the specified location. See also :meth:`df.concat() <dataframe.concat>`
+        and :meth:`df.appendrow() <dataframe.appendrow>`.
 
         Args:
             index (int): index at which to insert new row(s)
             value (array): the row(s) to insert
             reset_index (bool): update the index
             inplace (bool): whether to modify in-place
-            kwargs (dict): passed to ``df.concat()``
+            kwargs (dict): passed to `:meth:`df.concat() <dataframe.concat>`
         
[…]
