+++ sciris/sc_datetime.py
@@ -2,10 +2,10 @@
 Time/date utilities.
 
 Highlights:
-    - :func:`tic` / :func:`toc` / :func:`timer`: simple methods for timing durations
-    - :func:`readdate`: convert strings to dates using common formats
-    - :func:`daterange`: create a list of dates
-    - :func:`datedelta`: perform calculations on date strings
+    - :func:`:func:`sc.tic() <tic>`` / :func:`sc.toc() <toc>` / :class:`sc.timer() <timer>`: simple methods for timing durations
+    - :func:`sc.readdate() <readdate>`: convert strings to dates using common formats
+    - :func:`sc.daterange() <daterange>`: create a list of dates
+    - :func:`sc.datedelta() <datedelta>`: perform calculations on date strings
 '''
 
 import time as pytime
@@ -31,7 +31,7 @@ def time():
     '''
     Get current time in seconds -- alias to time.time()
     
-    See also :func:`now()` to return a datetime object, and :func:`getdate()` to 
+    See also :func:`sc.now() <now>` to return a datetime object, and :func:`sc.getdate() <getdate>` to
     return a string.
     
     *New in version 2.2.0.*
@@ -43,11 +43,11 @@ def now(astype='dateobj', timezone=None, utc=False, tostring=False, dateformat=N
     '''
     Get the current time as a datetime object, optionally in UTC time.
 
-    ``sc.now()`` is similar to ``sc.getdate()``, but ``sc.now()`` returns a datetime
-    object by default, while ``sc.getdate()`` returns a string by default.
+    :func:`sc.now() <now>` is similar to :func:`sc.getdate() <getdate>`, but :func:`sc.now() <now>` returns a datetime
+    object by default, while :func:`sc.getdate() <getdate>` returns a string by default.
 
     Args:
-        astype     (str)  : what to return; choices are "dateobj", "str", "float"; see ``sc.getdate()`` for more
+        astype     (str)  : what to return; choices are "dateobj", "str", "float"; see :func:`sc.getdate() <getdate>` for more
         timezone   (str)  : the timezone to set the itme to
         utc        (bool) : whether the time is specified in UTC time
         dateformat (str)  : if ``astype`` is ``'str'``, use this output format
@@ -80,7 +80,7 @@ def getdate(obj=None, astype='str', dateformat=None):
         '''
         Alias for converting a date object to a formatted string.
 
-        See also ``sc.now()``.
+        See also :func:`sc.now() <now>`.
 
         Args:
             obj (datetime): the datetime object to convert
@@ -251,9 +251,9 @@ def date(obj=None, *args, start_date=None, readformat=None, to='date', as_date=N
     list/array of any of those -- to a date object (or string, pandas, or numpy
     date). To convert an integer to a date, you must supply a start date.
 
-    Caution: while this function and ``sc.readdate()`` are similar, and indeed this function
-    calls ``sc.readdate()`` if the input is a string, in this function an integer is treated
-    as a number of days from start_date, while for ``sc.readdate()`` it is treated as a
+    Caution: while this function and :func:`sc.readdate() <readdate>` are similar, and indeed this function
+    calls :func:`sc.readdate() <readdate>` if the input is a string, in this function an integer is treated
+    as a number of days from start_date, while for :func:`sc.readdate() <readdate>` it is treated as a
     timestamp in seconds.
 
     Note: in this and other date functions, arguments work either with or without
@@ -263,7 +263,7 @@ def date(obj=None, *args, start_date=None, readformat=None, to='date', as_date=N
[…]
