+++ docs/_templates/custom-module-template.rst
@@ -14,20 +14,6 @@
    {% endif %}
    {% endblock %}
 
-   {% block functions %}
-   {% if functions %}
-   .. rubric:: {{ _('Functions') }}
-
-   .. autosummary::
-      :toctree:
-      :template: custom-function-template.rst
-      :nosignatures:
-   {% for item in functions %}
-      {{ item }}
-   {%- endfor %}
-   {% endif %}
-   {% endblock %}
-
    {% block classes %}
    {% if classes %}
    .. rubric:: {{ _('Classes') }}
@@ -37,6 +23,20 @@
       :template: custom-class-template.rst
       :nosignatures:
    {% for item in classes %}
+      {{ item }}
+   {%- endfor %}
+   {% endif %}
+   {% endblock %}
+
+   {% block functions %}
+   {% if functions %}
+   .. rubric:: {{ _('Functions') }}
+
+   .. autosummary::
+      :toctree:
+      :template: custom-function-template.rst
+      :nosignatures:
+   {% for item in functions %}
       {{ item }}
    {%- endfor %}
    {% endif %}
