+++ .github/workflows/tests.yaml
@@ -4,10 +4,10 @@ on: [pull_request]
 jobs:
   install_and_test:
     runs-on: ubuntu-latest
-    timeout-minutes: 5
+    timeout-minutes: 8
     strategy:
       fail-fast: false
-      max-parallel: 8
+      max-parallel: 1
       matrix:
         python-version: [ '3.9' ]
     name: Install and test
