+++ sciris/sc_fileio.py
@@ -2,13 +2,13 @@
 Functions for reading/writing to files, including pickles, JSONs, and Excel.
 
 Highlights:
-    - :func:`save` / :func:`load`: efficiently save/load any Python object (via pickling)
-    - :func:`savetext` / :func:`loadtext`: likewise, for text
-    - :func:`savejson` / :func:`loadjson`: likewise, for JSONs
-    - :func:`saveyaml` / :func:`saveyaml`: likewise, for YAML
-    - :func:`thisdir`: get current folder
-    - :func:`getfilelist`: easy way to access glob
-    - :func:`rmpath`: remove files and folders
+    - :func:`sc.save() <save>` / :func:`sc.load() <load>`: efficiently save/load any Python object (via pickling)
+    - :func:`sc.savetext() <savetext>` / :func:`sc.loadtext() <loadtext>`: likewise, for text
+    - :func:`sc.savejson() <savejson>` / :func:`sc.loadjson() <loadjson>`: likewise, for JSONs
+    - :func:`sc.saveyaml() <saveyaml>` / :func:`sc.saveyaml() <saveyaml>`: likewise, for YAML
+    - :func:`sc.thisdir() <thisdir>`: get current folder
+    - :func:`sc.getfilelist() <getfilelist>`: easy way to access glob
+    - :func:`sc.rmpath() <rmpath>`: remove files and folders
 """
 
 ##############################################################################
@@ -128,9 +128,9 @@ def _load_filestr(filename, folder):
 def load(filename=None, folder=None, verbose=False, die=None, remapping=None, 
          method='pickle', auto_remap=True, **kwargs):
     '''
-    Load a file that has been saved as a gzipped pickle file, e.g. by :func:`save`.
+    Load a file that has been saved as a gzipped pickle file, e.g. by :func:`sc.save() <save>`.
     Accepts either a filename (standard usage) or a file object as the first argument.
-    Note that `:func:`load`/:func:`loadobj` are aliases of each other.
+    Note that :func:`sc.load() <load>`/:func:`sc.loadobj() <loadobj>` are aliases of each other.
 
     **Note 1**: Since this function relies on pickle, it can potentially execute arbitrary
     code, so you should only use it with sources you trust. For more information, see:
@@ -139,8 +139,9 @@ def load(filename=None, folder=None, verbose=False, die=None, remapping=None,
     **Note 2**: When a pickle file is loaded, Python imports any modules that are referenced
     in it. This is a problem if module has been renamed. In this case, you can
     use the ``remapping`` argument to point to the new modules or classes. For
-    more robustness, use the :func:`savewithmetadata`/:func:`loadwithmetadata`
+    more robustness, use the :func:`sc.savewithmetadata() <savewithmetadata>`/:func:`sc.loadwithmetadata() <loadwithmetadata>`
     functions.
+
 
     Args:
         filename   (str/Path): the filename (or full path) to load
@@ -188,15 +189,16 @@ def save(filename='default.obj', obj=None, folder=None, method='pickle', compres
     '''
     Save any object to disk
     
-    This function is similar to ``pickle.dump()`` in that it serializes the object
+    This function is similar to :func:`pickle.dump()` in that it serializes the object
     to a file. Key differences include:
         
         - It takes care of opening/closing the file for writing
         - It compresses the data by default
         - It supports different serialization methods (e.g. pickle or dill)
     
-    Once an object is saved, it can be loaded with :func:`load`. Note that 
-    :func:`save()`/:func:`saveobj()` are identical.
+    Once an object is saved, it can be loaded with :func:`sc.load() <load>`. Note that 
+    :func:`sc.save() <save>`/:func:`sc.saveobj() <saveobj>` are identical.
+    
     
[…]
