+++ sciris/sc_printing.py
@@ -5,7 +5,7 @@ Highlights:
     - :func:`heading`: print text as a 'large' heading
     - :func:`colorize`: print text in a certain color
     - :func:`pr`: print full representation of an object, including methods and each attribute
-    - :func:`sigfigs`: truncate a number to a certain number of significant figures
+    - :func:`sigfig`: truncate a number to a certain number of significant figures
     - :func:`progressbar`: show a (text-based) progress bar
     - :func:`capture`: capture text output (e.g., stdout) as a variable
 '''
