+++ sciris/sc_math.py
@@ -2,10 +2,10 @@
 Extensions to Numpy, including finding array elements and smoothing data.
 
 Highlights:
-    - :func:`findinds`: find indices of an array matching a condition
-    - :func:`findnearest`: find nearest matching value
-    - :func:`rolling`: calculate rolling average
-    - :func:`smooth`: simple smoothing of 1D or 2D arrays
+    - :func:`sc.findinds() <findinds>`: find indices of an array matching a condition
+    - :func:`sc.findnearest() <findnearest>`: find nearest matching value
+    - :func:`sc.rolling() <rolling>`: calculate rolling average
+    - :func:`sc.smooth() <smooth>`: simple smoothing of 1D or 2D arrays
 '''
 
 import numpy as np
@@ -90,18 +90,18 @@ def findinds(arr=None, val=None, *args, eps=1e-6, first=False, last=False, ind=N
     If one argument, find nonzero values. With two arguments, check for equality
     using eps (by default 1e-6, to handle single-precision floating point). Returns 
     a tuple of arrays if val1 is multidimensional, else returns an array. Similar 
-    to calling ``np.nonzero(np.isclose(arr, val))[0]``.
+    to calling :func:`np.nonzero(np.isclose(arr, val))[0] <numpy.nonzero>`.
 
     Args:
         arr    (array): the array to find values in
         val    (float): if provided, the value to match
         args   (list):  if provided, additional boolean arrays
-        eps    (float): the precision for matching (default 1e-6, equivalent to ``np.isclose()``'s atol)
+        eps    (float): the precision for matching (default 1e-6, equivalent to :func:`numpy.isclose`'s atol)
         first  (bool):  whether to return the first matching value (equivalent to ind=0)
         last   (bool):  whether to return the last matching value (equivalent to ind=-1)
         ind    (int):   index of match to retrieve
         die    (bool):  whether to raise an exception if first or last is true and no matches were found
-        kwargs (dict):  passed to ``np.isclose()``
+        kwargs (dict):  passed to :func:`numpy.isclose()`
 
     **Examples**::
 
@@ -218,19 +218,19 @@ def count(arr=None, val=None, eps=1e-6, **kwargs):
     '''
     Count the number of matching elements.
 
-    Similar to ``np.count_nonzero()``, but allows for slight mismatches (e.g.,
+    Similar to :func:`numpy.count_nonzero()`, but allows for slight mismatches (e.g.,
     floats vs. ints). Equivalent to ``len(sc.findinds())``.
 
     Args:
         arr (array): the array to find values in
         val (float): if provided, the value to match
-        eps (float): the precision for matching (default 1e-6, equivalent to np.isclose's atol)
-        kwargs (dict): passed to ``np.isclose()``
+        eps (float): the precision for matching (default 1e-6, equivalent to :func:`numpy.isclose`'s atol)
+        kwargs (dict): passed to :func:`numpy.isclose()` 
 
     **Examples**::
 
         sc.count(rand(10)<0.5) # returns e.g. 4
-        sc.count([2,3,6,3], 3) # returs 2
+        sc.count([2,3,6,3], 3) # returns 2
 
     *New in version 2.0.0.*
     '''
@@ -305,7 +305,7 @@ def getvaliddata(data=None, filterdata=None, defaultind=0): # pragma: no cover
 
 def sanitize(data=None, returninds=False, replacenans=None, defaultval=None, die=True, verbose=False, label=None):
         '''
-        Sanitize input to remove NaNs. (NB: ``sc.sanitize()`` and ``sc.rmnans()`` are aliases.)
[…]
