+++ sciris/sc_plotting.py
@@ -3,13 +3,13 @@
 Extensions to Matplotlib, including 3D plotting and plot customization.
 
 Highlights:
-    - :func:`plot3d`: easy way to render 3D plots
-    - :func:`boxoff`: turn off top and right parts of the axes box
-    - :func:`commaticks`: convert labels from "10000" and "1e6" to "10,000" and "1,000,0000"
-    - :func:`SIticks`: convert labels from "10000" and "1e6" to "10k" and "1m"
-    - :func:`maximize`: make the figure fill the whole screen
-    - :func:`savemovie`: save a sequence of figures as an MP4 or other movie
-    - :func:`fonts`: list available fonts or add new ones
+    - :func:`sc.plot3d() <plot3d>`: easy way to render 3D plots
+    - :func:`sc.boxoff() <boxoff>`: turn off top and right parts of the axes box
+    - :func:`sc.commaticks() <commaticks>`: convert labels from "10000" and "1e6" to "10,000" and "1,000,0000"
+    - :func:`sc.SIticks() <SIticks>`: convert labels from "10000" and "1e6" to "10k" and "1m"
+    - :func:`sc.maximize() <maximize>`: make the figure fill the whole screen
+    - :func:`sc.savemovie() <savemovie>`: save a sequence of figures as an MP4 or other movie
+    - :func:`sc.fonts() <fonts>`: list available fonts or add new ones
 '''
 
 ##############################################################################
@@ -42,7 +42,7 @@ def fig3d(num=None, returnax=False, figkwargs=None, axkwargs=None, **kwargs):
     '''
     Shortcut for creating a figure with 3D axes.
 
-    Usually not invoked directly; kwargs are passed to ``pl.figure()``
+    Usually not invoked directly; kwargs are passed to ``pl.figure()`` # auto_docfix
     '''
     figkwargs = scu.mergedicts(figkwargs, kwargs, num=num)
     axkwargs = scu.mergedicts(axkwargs)
@@ -291,7 +291,7 @@ def stackedbar(x=None, values=None, colors=None, labels=None, transpose=False,
         flipud    (bool)     : whether to flip the array upside down prior to plotting
         cum       (bool)     : whether the array is already a cumulative sum
         barh      (bool)     : whether to plot as a horizontal instead of vertical bar
-        kwargs    (dict)     : passed to ``pl.bar()``
+        kwargs    (dict)     : passed to ``pl.bar()`` # auto_docfix
     
     **Example**::
         
@@ -475,13 +475,13 @@ def setaxislim(which=None, ax=None, data=None):
 
 
 def setxlim(data=None, ax=None):
-    ''' Alias for ``sc.setaxislim(which='x')`` '''
+    ''' Alias for :func:`sc.setaxislim(which='x') <setaxislim>` '''
     return setaxislim(data=data, ax=ax, which='x')
 
 
 def setylim(data=None, ax=None):
     '''
-    Alias for ``sc.setaxislim(which='y')``.
+    Alias for :func:`sc.setaxislim(which='y') <setaxislim>`.
 
     **Example**::
 
@@ -513,7 +513,7 @@ def commaticks(ax=None, axis='y', precision=2, cursor_precision=0):
     '''
     Use commas in formatting the y axis of a figure (e.g., 34,000 instead of 34000).
 
-    To use something other than a comma, set the default separator via e.g. ``sc.options(sep='.')``.
+    To use something other than a comma, set the default separator via e.g. :class:`sc.options(sep='.') <ScirisOptions>`.
 
     Args:
         ax (any): axes to modify; if None, use current; else can be a single axes object, a figure, or a list of axes
[…]
