+++ sciris/sc_printing.py
@@ -2,12 +2,12 @@
 Printing/notification functions.
 
 Highlights:
-    - :func:`heading`: print text as a 'large' heading
-    - :func:`colorize`: print text in a certain color
-    - :func:`pr`: print full representation of an object, including methods and each attribute
-    - :func:`sigfig`: truncate a number to a certain number of significant figures
-    - :func:`progressbar`: show a (text-based) progress bar
-    - :func:`capture`: capture text output (e.g., stdout) as a variable
+    - :func:`sc.heading() <heading>`: print text as a 'large' heading
+    - :func:`sc.colorize() <colorize>`: print text in a certain color
+    - :func:`sc.pr() <pr>`: print full representation of an object, including methods and each attribute
+    - :func:`sc.sigfig() <sigfig>`: truncate a number to a certain number of significant figures
+    - :func:`sc.progressbar() <progressbar>`: show a (text-based) progress bar
+    - :func:`sc.capture() <capture>`: capture text output (e.g., stdout) as a variable
 '''
 
 import io
@@ -278,7 +278,7 @@ def indent(prefix=None, text=None, suffix='\n', n=0, pretty=False, width=70, **k
         n (int): if prefix is not specified, the size of the indent
         pretty (bool): whether to use pprint to format the text
         width (int): maximum width before wrapping (if None, don't wrap)
-        kwargs (dict): passed to ``textwrap.fill()``
+        kwargs (dict): passed to :func:`textwrap.fill()`
 
     **Examples**::
 
@@ -410,7 +410,7 @@ def printmean(data, stds=2, mean_sf=None, err_sf=None, doprint=True, **kwargs):
         mean_sf (int): if provided, use this number of significant figures for the mean rather than the auto-calculated
         err_sf (int): ditto, but for the error (standard deviation)
         doprint (bool): whether to print (else, return the string)
-        kwargs (dict): passed to ``sc.sigfig()``
+        kwargs (dict): passed to :func:`sc.sigfig() <sigfig>`
     
     **Example**::
         
@@ -457,7 +457,7 @@ def printmedian(data, ci=95, sf=3, doprint=True, **kwargs):
         ci (int/float/list/str): the confidence interval to use to use (see above for details)
         sf (int): number of significant figures to use
         doprint (bool): whether to print (else, return the string)
-        kwargs (dict): passed to ``sc.sigfig()``
+        kwargs (dict): passed to :func:`sc.sigfig() <sigfig>`
     
     **Examples**::
         
@@ -726,7 +726,7 @@ def colorize(color=None, string=None, doprint=None, output=False, enable=True, s
         string (str): the text to be colored
         doprint (bool): whether to print the string (default true unless output)
         output (bool): whether to return the modified version of the string (default false)
-        enable (bool): switch to allow ``sc.colorize()`` to be easily turned off without converting to a ``print()`` statement
+        enable (bool): switch to allow :func:`sc.colorize() <colorize>` to be easily turned off without converting to a :func:`print()` statement
         showhelp (bool): show help rather than changing colors
         fg (str): foreground colour
         bg (str): background colour
@@ -741,7 +741,7 @@ def colorize(color=None, string=None, doprint=None, output=False, enable=True, s
         sc.colorize() # Stop typing in magenta
         sc.colorize('cat in the hat', fg='#ffa044', bg='blue', style='italic+underline') # Alternate usage example
 
-    To get available colors, type ``sc.colorize(showhelp=True)``.
+    To get available colors, type :func:`sc.colorize(showhelp=True) <colorize>`.
 
     | *New in version 1.3.1:* "doprint" argument; ansicolors shortcut
     '''
[…]
