+++ docs/conf.py
@@ -112,6 +112,7 @@ intersphinx_mapping = {
     "pandas": ("https://pandas.pydata.org/pandas-docs/stable/", None),
     "scipy": ("https://docs.scipy.org/doc/scipy/", None),
     "matplotlib": ("https://matplotlib.org/stable/", None),
+    "psutil": ("https://psutil.readthedocs.io/en/latest/", None),
 }
 
 def setup(app):
