+++ sciris/sc_utils.py
@@ -6,15 +6,15 @@ ignore a function that you don't need than write one from scratch that you
 do need.
 
 Highlights:
-    - :func:`dcp`: shortcut to ``copy.deepcopy()``
-    - :func:`pp`: shortcut to ``pprint.pprint()``
-    - :func:`isnumber`: checks if something is any number type
-    - :func:`tolist`: converts any object to a list, for easy iteration
-    - :func:`toarray`: tries to convert any object to an array, for easy use with numpy
-    - :func:`mergedicts`: merges any set of inputs into a dictionary
-    - :func:`mergelists`: merges any set of inputs into a list
-    - :func:`runcommand`: simple way of executing a shell command
-    - :func:`download`: download multiple URLs in parallel
+    - :func:`sc.dcp() <dcp>`: shortcut to :func:`copy.deepcopy()`
+    - :func:`sc.pp() <pp>`: shortcut to :func:`pprint.pprint()`
+    - :func:`sc.isnumber() <isnumber>`: checks if something is any number type
+    - :func:`sc.tolist() <tolist>`: converts any object to a list, for easy iteration
+    - :func:`sc.toarray() <toarray>`: tries to convert any object to an array, for easy use with numpy
+    - :func:`sc.mergedicts() <mergedicts>`: merges any set of inputs into a dictionary
+    - :func:`sc.mergelists() <mergelists>`: merges any set of inputs into a list
+    - :func:`sc.runcommand() <runcommand>`: simple way of executing a shell command
+    - :func:`sc.download() <download>`: download multiple URLs in parallel
 '''
 
 ##############################################################################
@@ -59,8 +59,8 @@ __all__ = ['fast_uuid', 'uuid', 'dcp', 'cp', 'pp', 'sha', 'traceback',
 
 def fast_uuid(which=None, length=None, n=1, secure=False, forcelist=False, safety=1000, recursion=0, recursion_limit=10, verbose=True):
     '''
-    Create a fast UID or set of UIDs. Note: for certain applications, :func:`uuid`
-    is faster than :func:`fast_uuid`!
+    Create a fast UID or set of UIDs. Note: for certain applications, :func:`sc.uuid() <uuid>`
+    is faster than :func:`sc.fast_uuid() <fast_uuid>`!
 
     Args:
         which (str): the set of characters to choose from (default ascii)
@@ -236,7 +236,7 @@ def dcp(obj, die=True, verbose=True):
     '''
     Shortcut to perform a deep copy operation
 
-    Almost identical to ``copy.deepcopy()``, but optionally fall back to copy()
+    Almost identical to :func:`copy.deepcopy()`, but optionally fall back to :func:`copy.copy()`
     if deepcopy fails.
 
     Args:
@@ -259,7 +259,7 @@ def cp(obj, die=True, verbose=True):
     '''
     Shortcut to perform a shallow copy operation
 
-    Almost identical to ``copy.copy()``, but optionally allow failures
+    Almost identical to :func:`copy.copy()`, but optionally allow failures
 
     *New in version 2.0.0:* default die=True instead of False
     '''
@@ -294,8 +294,8 @@ def pp(obj, jsonify=True, doprint=None, output=False, verbose=False, **kwargs):
     '''
     Shortcut for pretty-printing the object.
 
-    Almost identical to ``pprint.pprint()``, but can also be used as an alias for
-    ``pprint.pformat()``.
+    Almost identical to :func:`pprint.pprint()`, but can also be used as an alias for
+    :func:`pprint.pformat()`.
 
     Args:
         obj     (any):  object to print
@@ -303,7 +303,7 @@ def pp(obj, jsonify=True, doprint=None, output=False, verbose=False, **kwargs):
         doprint (bool): whether to show output (default true)
[…]
