+++ sciris/sc_versioning.py
@@ -3,12 +3,12 @@ Functions for checking and saving versioning information, such as Python package
 versions, git versions, etc.
 
 Highlights:
-    - :func:`freeze`: programatically store "pip freeze" output
-    - :func:`require`: require a specific version of a package
-    - :func:`gitinfo`: gets the git information (if available) of a given file
-    - :func:`compareversions`: easy way to compare version numbers
-    - :func:`storemetadata`: collects relevant metadata into a dictionary
-    - :func:`savewithmetadata`: saves data as a zip file including versioning metadata
+    - :func:`sc.freeze() <freeze>`: programmatically store "pip freeze" output
+    - :func:`sc.require() <require>`: require a specific version of a package
+    - :func:`sc.gitinfo() <gitinfo>`: gets the git information (if available) of a given file
+    - :func:`sc.compareversions() <compareversions>`: easy way to compare version numbers
+    - :func:`sc.metadata() <metadata>`: collects relevant metadata into a dictionary
+    - :func:`sc.savewithmetadata() <savewithmetadata>`: saves data as a zip file including versioning metadata
 '''
 
 import os
@@ -308,7 +308,7 @@ def compareversions(version1, version2):
 def getcaller(frame=2, tostring=True, includelineno=False, includeline=False, relframe=0, die=False):
     '''
     Try to get information on the calling function, but fail gracefully. See also
-    :func:`thisfile`.
+    :func:`sc.thisfile() <thisfile>`.
 
     Frame 1 is the file calling this function, so not very useful. Frame 2 is
     the default assuming it is being called directly. Frame 3 is used if
@@ -463,10 +463,10 @@ def loadmetadata(filename, load_all=False, die=True):
     '''
     Read metadata from a saved image; currently only PNG and SVG are supported.
 
-    Only for use with images saved with ``sc.savefig()``. Metadata retrieval for PDF
-    is not currently supported. To load metadata saved with ``sc.metadata()``, 
-    you can also use ``sc.loadjson()`` instead. To load metadata saved with ``sc.savewithmetadata()``,
-    use ``sc.loadwithmetadata()`` instead.
+    Only for use with images saved with :func:`sc.savefig() <savefig>`. Metadata retrieval for PDF
+    is not currently supported. To load metadata saved with :func:`sc.metadata() <metadata>`, 
+    you can also use :func:`sc.loadjson() <loadjson>` instead. To load metadata saved with :func:`sc.savewithmetadata() <savewithmetadata>`,
+    use :func:`sc.loadwithmetadata() <loadwithmetadata>` instead.
 
     Args:
         filename (str): the name of the file to load the data from
@@ -590,15 +590,15 @@ def savewithmetadata(filename, obj, folder=None, user=True, caller=True, git=Tru
     Args:
         filename (str/path): the file to save to (must end in .zip)
         obj (any): the object to save
-        caller (bool): store information on the current user in the metadata (see ``sc.metadata()``)
-        caller (bool): store information on the calling file in the metadata (see ``sc.metadata()``)
-        git (bool): store the git version in the metadata (see ``sc.metadata()``)
-        pipfreeze (bool): store the output of "pip freeze" in the metadata (see ``sc.metadata()``)
+        caller (bool): store information on the current user in the metadata (see :func:`sc.metadata() <metadata>`)
+        caller (bool): store information on the calling file in the metadata (see :func:`sc.metadata() <metadata>`)
+        git (bool): store the git version in the metadata (see :func:`sc.metadata() <metadata>`)
+        pipfreeze (bool): store the output of "pip freeze" in the metadata (see :func:`sc.metadata() <metadata>`)
         comments (str/dict): other comments/information to store in the metadata (must be JSON-compatible)
[…]
