+++ docs/overview.rst
@@ -1,6 +1,8 @@
 Getting started
 ===============
 
+.. currentmodule:: sciris
+
 .. image:: https://badgen.net/pypi/v/sciris/?color=blue
  :target: https://pypi.com/project/sciris
 
@@ -26,8 +28,8 @@ Highlights
 ~~~~~~~~~~
 Some highlights of Sciris (``import sciris as sc``):
 
-- **Powerful containers** – The ``sc.odict`` class is what ``OrderedDict`` (almost) could have been, allowing reference by position or key, casting to a NumPy array, sorting and enumeration functions, etc.
-- **Array operations** – Want to find the indices of an array that match a certain value or condition? ``sc.findinds()`` will do that. How about just the nearest value, regardless of exact match? ``sc.findnearest()``. What about the last matching value? ``sc.findlast()``. Yes, you could do ``np.nonzero()[0][-1]`` instead, but ``sc.findlast()`` is easier to read, type, and remember, and handles edge cases more elegantly.
+- **Powerful containers** – The :class:`sc.odict <sc_odict.odict>` class is what ``OrderedDict`` (almost) could have been, allowing reference by position or key, casting to a NumPy array, sorting and enumeration functions, etc.
+- **Array operations** – Want to find the indices of an array that match a certain value or condition? :func:`sc.findinds() <sc_math.findinds>` will do that. How about just the nearest value, regardless of exact match? :func:`sc.findnearest() <sc_math.findnearest>`. What about the last matching value? :func:`sc.findlast() <sc_math.findlast>`. Yes, you could do :func:`np.nonzero()[0][-1] <numpy.nonzero>` instead, but :func:`sc.findlast() <sc_math.findlast>` is easier to read, type, and remember, and handles edge cases more elegantly.
 - **File I/O** – One-liner functions for saving and loading text, JSON, spreadsheets, or even arbitrary Python objects.
 - **Plotting recipes** – Simple functions for mapping sequential or qualitative data onto colors, manipulating color data, and updating axis limits and tick labels, plus several new colormaps.
 
@@ -47,7 +49,7 @@ Where did Sciris come from?
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Development of Sciris began in 2014 to support development of the `Optima <http://optimamodel.com>`_ suite of models. We kept encountering the same issues and inconveniences over and over while building scientific webapps, and began collecting the tools we used to overcome these issues into a shared library. This library evolved into Sciris. (Note: while "Sciris" doesn't mean anything, "iris" means "rainbow" in Greek, and the name was loosely inspired by the wide spectrum of scientific computing features included in Sciris.)
 
-To give a based-on-a-true-story example, let's say you have a dictionary of results for multiple runs of your model, called ``results``. The output of each model run is itself a dictionary, with keys such as ``name`` and ``data``. Now let's say you want to access the data from the first model run. Using plain Python dictionaries, this would be ``results[list(results.keys())[0]]['data']``. Using a Sciris ``objdict``, this is ``results[0].data`` – almost 3x shorter.
[…]
