+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 What's new
 ==========
 
+.. currentmodule:: sciris
+
 All major updates to Sciris are documented here.
 
 By import convention, components of the Sciris library are listed beginning with ``sc.``, e.g. ``sc.odict()``.
@@ -11,9 +13,9 @@ Version 2.2.0 (2023-04-14)
 
 This version's major changes include:
 
-#. **New Parallel class**: A new ``sc.Parallel()`` class allows finer-grained managing of parallel processes, including automatic progress bars, better exception handling, and asynchronous running.
-#. **Better versioning**: New functions ``sc.metadata()``, ``sc.savewithmetadata()``, and ``sc.loadwithmetadata()`` make it easier to store and save metadata along with objects.
-#. **Easier imports**: ``sc.importbypath()`` lets you load a module into Python by providing the folder or filename (useful for loading one-off scripts, or two versions of the same library).
+#. **New Parallel class**: A new :class:`sc.Parallel() <sc_parallel.Parallel>` class allows finer-grained managing of parallel processes, including automatic progress bars, better exception handling, and asynchronous running.
+#. **Better versioning**: New functions :func:`sc.metadata() <sc_versioning.metadata>`, :func:`sc.savewithmetadata() <sc_versioning.savewithmetadata>`, and :func:`sc.loadwithmetadata() <sc_versioning.loadwithmetadata>` make it easier to store and save metadata along with objects.
+#. **Easier imports**: :func:`sc.importbypath() <sc_utils.importbypath>` lets you load a module into Python by providing the folder or filename (useful for loading one-off scripts, or two versions of the same library).
 #. **Better documentation**: A comprehensive set of tutorials has been added to the documentation, and the documentation has been rewritten in a new style.
 
 
@@ -22,67 +24,66 @@ Improvements and new features
 
 Parallelization
 ^^^^^^^^^^^^^^^
-#. There is a new ``sc.Parallel`` class, which is used to implement the (more or less unchanged) ``sc.parallelize()`` function.
-#. ``sc.parallelize()`` now has a ``progress`` argument that will show a progress bar; the ``returnpool`` argument has been removed (use ``sc.Parallel`` instead).
-
+#. There is a new :class:`sc.Parallel() <sc_parallel.Parallel>` class, which is used to implement the (more or less unchanged) :func:`sc.parallelize() <sc_parallel.parallelize>` function.
+#. :func:`sc.parallelize() <sc_parallel.parallelize>` now has a ``progress`` argument that will show a progress bar; the ``returnpool`` argument has been removed (use :class:`sc.Parallel() <sc_parallel.Parallel>` instead).
 
 Dataframe
 ^^^^^^^^^
-#. Better handling of (and preservation) of ``dtypes`` for dataframe columns, including a new ``df.set_dtypes()`` method.
+#. Better handling of (and preservation) of ``dtypes`` for dataframe columns, including a new :meth:`df.set_dtypes() <sc_dataframe.dataframe.set_dtypes>` method.
 #. Better implementation of underlying logic, leading to significant performance increases in some cases (e.g., iteratively appending rows).
 
 Time/date
 ^^^^^^^^^
-#. New ``sc.time()`` alias for ``time.time()``.
+#. New :func:`sc.time() <sc_datetime.time>` alias for :func:`time.time()`.
 #. Support for ``pandas`` and ``Numpy`` datetime objects.
-#. New ``sc.timer`` attributes and methods: ``rawtimings``, ``sum()``, ``min()``, ``max()``, ``mean()``, ``std()``.
[…]
