+++ CONTRIBUTING.rst
@@ -4,13 +4,13 @@ Contributing
 
 Welcome! We are thrilled you are interested in contributing to Sciris. (No, seriously.) This document will help you get started.
 
-- The Sciris community follows a `code of conduct`_. By participating in this project, you agree to abide by its terms.
+- We are serious about inclusion and believe the open-source software community still has a long way to go. The Sciris community follows a `code of conduct`_. By participating in this project, you agree to abide by its terms.
 - Take a look at our house `style guide`_. Sciris more or less follows Google's Python style guide, but with some exceptions. 
 - Feel free to `open an issue`_ on more or less anything! This project is small enough that we don't need a formal triage system.
 - Pull requests should be made against ``develop``. In addition to following the `style guide`_, please make sure the tests pass (``run_tests`` in the ``tests`` folder; they also run via GitHub actions).
 
 If you have any other questions, please reach out to us: info@sciris.org. Thank you!
 
-.. _code of conduct: https://github.com/sciris/sciris/blob/develop/CODE_OF_CONDUCT.rst
-.. _style guide: https://github.com/sciris/sciris/blob/develop/STYLE_GUIDE.rst
+.. _code of conduct: https://docs.sciris.org/code_of_conduct.html
+.. _style guide: https://docs.sciris.org/style_guide.html
 .. _open an issue: https://github.com/sciris/sciris/issues/new/choose
