+++ docs/requirements.txt
@@ -6,4 +6,5 @@ pypandoc
 sphinx>=1.8
 sphinx-autodoc-typehints
 pydata-sphinx-theme
+sphinx-design
 recommonmark
