+++ docs/index.rst
@@ -1,19 +1,18 @@
 .. title:: Sciris
 
+Welcome to Sciris
+=================
+
 .. image:: sciris-logo-grey.png
     :scale: 20%
     :align: center
 
-`Sciris <http://sciris.org>`__ is library of functions for scientific computing that provide additional flexibility and ease-of-use on top of Numpy, Scipy, and Matplotlib.
-
-ScirisWeb, built on Sciris, is a toolbox that makes building webapps easy -- think of it as `Shiny <https://shiny.rstudio.com/>`__ for Python.
-
-========
-Contents
-========
+.. include:: ../README.rst
+   :start-line: 1
 
 .. toctree::
    :maxdepth: 1
+   :hidden:
 
    overview
    whatsnew
