+++ README.rst
@@ -17,7 +17,7 @@ Welcome to Sciris
 What is Sciris?
 ---------------
 
-Sciris is a library of tools that can help make writing scientific Python code easier and more pleasant. Built on top of `NumPy <https://numpy.org/>`__ and `Matplotlib <https://matplotlib.org/>`__, Sciris provides functions covering a wide range of common math, file I/O, and plotting operations. This means you can get more done with less code, and spend less time looking things up on StackOverflow. It was originally written to help epidemiologists and neuroscientists focus on doing science, rather than on writing coding, but Sciris is applicable across scientific domains.
+Sciris is a library of tools that can help make writing scientific Python code easier and more pleasant. Built on top of `NumPy <https://numpy.org/>`_ and `Matplotlib <https://matplotlib.org/>`_, Sciris provides functions covering a wide range of common math, file I/O, and plotting operations. This means you can get more done with less code, and spend less time looking things up on StackOverflow. It was originally written to help epidemiologists and neuroscientists focus on doing science, rather than on writing coding, but Sciris is applicable across scientific domains.
 
 For more information, see the full `documentation <http://docs.sciris.org/overview.html>`_, the `paper <http://paper.sciris.org>`_, or `GitHub <https://github.com/sciris/sciris>`_.
 
@@ -34,7 +34,7 @@ If you have questions, feature suggestions, or would like some help getting star
 
 
 Installation
-------------
+~~~~~~~~~~~~
 
 Using pip: ``pip install sciris``
 
@@ -42,7 +42,7 @@ Using pip: ``pip install sciris``
 
 
 Tests
------
+~~~~~
 
 Sciris comes with an automated test suite covering all functions. You almost certainly don't need to run these, but if you want to, go to the ``tests`` folder and run ``pytest``. See the readme in that folder for more information.
 
