+++ docs/api/index.rst
@@ -2,6 +2,8 @@
 API reference
 =============
 
+There's a lot here! Below is absolutely everything that's included in Sciris. You can use the links on the left for navigation.
+
 .. toctree::
    :maxdepth: 3
 
