+++ docs/tutorials/index.rst
@@ -1,0 +1,9 @@
+=========
+Tutorials
+=========
+
+.. toctree::
+   :maxdepth: 3
+
+   tut_intro
+
