### 2023-04-16 1999be6
reimplement dictobj
