+++ sciris/sc_dataframe.py
@@ -28,32 +28,34 @@ class dataframe(pd.DataFrame):
         nrows (int): the number of arrows to preallocate (default 0)
         dtypes (list): optional list of data types to set each column to
         kwargs (dict): passed to :class:`pd.DataFrame() <pandas.DataFrame>`
+    
+    *Hint*: Run the example below line by line to get a sense of how the dataframe
+    changes.
 
-    **Examples**::
+    **Example**::
 
-        a = sc.dataframe(cols=['x','y'], data=[[1238,2],[384,5],[666,7]]) # Create data frame
-        a['x'] # Print out a column
-        a[0] # Print out a row
-        a['x',0] # Print out an element
-        a[0] = [123,6]; print(a) # Set values for a whole row
-        a['y'] = [8,5,0]; print(a) # Set values for a whole column
-        a['z'] = [14,14,14]; print(a) # Add new column
-        a.addcol('z', [14,14,14]); print(a) # Alternate way to add new column
-        a.rmcol('z'); print(a) # Remove a column
-        a.pop(1); print(a) # Remove a row
-        a.append([555,2,14]); print(a) # Append a new row
-        a.insert(1,[555,2,14]); print(a) # Insert a new row
-        a.sort(); print(a) # Sort by the first column
-        a.sort('y'); print(a) # Sort by the second column
-        a.addrow([555,2,14]); print(a) # Replace the previous row and sort
-        a.getrow(1) # Return the row starting with value '1'
-        a.rmrow(); print(a) # Remove last row
-        a.rmrow(1238); print(a) # Remove the row starting with element '3'
+        df = sc.dataframe(cols=['x','y'], data=[[1238,2],[384,5],[666,7]]) # Create data frame
+        df['x'] # Print out a column
+        df[0] # Print out a row
+        df['x',0] # Print out an element
+        df[0,:] = [123,6]; print(df) # Set values for a whole row
+        df['y'] = [8,5,0]; print(df) # Set values for a whole column
+        df['z'] = [14,14,14]; print(df) # Add new column
+        df.rmcol('z'); print(df) # Remove a column
+        df.addcol('z', [14,14,14]); print(df) # Alternate way to add new column
+        df.poprow(1); print(df) # Remove a row
+        df.append([555,2,14]); print(df) # Append a new row
+        df.insertrow(1,[556,2,14]); print(df) # Insert a new row
+        df.sort(); print(df) # Sort by the first column
+        df.sort('y'); print(df) # Sort by the second column
+        df.findrow(123) # Return the row starting with value 123
+        df.rmrow(); print(df) # Remove last row
+        df.rmrow(555); print(df) # Remove the row starting with element '555'
 
     The dataframe can be used for both numeric and non-numeric data.
 
     | *New in version 2.0.0:* subclass pandas DataFrame
-    | *New in version 2.2.0:* "dtypes" argument
+    | *New in version 2.2.0:* "dtypes" argument; handling of item setting
     '''
 
     def __init__(self, data=None, columns=None, nrows=None, dtypes=None, **kwargs):
@@ -122,16 +124,8 @@ class dataframe(pd.DataFrame):
         cols = self.cols
         if col is None:
             output = 0 # If not supplied, assume first column is intended
-        elif scu.isstring(col):
-            try:
-                output = cols.index(col) # Convert to index
-            except Exception as E: # pragma: no cover
-                errormsg = f'Could not get index of column "{col}"; columns are:\n{scu.newlinejoin(cols)}\n{E}'
-                if die:
-                    raise scu.KeyNotFoundError(errormsg) from E
-                else:
-                    print(errormsg)
-                    output = None
+        elif col in cols:
[…]
