+++ docs/api/index.rst
@@ -13,4 +13,3 @@ There's a lot here! Below is absolutely everything that's included in Sciris. Yo
    printing
    math
    utils
-
