+++ tests/test_dataframe.py
@@ -26,16 +26,16 @@ def test_dataframe():
     a['y'] = [8,5,0]; dfprint('Set values for a whole column', a)
     a['z'] = [14,14,14]; dfprint('Add new column', a)
     a.addcol('m', [14,15,16]); dfprint('Alternate way to add new column', a)
-    a.rmcol('z'); dfprint('Remove a column', a)
-    a.poprow(1); dfprint('Remove a row', a)
+    a.popcols('z'); dfprint('Remove a column', a)
+    a.poprows(1); dfprint('Remove a row', a)
     a.appendrow([555,2,-1]); dfprint('Append a new row', a)
     a.concat([[1,2,3],[4,5,6]], [9,9,9]); dfprint('Concatenate', a)
     a.insertrow(0, [660,3,-2]); dfprint('Insert a new row', a)
     a.sortrows(); dfprint('Sort by the first column', a)
     a.sortrows('y'); dfprint('Sort by the second column', a)
     dfprint('Return the row starting with value "555"', a.findrow(555))
-    a.rmrow(); dfprint('Remove last row', a)
-    a.rmrow(value=666); dfprint('Remove the row starting with element "666"', a)
+    a.poprows(); dfprint('Remove last row', a)
+    a.poprows(value=666); dfprint('Remove the row starting with element "666"', a)
     p = a.to_pandas(); dfprint('Convert to pandas', p)
     b = a.filtercols(['m','x']); dfprint('Filter to columns m and x', b)
     
@@ -85,7 +85,7 @@ def test_methods():
     assert isinstance(out2, float)
 
     subheading('Other')
-    df.rmrows([1,3]); dfprint('Removing rows', df)
+    df.poprows([1,3]); dfprint('Removing rows', df)
     df.replacecol('a', 300, 333); dfprint('Replacing 300→333', df)
     od = df.to_odict(); dfprint('To dict', od)
     df.appendrow(np.random.rand(2,3)); dfprint('Adding more rows', df)
