+++ sciris/sc_dataframe.py
@@ -566,6 +566,8 @@ class dataframe(pd.DataFrame):
         rather than the index of the row. See :meth:`df.col_index() <dataframe.col_index>` for the column
         equivalent.
         
+        
+        
         Args:
             value (any): the value to look for (default: return last row index)
             col (str): the column to look in (default: first)
@@ -597,42 +599,39 @@ class dataframe(pd.DataFrame):
                 else:
                     return
         return index
-
-
-    def rmrow(self, value=None, col=None, returnval=False, die=True):
-        #TODO 
-        # rename to popbyvalue? or delete and replace with findrow(pop=True)?
-        '''
-        Like pop, but removes by matching the value in the given column instead of the index
-        '''
-        index = self.row_index(value=value, col=col, die=die)
-        if index is not None: self.poprow(index, returnval=returnval)
-        return self
-
+    
 
     def _diffinds(self, inds=None):
         ''' For a given set of indices, get the inverse, in set-speak '''
         if inds is None: inds = []
-        ind_set = set(np.array(inds))
+        ind_set = set(np.arange(len(self))[inds])
         all_set = set(np.arange(self.nrows))
         diff_set = np.array(list(all_set - ind_set))
         return diff_set
 
 
-    def poprows(self, inds=None, reset_index=True, inplace=True):
+    def poprows(self, inds=-1, value=None, col=None, reset_index=True, inplace=True, **kwargs):
         '''
-        Remove multiple rows by index
+        Remove multiple rows by index or value
         
         Args:
             inds (list): the rows to remove
+            values (list): alternatively, search for these values to remove; see :meth:`df.findinds <dataframe.findinds>` for details
+            col (str): if removing by value, use this column to find the values
             reset_index (bool): update the index
             inplace (bool): whether to modify in-place
+            kwargs (dict): passed to :meth:`df.findinds <dataframe.findinds>`
         
-        **Example**::
+        **Examples**::
             
             df = sc.dataframe(np.random.rand(10,3))
-            df.rmrows([3,4,5])
+            df.poprows([3,4,5])
+            
+            df = sc.dataframe(dict(x=[0,1,2,3,4], y=[2,3,2,7,8]))
+            df.poprows(value=2, col='y')
         '''
+        if value is not None:
+            inds = self.findinds(value=value, col=col, **kwargs)
         keep_set = self._diffinds(inds)
         keep_data = self.iloc[keep_set,:]
         newdf = self._constructor(data=keep_data, cols=self.cols)
@@ -663,25 +662,26 @@ class dataframe(pd.DataFrame):
         return output
 
 
-    def findrow(self, value=None, col=None, default=None, closest=False, asdict=False, die=False):
+    def findrow(self, value=None, col=None, default=None, closest=False, pop=False, asdict=False, die=False):
         '''
         Return a row by searching for a matching value.
         
         See :meth:`df.row_index() <dataframe.row_index>` for the equivalent to return the index of the row
-        rather than the row itself.
+        rather than the row itself, and :meth:`df.findinds() <dataframe.findinds>`
+        to find multiple row indices.
 
         Args:
             value (any): the value to look for
             col (str): the column to look for this value in
             default (any): the value to return if key is not found (overrides die)
             closest (bool): whether or not to return the closest row (overrides default and die)
-            returnind (bool): return the index of the row rather than the row itself
+            pop (bool): if True, pop the matching row
             asdict (bool): whether to return results as dict rather than list
             die (bool): whether to raise an exception if the value is not found
 
[…]
