+++ sciris/sc_dataframe.py
@@ -111,36 +111,113 @@ class dataframe(pd.DataFrame):
         return
 
 
-    def col_index(self, col, die=True):
+    def col_index(self, col=None, *args, die=True):
         '''
-        Get the index of the specified column.
+        Get the index of the column named ``col``.
+        
+        Similar to ``df.columns.get_loc(col)``, and opposite of :meth:`df.col_name <dataframe.col_name>`.
         
         Args:
-            col (str): the column to get the index of (return 0 if None)
+            col (str/list): the column(s) to get the index of (return 0 if None)
+            args (list): additional column(s) to get the index of
             die (bool): whether to raise an exception if the column could not be found (else, return None)
         
-        *New in version 2.2.0:* renamed from "_sanitizecols"
+        **Examples**::
+            
+            df = sc.dataframe(dict(a=[1,2,3], b=[4,5,6], c=[7,8,9]))
+            df.col_index('b') # Returns 1
+            df.col_index(1) # Returns 1
+            df.col_index('a', 'c') # Returns [0, 2]
+            
+        *New in version 2.2.0:* renamed from "_sanitizecols"; multiple arguments
         '''
+        arglist = scu.mergelists(col, list(args))
+        outputlist = []
         cols = self.cols
-        if col is None:
-            output = 0 # If not supplied, assume first column is intended
-        elif col in cols:
-            output = cols.index(col) # Convert to index
-        elif scu.isnumber(col):
-            try:
-                cols[col]
-            except Exception as E: # pragma: no cover
-                errormsg = f'Column "{col}" is not a valid index'
-                raise IndexError(errormsg) from E
-            output = col
-        else: # pragma: no cover
-            errormsg = f'Unrecognized column/column type "{col}" {type(col)}'
-            if die:
-                raise TypeError(errormsg)
-            else:
-                print(errormsg)
-                output = None
-        return output
+        for col in arglist:
+            if col is None:
+                output = 0 # If not supplied, assume first column is intended
+            elif col in cols:
+                output = cols.index(col) # Convert to index
+            elif scu.isnumber(col):
+                try:
+                    cols[col]
+                except IndexError as E: # pragma: no cover
+                    errormsg = f'Column "{col}" is not a valid index; there are {len(cols)} columns'
+                    raise IndexError(errormsg) from E
+                output = col
+            else: # pragma: no cover
+                errormsg = f'Unrecognized column/column type "{col}" {type(col)}'
+                if die:
+                    raise TypeError(errormsg)
+                else:
+                    print(errormsg)
+                    output = None
+            outputlist.append(output)
+        if len(outputlist) == 1:
+            outputlist = outputlist[0]
+        return outputlist
+    
+    
+    def col_name(self, col=None, *args, die=True):
+        '''
+        Get the name of the column(s) with index ``col``.
+        
+        Similar to ``df.columns[col]``, and opposite of :meth:`df.col_index <dataframe.col_index>`.
+        
+        **Note**: This method always looks for named columns first. If ``col`` is 
+        name of a column, it will return ``col`` rather than ``columns[col]``. See
+        example below for more information.
+        
+        Args:
+            col (int/list): the column(s) to get the index of (return 0 if None)
+            args (list): additional column(s) to get the index of
+            die (bool): whether to raise an exception if the column could not be found (else, return None)
+        
+        **Examples**::
+            
+            df = sc.dataframe(dict(a=[1,2,3], b=[4,5,6], c=[7,8,9]))
+            df.col_name(1) # Returns 'b'
[…]
