### 2023-04-16 c312a5c
update odict init
