+++ sciris/sc_dataframe.py
@@ -55,7 +55,7 @@ class dataframe(pd.DataFrame):
     The dataframe can be used for both numeric and non-numeric data.
 
     | *New in version 2.0.0:* subclass pandas DataFrame
-    | *New in version 2.2.0:* "dtypes" argument; handling of item setting
+    | *New in version 3.0.0:* "dtypes" argument; handling of item setting
     '''
 
     def __init__(self, data=None, columns=None, nrows=None, dtypes=None, **kwargs):
@@ -102,7 +102,7 @@ class dataframe(pd.DataFrame):
         '''
         Set dtypes in-place (see :meth:`df.astype() <pandas.DataFrame.astype>` for the user-facing version)
         
-        *New in version 2.2.0.*
+        *New in version 3.0.0.*
         '''
         if not isinstance(dtypes, dict):
             dtypes = {col:dtype for col,dtype in zip(self.columns, dtypes)}
@@ -129,7 +129,7 @@ class dataframe(pd.DataFrame):
             df.col_index(1) # Returns 1
             df.col_index('a', 'c') # Returns [0, 2]
             
-        *New in version 2.2.0:* renamed from "_sanitizecols"; multiple arguments
+        *New in version 3.0.0:* renamed from "_sanitizecols"; multiple arguments
         '''
         arglist = scu.mergelists(col, list(args), keepnone=True)
         outputlist = []
@@ -181,7 +181,7 @@ class dataframe(pd.DataFrame):
             df.col_name('b') # Returns 'b'
             df.col_name(0, 2) # Returns ['a', 'c']
         
-        *New in version 2.2.0.*
+        *New in version 3.0.0.*
         '''
         arglist = scu.mergelists(col, list(args), keepnone=True)
         outputlist = []
@@ -366,7 +366,7 @@ class dataframe(pd.DataFrame):
             reset_index (bool): update the index
             inplace (bool): whether to modify in-place
         
-        *New in version 2.2.0:* improved dtype handling
+        *New in version 3.0.0:* improved dtype handling
         '''
         if newdf is None:
             newdf = self._constructor(data=newdata, columns=self.columns)
@@ -411,7 +411,7 @@ class dataframe(pd.DataFrame):
             df.appendrow(['cat', 3, 0.3])           # Append a list
             df.appendrow(dict(a='dog', b=4, c=0.7)) # Append a dict
             
-        *New in version 2.2.0:* renamed "value" to "row"; improved performance
+        *New in version 3.0.0:* renamed "value" to "row"; improved performance
         '''
         return self.concat(row, reset_index=reset_index, inplace=inplace)
     
@@ -420,7 +420,7 @@ class dataframe(pd.DataFrame):
         '''
         Alias to :meth:`appendrow() <dataframe.appendrow>`.
         
-        *New in version 2.2.0.*
+        *New in version 3.0.0.*
         '''
         return self.concat(row, reset_index=reset_index, inplace=inplace)
 
@@ -454,7 +454,7 @@ class dataframe(pd.DataFrame):
             df.insertrow(1, ['bar', 2, 0.2])           # Insert a list
             df.insertrow(0, dict(a='rat', b=0, c=0.7)) # Insert a dict
         
-        *New in version 2.2.0:* renamed "row" to "index"
+        *New in version 3.0.0:* renamed "row" to "index"
         '''
         before = self.iloc[:index,:]
         after  = self.iloc[index:,:]
@@ -494,7 +494,7 @@ class dataframe(pd.DataFrame):
             **kwargs (dict): passed to :func:`pd.concat() <pandas.concat>`
         
         | *New in version 2.0.2:* "inplace" defaults to False
-        | *New in version 2.2.0:* improved type handling
+        | *New in version 3.0.0:* improved type handling
[…]
