+++ docs/api/all.rst
@@ -1,0 +1,18 @@
+===========
+All modules
+===========
+
+Below is absolutely everything that's included in Sciris. 
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-module-template.rst
+   :recursive:
+   :hidden:
+
+   sciris
+
+.. toctree::
+   :maxdepth: 3
+
+   _autosummary/sciris
