+++ docs/api/all.rst
@@ -4,15 +4,14 @@ All modules
 
 Below is absolutely everything that's included in Sciris. 
 
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-module-template.rst
-   :recursive:
-   :hidden:
-
-   sciris
-
 .. toctree::
    :maxdepth: 3
 
    _autosummary/sciris
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-module-template.rst
+   :recursive:
+
+   sciris
