+++ docs/api/all.rst
@@ -4,11 +4,6 @@ All modules
 
 Below is absolutely everything that's included in Sciris. 
 
-.. toctree::
-   :maxdepth: 3
-
-   _autosummary/sciris
-
 .. autosummary::
    :toctree: _autosummary
    :template: custom-module-template.rst
