+++ docs/api/index.rst
@@ -2,10 +2,15 @@
 API reference
 =============
 
+.. currentmodule:: sciris
+
 There's a lot here! Use the links on the left for navigation.
 
+Note: functions in Sciris are of course organized into submodules. However, standard
+usage is to call the functions directly, e.g. :func:`sc.parallelize() <sc_parallel.parallelize>` instead
+of ``sc.sc_parallel.parallelize()``.
+
 .. toctree::
-   :maxdepth: 3
    :hidden:
 
    containers
