### 2023-04-16 95ae331
updating readthedocs config file
