+++ .readthedocs.yaml
@@ -10,10 +10,6 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.9"
-    # You can also specify other tool versions:
-    # nodejs: "19"
-    # rust: "1.64"
-    # golang: "1.19"
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
@@ -28,4 +24,4 @@ python:
    install:
     - method: pip
       path: .
-   - requirements: docs/requirements.txt
+    - requirements: docs/requirements.txt
