+++ .readthedocs.yaml
@@ -25,3 +25,4 @@ python:
     - method: pip
       path: .
     - requirements: docs/requirements.txt
+  system_packages: true # Saves on installation time
