+++ .readthedocs.yaml
@@ -21,7 +21,7 @@ sphinx:
 
 # Optionally declare the Python requirements required to build your docs
 python:
-   install:
+  install:
     - method: pip
       path: .
     - requirements: docs/requirements.txt
