### 2023-04-16 dcb07e8
update tests and test run script
