+++ .github/workflows/tests.yaml
@@ -4,12 +4,12 @@ on: [pull_request]
 jobs:
   install_and_test:
     runs-on: ubuntu-latest
-    timeout-minutes: 8
+    timeout-minutes: 5
     strategy:
       fail-fast: false
-      max-parallel: 1
+      max-parallel: 8
       matrix:
-        python-version: [ '3.9' ]
+        python-version: [ '3.9', '3.10', '3.11' ]
     name: Install and test
     steps:
       - name: Checkout sources
