+++ .github/workflows/tests.yaml
@@ -1,14 +1,14 @@
-name: Sciris CI workflow
+name: Sciris CI tests
 on: [pull_request]
 
 jobs:
   install_and_test:
-    runs-on: ubuntu-latest
     timeout-minutes: 5
     strategy:
       fail-fast: false
       max-parallel: 8
       matrix:
+        os: ['ubuntu-latest', 'windows-latest']
         python-version: [ '3.9', '3.10', '3.11' ]
     name: Install and test
     steps:
