+++ .github/workflows/tests.yaml
@@ -3,12 +3,12 @@ on: [pull_request]
 
 jobs:
   install_and_test:
+    runs-on: ['ubuntu-latest', 'windows-latest']
     timeout-minutes: 5
     strategy:
       fail-fast: false
       max-parallel: 8
       matrix:
-        os: ['ubuntu-latest', 'windows-latest']
         python-version: [ '3.9', '3.10', '3.11' ]
     name: Install and test
     steps:
