### 2023-04-16 acdaa77
fix config setup
