+++ .github/workflows/tests.yaml
@@ -3,13 +3,14 @@ on: [pull_request]
 
 jobs:
   install_and_test:
-    runs-on: ['ubuntu-latest', 'windows-latest']
     timeout-minutes: 5
     strategy:
       fail-fast: false
       max-parallel: 8
       matrix:
+        os: ['ubuntu-latest', 'windows-latest']
         python-version: [ '3.9', '3.10', '3.11' ]
+    runs-on: ${{ matrix.os }}
     name: Install and test
     steps:
       - name: Checkout sources
