+++ .github/workflows/tests.yaml
@@ -26,7 +26,7 @@ jobs:
         run: pip install -r requirements_test.txt
       - name: Run API tests
         working-directory: ./tests
-        run: pytest test_*.py --durations=0 --junitxml=test-results.xml # Run actual tests
+        run: pytest --durations=0 --junitxml=test-results.xml # Run actual tests
       - name: Upload test results
         uses: actions/upload-artifact@v1
         if: failure()
