### 2023-04-16 1aa70d9
update installation and testing setup
