+++ .github/workflows/tests.yaml
@@ -23,7 +23,7 @@ jobs:
         run: pip install -e .
       - name: Install tests
         working-directory: ./tests
-        run: pip install -r requirements_test.txt
+        run: pip install -r requirements.txt
       - name: Run API tests
         working-directory: ./tests
         run: pytest --durations=0 --junitxml=test-results.xml # Run actual tests
