+++ sciris/sc_parallel.py
@@ -126,7 +126,6 @@ class Parallel:
         self.reset()
         self.set_defaults()
         self.validate_args()
-        self.make_argslist()
         self.set_ncpus()
         self.set_method()
         return
@@ -140,10 +139,10 @@ class Parallel:
         self.argslist     = None
         self.method       = None
         self.pool         = None
-        self.manager      = mpi.Manager() # Create a manager for sharing resources across jobs
-        self.globaldict   = self.manager.dict() # Create a dict for sharing progress of each job
         self.map_func     = None
         self.is_async     = None
+        self.manager      = None
+        self.globaldict   = None
         self.jobs         = None
         self.results      = None
         self.success      = None
@@ -247,46 +246,6 @@ class Parallel:
         else:
             self.njobs = njobs
             
-        return
-    
-    
-    def make_argslist(self):
-        ''' Construct argument list '''
-
-        # Initialize
-        iterarg    = self.iterarg
-        iterkwargs = self.iterkwargs
-        argslist   = []
-        
-        # Check for embarrassingly parallel run -- should already be validated
-        if self.embarrassing:
-            iterarg = np.arange(iterarg)
-        
-        # Construct the argument list for each job
-        for index in range(self.njobs):
-            if iterarg is None:
-                iterval = None
-            else:
-                iterval = iterarg[index]
-            if iterkwargs is None:
-                iterdict = None
-            else:
-                if isinstance(iterkwargs, dict): # Dict of lists
-                    iterdict = {}
-                    for key,val in iterkwargs.items():
-                        iterdict[key] = val[index]
-                elif isinstance(iterkwargs, list): # List of dicts
-                    iterdict = iterkwargs[index]
-                else:  # pragma: no cover # Should be caught by previous checking, so shouldn't happen
-                    errormsg = f'iterkwargs type not understood ({type(iterkwargs)})'
-                    raise TypeError(errormsg)
-            taskargs = TaskArgs(func=self.func, index=index, njobs=self.njobs, iterval=iterval, iterdict=iterdict,
-                                args=self.args, kwargs=self.kwargs, maxcpu=self.maxcpu, maxmem=self.maxmem,
-                                interval=self.interval, embarrassing=self.embarrassing, callback=self.callback, 
-                                progress=self.progress, globaldict=self.globaldict, die=self.die)
-            argslist.append(taskargs)
-        
-        self.argslist = argslist
         return
     
     
@@ -396,16 +355,100 @@ class Parallel:
             errormsg = f'Invalid parallelizer "{self.parallelizer}"'
             raise ValueError(errormsg)
             
+        # Create a manager for sharing resources across jobs
+        if method == 'multiprocess': # Special case: can't share multiprocessing managers with multiprocess
+            manager = mp.Manager() 
+        else:
+            manager = mpi.Manager()
+        globaldict = manager.dict() # Create a dict for sharing progress of each job
+        
         # Reset
         self.pool       = pool
         self.map_func   = map_func
         self.is_async   = is_async
+        self.manager    = manager
+        self.globaldict = globaldict
         self.jobs       = None
         self.rawresults = None
         self.results    = None
         
         return
 
+
+    def make_argslist(self):
+        ''' Construct argument list '''
+
+        # Initialize
+        iterarg    = self.iterarg
+        iterkwargs = self.iterkwargs
+        argslist   = []
+        
+        # Check for embarrassingly parallel run -- should already be validated
+        if self.embarrassing:
+            iterarg = np.arange(iterarg)
+        
+        # Construct the argument list for each job
+        for index in range(self.njobs):
+            if iterarg is None:
+                iterval = None
+            else:
[…]
