### 2023-04-16 0d5531b
fix path checking for windows
