+++ tests/pytest.ini
@@ -1,3 +1,5 @@
 [pytest]
 filterwarnings =
     ignore::pytest.PytestReturnNotNoneWarning
+env =
+    SCIRIS_BACKEND=agg
