### 2023-04-17 9db240e
allow 2d input for scatter
