+++ docs/tutorials/tut_intro.rst
@@ -1,5 +1,0 @@
-==========
-Tutorial 1
-==========
-
-Test
