+++ docs/requirements.txt
@@ -1,10 +1,8 @@
-# Dependencies for testing/documentation
-pytest
-pytest-xdist
+# Dependencies for documentation
 pandoc
 pypandoc
 sphinx>=1.8
 sphinx-autodoc-typehints
 pydata-sphinx-theme
 sphinx-design
-recommonmark
+nbsphinx
