+++ docs/requirements.txt
@@ -5,4 +5,5 @@ sphinx>=1.8
 sphinx-autodoc-typehints
 pydata-sphinx-theme
 sphinx-design
+ipykernel
 nbsphinx
