+++ .readthedocs.yaml
@@ -2,8 +2,12 @@
 # Read the Docs configuration file
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 
-# Required
+# YAML version
 version: 2
+
+# Set the conda envrionment
+conda:
+  environment: docs/environment.yml
 
 # Set the version of Python and other tools you might need
 build:
@@ -25,4 +29,4 @@ python:
     - method: pip
       path: .
     - requirements: docs/requirements.txt
-  system_packages: true # Saves on installation time
+  system_packages: false # Saves on installation time, but fails
