+++ .readthedocs.yaml
@@ -2,18 +2,18 @@
 # Read the Docs configuration file
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 
-# YAML version
+# Required
 version: 2
+
+# Set the version of Python and other tools you might need
+build:
+  os: "ubuntu-20.04"
+  tools:
+    python: "mambaforge-4.10"
 
 # Set the conda envrionment
 conda:
   environment: docs/environment.yml
-
-# Set the version of Python and other tools you might need
-build:
-  os: ubuntu-22.04
-  tools:
-    python: "3.9"
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
@@ -28,5 +28,3 @@ python:
   install:
     - method: pip
       path: .
-    - requirements: docs/requirements.txt
-  system_packages: false # Saves on installation time, but fails
