+++ .readthedocs.yaml
@@ -10,6 +10,7 @@ build:
   os: "ubuntu-20.04"
   tools:
     python: "mambaforge-4.10"
+  commands: touch sciris/extras/__init__.py
 
 # Set the conda envrionment
 conda:
