+++ .readthedocs.yaml
@@ -10,7 +10,8 @@ build:
   os: "ubuntu-20.04"
   tools:
     python: "mambaforge-4.10"
-  commands: touch sciris/extras/__init__.py
+  commands: 
+    - touch sciris/extras/__init__.py
 
 # Set the conda envrionment
 conda:
