+++ .readthedocs.yaml
@@ -11,7 +11,11 @@ build:
   tools:
     python: "mambaforge-4.10"
   commands: 
-    - touch sciris/extras/__init__.py
+    - pwd
+    - pip install -e .
+    - cd docs
+    - ./build_docs
+
 
 # Set the conda envrionment
 conda:
