+++ .readthedocs.yaml
@@ -15,7 +15,7 @@ build:
     - ls
     - ls docs
     - pip install -e .
-    - pip install ./docs/requirements.txt
+    - pip install -r ./docs/requirements.txt
     - cd docs; build_docs
 
 
