+++ .readthedocs.yaml
@@ -16,7 +16,7 @@ build:
     - ls docs
     - pip install -e .
     - pip install -r ./docs/requirements.txt
-    - cd docs; build_docs
+    - cd docs; ./build_docs
 
 
 # Set the conda envrionment
