+++ .readthedocs.yaml
@@ -17,6 +17,8 @@ build:
     - pip install -e .
     - pip install -r ./docs/requirements.txt
     - cd docs; ./build_docs
+    - mkdir -p _readthedocs/
+    - cp ./docs/_build/* _readthedocs/
 
 
 # Set the conda envrionment
