+++ .readthedocs.yaml
@@ -7,23 +7,9 @@ version: 2
 
 # Set the version of Python and other tools you might need
 build:
-  os: "ubuntu-20.04"
+  os: ubuntu-22.04
   tools:
-    python: "mambaforge-4.10"
-  commands: 
-    - pwd
-    - ls
-    - ls docs
-    - pip install -e .
-    - pip install -r ./docs/requirements.txt
-    - cd docs; ./build_docs
-    - mkdir -p _readthedocs/
-    - cp -r ./docs/_build/* _readthedocs/
-
-
-# Set the conda envrionment
-conda:
-  environment: docs/environment.yml
+    python: "3.9"
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
@@ -38,3 +24,5 @@ python:
   install:
     - method: pip
       path: .
+    - requirements: docs/requirements.txt
+  system_packages: false # Saves on installation time
