+++ .readthedocs.yaml
@@ -7,9 +7,13 @@ version: 2
 
 # Set the version of Python and other tools you might need
 build:
-  os: ubuntu-22.04
+  os: "ubuntu-20.04"
   tools:
-    python: "3.9"
+    python: "mambaforge-4.10"
+
+# Set the conda envrionment
+conda:
+  environment: docs/environment.yml
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
