### 2023-04-17 e6f7d62
tidy up conf.py
