+++ sciris/sc_asd.py
@@ -1,8 +1,11 @@
 '''
-Adaptive stochastic descent optimization algorithm, building on scipy.optimize.
+Adaptive stochastic descent optimization algorithm, building on :mod:`scipy.optimize`.
 
-This algorithm is published as "Optimization by adaptive stochastic descent" by
-Kerr et al. (2018) (DOI: https://doi.org/10.1371/journal.pone.0192944).
+This algorithm is published as:
+
+  Kerr CC, Dura-Bernal S, Smolinski TG, Chadderdon GL, Wilson DP (2018).
+  **Optimization by Adaptive Stochastic Descent**. *PLoS ONE* 13(3): e0192944. 
+  https://doi.org/10.1371/journal.pone.0192944
 '''
 
 import numpy as np
