+++ sciris/sc_asd.py
@@ -8,9 +8,8 @@ This algorithm is published as:
   https://doi.org/10.1371/journal.pone.0192944
 '''
 
+import time
 import numpy as np
-import numpy.random as nr
-from time import time
 from . import sc_utils as scu
 from . import sc_printing as scp
 from . import sc_odict as sco
@@ -93,11 +92,10 @@ def asd(function, x, args=None, stepsize=0.1, sinc=2, sdec=2, pinc=2, pdec=2,
         Optimization by adaptive stochastic descent.
         PLOS ONE 13 (3), e0192944.
 
-    Version: 2019jul08
+    *New in version 2.2.0:* Uses its own random number stream
     """
-    if randseed is not None:
-        nr.seed(int(randseed)) # Don't reset it if not supplied
-        if verbose >= 2: print(f'ASD: Launching with random seed is {randseed}; sample: {nr.random()}')
+    rng = np.random.default_rng(int(randseed))
+    if verbose >= 2: print(f'ASD: Launching with random seed is {randseed}')
 
     def consistentshape(userinput, origshape=False):
         """
@@ -178,7 +176,7 @@ def asd(function, x, args=None, stepsize=0.1, sinc=2, sdec=2, pinc=2, pdec=2,
 
     # Loop
     count = 0 # Keep track of how many iterations have occurred
-    start = time() # Keep track of when we begin looping
+    start = time.time() # Keep track of when we begin looping
     offset = ' ' * 4 # Offset the print statements
     exitreason = 'Unknown exit reason' # Catch everything else
     while True:
@@ -190,7 +188,7 @@ def asd(function, x, args=None, stepsize=0.1, sinc=2, sdec=2, pinc=2, pdec=2,
         cumprobs = np.cumsum(probabilities) # Calculate the cumulative distribution
         inrange = False
         for r in range(maxrangeiters): # Try to find parameters within range
-            choice = np.flatnonzero(cumprobs > nr.random())[0] # Choose a parameter and upper/lower at random
+            choice = np.flatnonzero(cumprobs > rng.random())[0] # Choose a parameter and upper/lower at random
             par = np.mod(choice, nparams) # Which parameter was chosen
             pm = np.floor((choice) / nparams) # Plus or minus
             newval = x[par] + ((-1)**pm) * stepsizes[choice] # Calculate the new vector
@@ -232,7 +230,7 @@ def asd(function, x, args=None, stepsize=0.1, sinc=2, sdec=2, pinc=2, pdec=2,
                 if verbose >= 1: print('ASD: Warning, objective function returned NaN')
         if verbose > 0 and not (count % max(1, int(1.0/verbose))): # Print out every 1/verbose steps
             orig, best, new, diff = scp.sigfig([fvalorig, fvalold, fvalnew, fvalnew-fvalold])
-            print(offset + label + f' step {count} ({time()-start:0.1f} s) {flag} (orig:{orig} | best:{best} | new:{new} | diff:{diff})')
+            print(offset + label + f' step {count} ({time.time()-start:0.1f} s) {flag} (orig:{orig} | best:{best} | new:{new} | diff:{diff})')
 
         # Store output information
         fvals[count] = fval # Store objective function evaluations
@@ -243,7 +241,7 @@ def asd(function, x, args=None, stepsize=0.1, sinc=2, sdec=2, pinc=2, pdec=2,
             exitreason = 'Maximum iterations reached'
             break
         if (time() - start) > maxtime: # pragma: no cover
-            strtime, strmax = scp.sigfig([(time()-start), maxtime])
+            strtime, strmax = scp.sigfig([(time.time()-start), maxtime])
             exitreason = f'Time limit reached ({strtime} > {strmax})'
             break
[…]
