+++ docs/tutorials/index.rst
@@ -6,4 +6,13 @@ Tutorials
    :maxdepth: 3
 
    tut_intro
+   tut_arrays
+   tut_containers
+   tut_files
+   tut_printing
+   tut_plotting
+   tut_parallel
+   tut_dates
+   tut_utils
+   tut_advanced
 
