+++ docs/tutorials/index.rst
@@ -2,12 +2,92 @@
 Tutorials
 =========
 
+These tutorials illustrate the main features of Sciris. They roughly make the most sense in the order listed, but can be done in any order. Each one should take a few minutes to skim through, or 15-30 minutes to go through in detail. Each tutorial contains a link to an interactive version running on `Binder <http://mybinder.org>`_.
+
+Questions? Reach out to us at info@sciris.org. 
+
+.. grid:: 2
+    :gutter: 4
+
+    .. grid-item-card:: 1 Whirlwind tour
+        :link: tut_intro
+        :link-type: doc
+
+        A quick high-level overview of all of Sciris' most important features.
+
+
+    .. grid-item-card:: 2 Array tools
+        :link: tut_arrays
+        :link-type: doc
+
+        Arrays are at the core of scientific computing, and this tutorial will go through some tools that will make it easier to work with them.
+
+
+    .. grid-item-card:: 3 Dictionaries and dataframes
+        :link: tut_dicts
+        :link-type: doc
+
+        Data are only as good as the container they're stored in (?), so this tutorial goes through Sciris' main container types.
+
+
+    .. grid-item-card:: 4 Files and versioning
+        :link: tut_files
+        :link-type: doc
+        
+        All that data has to come from and go somewhere: this tutorial covers saving and loading files, including with version metadata.
+
+
+    .. grid-item-card:: 5 Printing
+        :link: tut_printing
+        :link-type: doc
+
+        This tutorial covers tools for quickly printing out information about objects and data.
+
+
+    .. grid-item-card:: 6 Plotting
+        :link: tut_plotting
+        :link-type: doc
+
+        Without a plot, did science even happen? This tutorial covers Sciris' plotting tools and shortcuts.
+
+
+    .. grid-item-card:: 7 Parallelization and profiling
+        :link: tut_parallel
+        :link-type: doc
+
+        When you need answers faster: this tutorial covers running your code in parallel, and profiling to figure out where to find performance improvements.
+
+
+    .. grid-item-card:: 8 Dates and times
+        :link: tut_dates
+        :link-type: doc
+        
+        Compared to nice, simple numbers, dates are a pain to work with. This tutorial goes through some ways to work with them more easily.
+
+
+    .. grid-item-card:: 9 Miscellaneous utilities
+        :link: tut_utils
+        :link-type: doc
+        
+        Sometimes, you just need to do some random task. There's a good chance it'll be covered by one of these tools.
+
+
+    .. grid-item-card:: 10 Advanced features
+        :link: tut_advanced
+        :link-type: doc
+        
+        You probably won't often need these tools, but they're there waiting for you in case you do.
+
+
+Full contents
+-------------
+
 .. toctree::
    :maxdepth: 3
 
    tut_intro
    tut_arrays
-   tut_containers
+   tut_dicts
    tut_files
    tut_printing
    tut_plotting
