+++ docs/conf.py
@@ -119,7 +119,10 @@ intersphinx_mapping = {
 
 def setup(app):
     app.add_css_file("custom.css")
+    app.add_css_file("collapsible.css")
+    app.add_js_file("collapsible.js")
     app.add_js_file("copybutton.js")
+    app.add_js_file("hi.js")
 
 
 # Modify this to not rerun the Jupyter notebook cells -- usually set by build_docs
