+++ sciris/sc_settings.py
@@ -500,6 +500,8 @@ class ScirisOptions(sco.objdict):
         as part of a ``with`` block to set the style just for that block (using
         this function outsde of a with block and with ``use=False`` has no effect, so
         don't do that!).
+        
+        Note: you can also just use :func:`pl.style.context() <matplotlib.style.context>`.
 
         Args:
             style_args (dict): a dictionary of style arguments
