+++ docs/Makefile
@@ -23,6 +23,7 @@ help:
 clean: # The -f is to ignore missing files, not to force remove
 	rm -rf $(BUILDDIR)
 	rm -rf $(AUTODIR)
+	rm -vf tutorials/*.obj # Do not do recursively and hard-code to make this safer!
 
 # Catch-all target: route all unknown targets to Sphinx using the new
 # "make mode" option.  $(O) is meant as a shortcut for $(SPHINXOPTS).
