### 2023-04-17 644cfca
working on array tutorial
