+++ docs/requirements.txt
@@ -5,5 +5,6 @@ sphinx>=1.8
 sphinx-autodoc-typehints
 pydata-sphinx-theme
 sphinx-design
-ipykernel
-nbsphinx
+ipykernel # For tutorials
+nbsphinx # For tutorials
+dask # For parallel examples
