### 2023-04-17 a728b29
example benchmarking startup, same as just importing numpy, pandas, and matplotlib
