### 2023-04-17 7032232
fix figure handling for 3d plots
