### 2023-04-18 e5daf2e
add manifest file and more plot tests
