+++ sciris/sc_settings.py
@@ -26,6 +26,10 @@ from . import sc_printing as scp
 
 __all__ = ['ScirisOptions', 'options', 'parse_env', 'help']
 
+#%% Define plotting options
+# See also simple.mplstyle and fancy.mplstyle files, which can be generated by:
+#   sc.saveyaml('simple.mplstyle', sc.sc_settings.rc_simple)
+#   sc.saveyaml('fancy.mplstyle',  sc.sc_settings.rc_fancy)
 
 # Define simple plotting options -- similar to Matplotlib default
 rc_simple = {
@@ -37,13 +41,14 @@ rc_simple = {
     'legend.frameon':    False,
 }
 
-# Define default plotting options -- based on Seaborn
+# Define default plotting options -- loosely inspired by Seaborn
 rc_fancy = scu.dcp(rc_simple)
 rc_fancy.update({
     'axes.facecolor': '#f2f2ff',
     'axes.grid':      True,
     'grid.color':     'white',
-    'grid.linewidth': 1,
+    'grid.linewidth':  1,
+    'lines.linewidth': 2,
 })
 
 
