+++ sciris/sc_fileio.py
@@ -520,7 +520,7 @@ def unzip(filename=None, outfolder='.', folder=None, members=None):
 
 
 def savezip(filename=None, files=None, data=None, folder=None, sanitizepath=True, 
-            basename=False, relpath=True, tobytes=True, verbose=True, **kwargs):
+            basename=False, tobytes=True, verbose=True, **kwargs):
     '''
     Create a zip file from the supplied list of files (or less commonly, supplied data)
 
@@ -531,7 +531,6 @@ def savezip(filename=None, files=None, data=None, folder=None, sanitizepath=True
         folder (str): optional additional folder for the filename
         sanitizepath (bool): whether to sanitize the path prior to saving
         basename (bool): whether to use only the file's basename as the name inside the zip file (otherwise, store folder info)
-        relpath (bool): keep only relative paths, i.e. remove any folders that are common to all input files
         tobytes (bool): if data is provided, convert it automatically to bytes (otherwise, up to the user)
         verbose (bool): whether to print progress
         kwargs (dict): passed to :func:`sc.save() <save>`
@@ -561,55 +560,24 @@ def savezip(filename=None, files=None, data=None, folder=None, sanitizepath=True
     else:
         
         # Handle subfolders
-        extfilelist = scu.dcp(origfilelist) # Keep all the original
-        for ofile in origfilelist:
-            if ofile.is_dir():
-                contents = getfilelist(ofile, recursive=True, aspath=True)
+        extfilelist = scu.dcp(origfilelist) # An extended file list, including recursion into subfolders
+        for orig in origfilelist:
+            if orig.is_dir():
+                contents = getfilelist(orig, abspath=False, recursive=True, aspath=True)
                 extfilelist.extend(contents[1:]) # Skip the first entry since it's the folder that's already in the list
-                print('DEBUG adding', contents[1:])
         
-        # Remove duplicates
+        # Remove duplicates to create the final file list
         filelist = []
         for efile in extfilelist:
             if efile not in filelist:
                 filelist.append(efile)
-            else:
-                print('DEBUG', efile)
-        
-        # Handle relative paths
-        filenames = scu.dcp(filelist)
-        if relpath:
-            print('DEBUG')
-            print(filenames)
-            pathparts = [f.parts for f in filelist]
-            print('pathparts')
-            print(pathparts)
-            transparts = scu.transposelist(pathparts, fix_uneven=True)
-            print('transparts')
-            print(transparts)
-            done = len(pathparts) <= 1 # Don't look for relative paths if only one file
-            while not done:
-                if len(set(transparts[0])) == 1: # They're all the same
-                    transparts = transparts[1:]
-                    print('trimmed ', set(transparts[0]))
-                    print(done, transparts)
-                else:
-                    print('did NOT trim', set(transparts[0]))
-                    done = True
-            pathparts = scu.transposelist(transparts) # Transpose back
-            print('pathparts')
-            print(pathparts)
-            filenames = [path(*row) for row in pathparts]
-            print('filelist')
-            print(filenames)
             
     # Write zip file
     with ZipFile(fullpath, 'w') as zf: # Create the zip file
         if data is None: # Main use case, save files
-            for thisfile,thisname in zip(filelist, filenames):
+            for thisfile in filelist:
                 thispath = makefilepath(filename=thisfile, abspath=False, makedirs=False)
-                if basename: # pragma: no cover
-                    thisname = os.path.basename(thisname)
+                thisname = os.path.basename(thisfile) if basename else thisfile
                 zf.write(thispath, thisname) # Actually save
         else: # Alternatively, save data
             for key,val in data.items():
