### 2023-04-18 b4efdd7
updates to jsonify
