### 2023-04-19 8b18c11
working on plotting tutorial
