+++ sciris/sc_datetime.py
@@ -599,21 +599,74 @@ def tic():
     return _tictime    # Return the same stored number
 
 
+def _convert_time_unit(unit, elapsed=None):
+    ''' Convert between different units of time; not for the user '''
+    
+    # Shortcut for speed
+    if unit == 's':
+        return 1, 's'
 
-def toc(start=None, label=None, baselabel=None, sigfigs=None, reset=False, output=False, doprint=None, elapsed=None):
+    # Standard use case
+    else:
+    
+        # Define the mapping -- in order of expected usage frequency for speed
+        mapping = {
+            's'  : dict(factor=   1, aliases=[None, 'default', 's', 'sec', 'secs', 'second', 'seconds']),
+            'ms' : dict(factor=1e-3, aliases=['ms', 'milisecond', 'miliseconds']),
+            'μs' : dict(factor=1e-6, aliases=['us', 'μs', 'microsecond', 'microseconds']),
+            'ns' : dict(factor=1e-9, aliases=['ns', 'nanosecond', 'nanoseconds']),
+            'min': dict(factor=  60, aliases=['m', 'min', 'mins', 'minute', 'minutes']),
+            'hr' : dict(factor=3600, aliases=['h', 'hr', 'hrs', 'hour', 'hours']),
+        }
+        
+        # Handle 'auto'
+        if unit == 'auto':
+            if elapsed is None:  unit = 's'
+            elif elapsed < 1e-7: unit = 'ns'
+            elif elapsed < 1e-4: unit = 'μs'
+            elif elapsed < 1e-1: unit = 'ms'
+            else:                unit = 's'
+    
+        # Perform the mapping
+        factor = None
+        for label,entry in mapping.items():
+            if unit in [label, entry['factor']] + entry['aliases']:
+                factor = entry['factor']
+                break
+        if factor is None:
+            errormsg = f'Could not understand "{unit}"; all possible values are:\n{mapping}'
+            raise ValueError(errormsg)
+    
+    return factor, label
+
+
+def toc(start=None, label=None, baselabel=None, sigfigs=None, reset=False, unit='s',
+        output=False, doprint=None, elapsed=None):
     '''
     With :func:`sc.tic() <tic>`, a little pair of functions to calculate a time difference. See
     also :class:`sc.timer() <timer>`.
+    
+    By default, output is displayed in seconds. You can change this with the ``unit``
+    argument, which can be a string or a float:
+        
+        - 'hr' or 3600
+        - 'min' or 60
+        - 's' or 1 (default)
+        - 'ms' or 1e-3
+        - 'us' or 1e-6
+        - 'ns' or 1e-9
+        - 'auto' to choose an appropriate unit
 
     Args:
-        start     (float): the starting time, as returned by e.g. :func:`sc.tic() <tic>`
-        label     (str): optional label to add
-        baselabel (str): optional base label; default is "Elapsed time: "
-        sigfigs   (int): number of significant figures for time estimate
+        start    (float): the starting time, as returned by e.g. :func:`sc.tic() <tic>`
+        label      (str): optional label to add
+        baselabel  (str): optional base label; default is "Elapsed time: "
+        sigfigs    (int): number of significant figures for time estimate
         reset     (bool): reset the time; like calling :func:`sc.toctic() <toctic>` or :func:`sc.tic() <tic>` again
+        unit (str/float): the unit of time to display; see options above
         output    (bool): whether to return the output (otherwise print); if output='message', then return the message string; if output='both', then return both
         doprint   (bool): whether to print (true by default)
-        elapsed   (float): use a pre-calculated elapsed time instead of recalculating (not recommneded)
[…]
