### 2023-04-20 f8d5849
add example scripts
