+++ docs/environment.yml
@@ -1,130 +1,0 @@
-name: sciris_docs_v3.0.0
-channels:
-  - forge
-  - defaults
-dependencies:
-  - _libgcc_mutex=0.1=main
-  - _openmp_mutex=5.1=1_gnu
-  - ca-certificates=2023.01.10=h06a4308_0
-  - certifi=2022.12.7=py39h06a4308_0
-  - ld_impl_linux-64=2.38=h1181459_1
-  - libffi=3.4.2=h6a678d5_6
-  - libgcc-ng=11.2.0=h1234567_1
-  - libgomp=11.2.0=h1234567_1
-  - libstdcxx-ng=11.2.0=h1234567_1
-  - ncurses=6.4=h6a678d5_0
-  - openssl=1.1.1t=h7f8727e_0
-  - pip=23.0.1=py39h06a4308_0
-  - python=3.9.16=h7a1cb2a_2
-  - readline=8.2=h5eee18b_0
-  - setuptools=65.6.3=py39h06a4308_0
-  - sqlite=3.41.2=h5eee18b_0
-  - tk=8.6.12=h1ccaba5_0
-  - wheel=0.38.4=py39h06a4308_0
-  - xz=5.2.10=h5eee18b_1
-  - zlib=1.2.13=h5eee18b_0
-  - pip:
-      - accessible-pygments==0.0.4
-      - alabaster==0.7.13
-      - asttokens==2.2.1
-      - attrs==23.1.0
-      - babel==2.12.1
-      - backcall==0.2.0
-      - beautifulsoup4==4.12.2
-      - bleach==6.0.0
-      - charset-normalizer==3.1.0
-      - comm==0.1.3
-      - contourpy==1.0.7
-      - cycler==0.11.0
-      - debugpy==1.6.7
-      - decorator==5.1.1
-      - defusedxml==0.7.1
-      - dill==0.3.6
-      - docutils==0.19
-      - et-xmlfile==1.1.0
-      - executing==1.2.0
-      - fastjsonschema==2.16.3
-      - fonttools==4.39.3
-      - gitdb==4.0.10
-      - gitpython==3.1.31
-      - idna==3.4
-      - imagesize==1.4.1
-      - importlib-metadata==6.4.1
-      - importlib-resources==5.12.0
-      - ipykernel==6.22.0
-      - ipython==8.12.0
-      - jedi==0.18.2
-      - jellyfish==0.11.2
-      - jinja2==3.1.2
-      - jsonpickle==3.0.1
-      - jsonschema==4.17.3
-      - jupyter-client==8.2.0
-      - jupyter-core==5.3.0
-      - jupyterlab-pygments==0.2.2
-      - kiwisolver==1.4.4
-      - line-profiler==4.0.3
-      - markupsafe==2.1.2
-      - matplotlib==3.7.1
-      - matplotlib-inline==0.1.6
-      - memory-profiler==0.61.0
-      - mistune==2.0.5
-      - multiprocess==0.70.14
-      - nbclient==0.7.3
-      - nbconvert==7.3.1
-      - nbformat==5.8.0
-      - nbsphinx==0.9.1
-      - nest-asyncio==1.5.6
-      - numpy==1.24.2
-      - openpyxl==3.1.2
-      - packaging==23.1
-      - pandas==2.0.0
-      - pandoc==2.3
-      - pandocfilters==1.5.0
-      - parso==0.8.3
-      - pexpect==4.8.0
-      - pickleshare==0.7.5
-      - pillow==9.5.0
-      - platformdirs==3.2.0
-      - plumbum==1.8.1
-      - ply==3.11
-      - prompt-toolkit==3.0.38
-      - psutil==5.9.4
-      - ptyprocess==0.7.0
[…]
