+++ .readthedocs.yaml
@@ -9,11 +9,7 @@ version: 2
 build:
   os: "ubuntu-20.04"
   tools:
-    python: "mambaforge-4.10"
-
-# Set the conda envrionment
-conda:
-  environment: docs/environment.yml
+    python: "3.9"
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
@@ -29,4 +25,4 @@ python:
     - method: pip
       path: .
     - requirements: docs/requirements.txt
-  system_packages: false # Saves on installation time
+  system_packages: false # Saves on installation time if True
