### 2023-04-20 2b17fad
fix thread copy
