+++ tests/test_utils.py
@@ -318,10 +318,10 @@ def test_misc():
     assert isinstance(lazynp, sc.LazyModule)
     lazynp.array(0)
     assert not isinstance(lazynp, sc.LazyModule)
-    test_other = sc.importbyname(path='./test_other.py', variable='test_other')
-    assert 'test_options' in dir(test_other)
-    test_other2 = sc.importbypath(path='./test_other.py')
-    assert 'test_options' in dir(test_other2)
+    test_set = sc.importbyname(path='./test_settings.py', variable='test_set')
+    assert 'test_options' in dir(test_set)
+    test_set2 = sc.importbypath(path='./test_settings.py')
+    assert 'test_options' in dir(test_set2)
 
     print('\nTesting get_caller()')
     o.caller = sc.getcaller(includeline=True)
