### 2023-04-21 2836c4f
update isjupyter
