### 2023-04-21 c7b668e
do not attempt to auto-set jupyter
