### 2023-04-21 7592690
fix vectocolor test
