+++ sciris/sc_version.py
@@ -5,5 +5,5 @@ Version and license information.
 __all__ = ['__version__', '__versiondate__', '__license__']
 
 __version__      = '3.0.0'
-__versiondate__  = '2023-04-17'
+__versiondate__  = '2023-04-20'
 __license__      = f'Sciris {__version__} ({__versiondate__}) – © 2014-2023 by the Sciris Development Team'
