+++ CHANGELOG.rst
@@ -8,7 +8,7 @@ All major updates to Sciris are documented here.
 By import convention, components of the Sciris library are listed beginning with ``sc.``, e.g. ``sc.odict()``.
 
 
-Version 3.0.0 (2023-04-17)
+Version 3.0.0 (2023-04-20)
 --------------------------
 
 This version's major changes include:
@@ -22,68 +22,84 @@ This version's major changes include:
 Improvements and new features
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-Parallelization
-^^^^^^^^^^^^^^^
+1. Parallelization
+^^^^^^^^^^^^^^^^^^
 #. There is a new :class:`sc.Parallel() <sc_parallel.Parallel>` class, which is used to implement the (more or less unchanged) :func:`sc.parallelize() <sc_parallel.parallelize>` function.
 #. :func:`sc.parallelize() <sc_parallel.parallelize>` now has a ``progress`` argument that will show a progress bar; the ``returnpool`` argument has been removed (use :class:`sc.Parallel() <sc_parallel.Parallel>` instead).
 
-Dataframe
-^^^^^^^^^
-#. Better handling of (and preservation) of ``dtypes`` for dataframe columns, including a new :meth:`df.set_dtypes() <sc_dataframe.dataframe.set_dtypes>` method.
+2. Dataframe
+^^^^^^^^^^^^
 #. Better implementation of underlying logic, leading to significant performance increases in some cases (e.g., iteratively appending rows).
+#. Better handling of (and preservation) of ``dtypes`` for dataframe columns, including a new :meth:`df.set_dtypes() <sc_dataframe.dataframe.set_dtypes>` method.
 
 Time/date
-^^^^^^^^^
-#. New :func:`sc.time() <sc_datetime.time>` alias for :func:`time.time()`.
+^^^^^^^^^^^^
 #. Support for ``pandas`` and ``Numpy`` datetime objects.
 #. New :class:`sc.timer <sc_datetime.timer>` attributes and methods: :obj:`sc.rawtimings <sc_datetime.timer.rawtimings>`, :meth:`sc.sum() <sc_datetime.timer.sum>`, :meth:`sc.min() <sc_datetime.timer.min>`, :meth:`sc.max() <sc_datetime.timer.max>`, :meth:`sc.mean() <sc_datetime.timer.mean>`, :meth:`sc.std() <sc_datetime.timer.std>`.
+#. New :func:`sc.time() <sc_datetime.time>` alias for :func:`time.time()`.
 #. More accurate computation of self-time in :func:`sc.timedsleep() <sc_datetime.timedsleep>`.
 
-Files
-^^^^^
+3. Files
+^^^^^^^^
 #. A new function :func:`sc.unzip() <sc_fileio.unzip>` extracts zip files to disk, while :func:`sc.loadzip() <sc_fileio.loadzip>` now defaults to loading the zip file contents to memory.
 #. If a saved file can't be unpickled, :func:`sc.load() <sc_fileio.load>` now defaults to using ``dill``, and has more robust error handling (see also "versioning" updates below).
 #. :func:`sc.makefilepath() <sc_fileio.makefilepath>` now defaults to ``makedirs=False``.
 #. File save functions now make new subfolders by default.
 #. :func:`sc.save() <sc_fileio.save>` now has an ``allow_empty`` argument (instead of ``die='never'``).
+#. :func:`sc.glob() <sc_fileio.glob>` is a new alias for :func:`sc.getfilelist() <sc_fileio.getfilelist>`.
 
-Printing
-^^^^^^^^
+4. Printing
+^^^^^^^^^^^
[…]
