+++ CHANGELOG.rst
@@ -27,41 +27,57 @@ Improvements and new features
 #. There is a new :class:`sc.Parallel() <sc_parallel.Parallel>` class, which is used to implement the (more or less unchanged) :func:`sc.parallelize() <sc_parallel.parallelize>` function.
 #. :func:`sc.parallelize() <sc_parallel.parallelize>` now has a ``progress`` argument that will show a progress bar; the ``returnpool`` argument has been removed (use :class:`sc.Parallel() <sc_parallel.Parallel>` instead).
 
+
 2. Dataframe
 ^^^^^^^^^^^^
 #. Better implementation of underlying logic, leading to significant performance increases in some cases (e.g., iteratively appending rows).
+#. Numerous methods have been renamed, modified, or added, specifically: ``append``, ``col_index``, ``col_name``, ``findind``, ``findinds``, ``merge``, ``popcols``, ``poprow``, ``poprows``, and ``sort``.
+#. Keyword arguments are now interpreted as columns, e.g. ``df = sc.dataframe(a=[1,2], b=[3,4])``.
 #. Better handling of (and preservation) of ``dtypes`` for dataframe columns, including a new :meth:`df.set_dtypes() <sc_dataframe.dataframe.set_dtypes>` method.
+#. Dataframes now support equality checks.
 
-Time/date
+
+3. Time/date
 ^^^^^^^^^^^^
 #. Support for ``pandas`` and ``Numpy`` datetime objects.
 #. New :class:`sc.timer <sc_datetime.timer>` attributes and methods: :obj:`sc.rawtimings <sc_datetime.timer.rawtimings>`, :meth:`sc.sum() <sc_datetime.timer.sum>`, :meth:`sc.min() <sc_datetime.timer.min>`, :meth:`sc.max() <sc_datetime.timer.max>`, :meth:`sc.mean() <sc_datetime.timer.mean>`, :meth:`sc.std() <sc_datetime.timer.std>`.
+#. :class:`sc.timer <sc_datetime.timer>` now displays time in human-appropriate units (e.g., 3.4 μs instead of 0.0000034 s) by default, or accepts a ``unit`` argument.
 #. New :func:`sc.time() <sc_datetime.time>` alias for :func:`time.time()`.
+#. :func:`sc.datedelta() <sc_datetime.datedelta>` can now operate on a list of dates.
+#. :func:`sc.randsleep() <sc_datetime.randsleep>` now accepts a ``seed`` argument.
 #. More accurate computation of self-time in :func:`sc.timedsleep() <sc_datetime.timedsleep>`.
 
-3. Files
+
+4. Files
 ^^^^^^^^
-#. A new function :func:`sc.unzip() <sc_fileio.unzip>` extracts zip files to disk, while :func:`sc.loadzip() <sc_fileio.loadzip>` now defaults to loading the zip file contents to memory.
+#. A new function :func:`sc.unzip() <sc_fileio.unzip>` extracts zip files to disk, while :func:`sc.loadzip() <sc_fileio.loadzip>` now defaults to loading the zip file contents to memory. :func:`sc.savezip() <sc_fileio.savezip>` can now save both data and files, and its ``filelist`` argument has been renamed ``files``.
 #. If a saved file can't be unpickled, :func:`sc.load() <sc_fileio.load>` now defaults to using ``dill``, and has more robust error handling (see also "versioning" updates below).
 #. :func:`sc.makefilepath() <sc_fileio.makefilepath>` now defaults to ``makedirs=False``.
-#. File save functions now make new subfolders by default.
+#. File save functions now make new subfolders by default
 #. :func:`sc.save() <sc_fileio.save>` now has an ``allow_empty`` argument (instead of ``die='never'``).
[…]
