### 2023-04-21 b1667c9
improve heading defaults
