+++ sciris/_extras/README.md
@@ -1,8 +1,7 @@
-======
-Extras
-======
+# Extras
 
 This folder contains two files:
+
 - ``ansicolors.py`` is a copy of https://github.com/jonathaneunice/colors/, to avoid install issues on conda;
 - ``legacy.py`` is a graveyard for old (but conceivably, if improbably, still useful) Sciris functions.
 
