### 2023-04-21 #5076 @cliffckerr
Sciris v3.0.0 has just been released, which includes many of the reviewer suggestions, such as [tutorials](https://sciris.readthedocs.io/en/latest/tutorials/index.html). The full set of changes (which I don't expect you to review!!) is [here](https://sciris.readthedocs.io/en/latest/whatsnew.html#version-3-0-0-2023-04-20). Thanks!
