### 2023-04-29 eb860d9
Add is_empty() method
