### 2023-04-29 414506b
#369 Truncate each row string to maxwidth; insert newline.
