### 2023-04-29 91cc034
Implement #184 basic functionality for one path in the svg file
