### 2023-04-29 a41f557
Extend functionality of svg2mpl (#184) to support extraction of multiple paths; include basic input validation
