### 2023-05-05 3ba8d17
add support for globaldict customization
