+++ sciris/sc_parallel.py
@@ -95,7 +95,7 @@ class Parallel:
     '''
     def __init__(self, func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncpus=None, 
                  maxcpu=None, maxmem=None, interval=None, parallelizer=None, serial=False, 
-                 progress=True, callback=None, globaldict=None, label=None, die=True, **func_kwargs):
+                 progress=False, callback=None, globaldict=None, label=None, die=True, **func_kwargs):
         
         # Store input arguments
         self.func         = func
@@ -111,7 +111,7 @@ class Parallel:
         self.serial       = serial
         self.progress     = progress
         self.callback     = callback
-        self.inputdict    = scu.mergedicts(globaldict)
+        self.inputdict    = globaldict
         self.label        = label
         self.die          = die
         
@@ -394,7 +394,10 @@ class Parallel:
         # Check for embarrassingly parallel run -- should already be validated
         if self.embarrassing:
             iterarg = np.arange(iterarg)
-        
+            
+        # Check for additional global arguments
+        useglobal = True if self.inputdict else False
+            
         # Construct the argument list for each job
         for index in range(self.njobs):
             if iterarg is None:
@@ -416,7 +419,7 @@ class Parallel:
             taskargs = TaskArgs(func=self.func, index=index, njobs=self.njobs, iterval=iterval, iterdict=iterdict,
                                 args=self.args, kwargs=self.kwargs, maxcpu=self.maxcpu, maxmem=self.maxmem,
                                 interval=self.interval, embarrassing=self.embarrassing, callback=self.callback, 
-                                progress=self.progress, globaldict=self.globaldict, die=self.die)
+                                progress=self.progress, globaldict=self.globaldict, useglobal=useglobal, die=self.die)
             argslist.append(taskargs)
         
         self.argslist = argslist
@@ -575,7 +578,7 @@ class Parallel:
 
 def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncpus=None, 
                 maxcpu=None, maxmem=None, interval=None, parallelizer=None, serial=False, 
-                progress=False, callback=None, die=True, **func_kwargs):
+                progress=False, callback=None, globaldict=None, die=True, **func_kwargs):
     '''
     Execute a function in parallel.
 
@@ -610,6 +613,7 @@ def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncp
         serial       (bool)      : whether to skip parallelization and run in serial (useful for debugging; equivalent to ``parallelizer='serial'``)
         progress     (bool)      : whether to show a progress bar
         callback     (func)      : an optional function to call from each worker
+        globaldict   (dict)      : an optional global dictionary to pass to each worker (note: may not update properly with low task latency)
         die          (bool)      : whether to stop immediately if an exception is encountered (otherwise, store the exception as the result)
         func_kwargs  (dict)      : merged with kwargs (see above)
 
@@ -725,12 +729,13 @@ def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncp
     | *New in version 2.0.0:* changed default parallelizer from ``multiprocess.Pool`` to ``concurrent.futures.ProcessPoolExecutor``; replaced ``maxload`` with ``maxcpu``/``maxmem``; added ``returnpool`` argument
     | *New in version 2.0.4:* added "die" argument; changed exception handling
     | *New in version 3.0.0:* new Parallel class; propagated "die" to jobs
+    | *New in version 3.0.1:* new "globaldict" argument
     '''
     # Create the parallel instance
     P = Parallel(func, iterarg=iterarg, iterkwargs=iterkwargs, args=args, kwargs=kwargs, 
[…]
