### 2023-05-16 1f53721
FIX: do not ignore user input in dataframe init
If the user passes both cols and columns to dataframe init columns was silently ignored.
