### 2023-05-16 7c09946
FIX: correct invocation of deps installation in tests README.rst
