+++ tests/README.rst
@@ -7,7 +7,7 @@ This folder contains the core tests for Sciris.
 Installation
 ------------
 
-To install test dependencies, use ``pip install -r requirements_test.txt``. (Note: ``pytest-parallel`` is used instead of ``pytest-xdist`` since it has much less overhead for running fast tests.)
+To install test dependencies, use ``pip install -r requirements.txt``. (Note: ``pytest-parallel`` is used instead of ``pytest-xdist`` since it has much less overhead for running fast tests.)
 
 Usage
 -----
