### 2023-05-17 #5076 @tacaswell
> Hi @danielskatz -- I think I've addressed all the review comments. I've found the "branch of main software repo" structure challenging to work with, so have been working on the paper in its own repo:
https://github.com/sciris/sciris_paper

https://github.com/openjournals/joss-reviews/issues/5076#issuecomment-1499643788
