### 2023-05-17 #5076 @danielskatz
Ah, I see.

👋 @cliffckerr, this isn't something JOSS is really set up to handle. Can you put the updated paper in the joss branch, which we can handle, or in the main repo, which would also be fine. Having it be in a separate repo doesn't work for our system
