### 2023-05-20 ef6f276
Adapt to more traditional description, and parallelise sampling
