+++ sciris/sc_alias.py
@@ -19,19 +19,20 @@ D. Knuth. The Art of Computer Programming, Vol 2: Seminumerical Algorithms, sect
 import numpy.random as npr
 import numpy as np
 from . import sc_utils as scu
+from . import sc_math as scm
+from . import sc_parallel as scp
 
 __all__ = ['alias_sampler']
 
 
-def sample(n, q, J, r1, r2):
-    res = np.zeros(n, dtype=np.int32)
-    lj = len(J)
-    for i in range(n):
-        kk = int(np.floor(r1[i] * lj))
-        if r2[i] < q[kk]:
-            res[i] = kk
-        else:
-            res[i] = J[kk]
+def sample(r1, r2, probs_table=None, alias_table=None):
+    res = scu.toarray(0, dtype=np.int32)
+    lj = len(alias_table)
+    kk = int(np.floor(r1 * lj))
+    if r2 < probs_table[kk]:
+        res[0] = kk
+    else:
+        res[0] = alias_table[kk]
     return res
 
 
@@ -54,13 +55,15 @@ class alias_sampler:
 
     """
 
-    def __init__(self, probs, vals=None, randseed=None, verbose=False):
+    def __init__(self, probs, vals=None, randseed=None, verbose=False, parallel=True):
         self.probs = scu.toarray(probs)
         self.vals = scu.toarray(vals)  # If vals is None, then self.vals is an empty array
         self.rng = np.random.default_rng(randseed)  # Construct a new Generator
-        self.K = K = len(probs)
-        self.q = q = np.zeros(K)
-        self.J = J = np.zeros(K, dtype=np.int32)
+        self.n = n = len(probs)
+        # Initialise tables
+        self.probs_table = probs_table = n * self.probs
+        self.alias_table = alias_table = np.empty((n,), dtype=np.int32)
+        self.parallel = parallel
 
         # Do some checks if both probs and vals are given
         if vals is not None:
@@ -72,22 +75,49 @@ class alias_sampler:
         # Run checks on probs
         self._check_probs(verbose)
 
-        smaller, larger = [], []
-        for kk, prob in enumerate(probs):
-            q[kk] = K * prob
-            if q[kk] < 1.0:  # underfull group
-                smaller.append(kk)
-            else:
-                larger.append(kk)
+        # Divide the table entries into three categories,
+        overfull = probs_table > 1.0
+        underfull = probs_table < 1.0
+        exactly_full = probs_table == 1.0
 
-        while len(smaller) > 0 and len(larger) > 0:
-            small, large = smaller.pop(), larger.pop()
-            J[small] = large
-            q[large] = q[large] - (1.0 - q[small])
-            if q[large] < 1.0:
-                smaller.append(large)
+        # Alias table K, initialise exactly full entries
+        alias_table[exactly_full] = scm.findinds(exactly_full)
+
+        # As long as not all prob table entries are exactly full, repeat the following steps:
+        while any(underfull) and any(overfull):
+            # Find True indices
+            uf = scm.findinds(underfull)
+            of = scm.findinds(overfull)
+
+            # Arbitrarily choose an overfull entry (i) and an underfull entry (j)
+            uf_j = uf(self.rng.choice(uf, 1))
+            of_i = of(self.rng.choice(of, 1))
+
+            # Allocate the unused space in entry j to outcome i, by setting Kj = i.
+            alias_table[uf_j] = of_i
+
+            # Remove the allocated space from entry i by changing
+            probs_table[of_i] = probs_table[of_i] - (1.0 - probs_table[uf_j])
+
+            # Entry j is now exactly full, so update the corresponding category arrays
+            underfull[uf_j] = False
+            exactly_full[uf_j] = True
+
[…]
