### 2023-05-20 e81a46b
WIP: split sample functions, one to be used by parallelise and a vector based
