### 2023-05-21 1dceb6a
Remove redundant operation in vectorised sampling
