### 2023-05-21 9efdbbd
Return values if values were passed
