### 2023-05-21 caee472
Write basic tests for new alias sampler
