### 2023-05-21 a215d9c
Verify generated samples follow the desired distribution
