+++ tests/test_alias.py
@@ -5,6 +5,8 @@ Test Sciris alias_sampler class
 import sciris as sc
 import pytest
 import sciris as sc
+import numpy as np
+import scipy.stats as stats
 import numpy.random as npr
 
 
@@ -12,7 +14,7 @@ import numpy.random as npr
 def alias_sampler_fixture():
     # Fixture to create an instance of alias_sampler for testing
     probs = [0.1, 0.1, 0.05, 0.25, 0.1, 0.25, 0.15]
-    return sc.alias_sampler(probs)
+    return sc.alias_sampler(probs, randseed=42)
 
 
 def test_draw_indices(alias_sampler_fixture):
@@ -57,11 +59,39 @@ def test_probs_not_add_up_to_one():
     assert pytest.approx(sum(sampler.probs)) == 1.0
 
 
+def test_sample_distribution():
+    """
+    Verify that the generated samples from the draw method of the alias_sampler
+    class follow the desired distribution using a statistical test:
+    """
+    # Define a normal probability distribution -- will have to be normalised to add up to 1
+    probs = np.array([0.0021, 0.0047, 0.0097, 0.0181, 0.0309, 0.0483, 0.069, 0.0902,
+                      0.1078, 0.1179, 0.1179, 0.1078, 0.0902, 0.069, 0.0483, 0.0309,
+                      0.0181, 0.0097, 0.0047, 0.0021])
+
+    vals = ['lol', 'btw', 'omg', 'idk', 'wtf', 'btw', 'npo', 'omw', 'brb', 'gtg',
+            'tmi', 'fyi', 'imo', 'atm', 'bff', 'tbh', 'irl', 'smh', 'tyt', 'lmk']
+
+    # Create an instance of alias_sampler
+    sampler = sc.alias_sampler(probs, vals)
+
+    # Generate a larg-ish number of samples
+    n_samples = 2**16
+    samples = sampler.draw(n_samples)
+
+    # Perform a chi-square goodness-of-fit test
+    observed, _ = np.histogram(samples, bins=len(vals))
+    expected = n_samples * probs
+    _, p_value = stats.chisquare(observed, expected)
+
+    # Verify that the p-value is above a certain threshold (e.g., 0.05)
+    assert p_value > 0.05
+
 # %% Run as a script
 if __name__ == '__main__':
     sc.tic()
 
-    # Add something here
+    test_sample_distribution()
 
     sc.toc()
     print('Done.')
