### 2023-05-21 d70bbd8
Replace categorical values by integer
