### 2023-05-21 690a7ba
WIP: Set seed that causes index error
