### 2023-05-21 5bcc716
Debug: initialise tables after checking input probs
