### 2023-05-21 a0cdc85
Actually raise ValueError if there's a negative prob
