### 2023-05-21 323718c
Use sampler probabilities to ensure that the observed and expected sums are the same
