+++ sciris/sc_dataframe.py
@@ -69,7 +69,7 @@ class dataframe(pd.DataFrame):
         # Handle inputs
         if 'cols' in kwargs:
             if columns is not None:
-                raise ValueError('you supplied both *cols* and *columns*, you may only supply one of them')
+                raise ValueError('You can supply "cols" or "columns", but not both')
             columns = kwargs.pop('cols')
         if nrows and data is None:
             ncols = len(columns)
