+++ tests/README.rst
@@ -2,14 +2,14 @@
 Tests
 =====
 
-This folder contains the core tests for Sciris.
+This folder contains the tests for Sciris.
 
 Installation
 ------------
 
-To install test dependencies, use ``pip install -r requirements.txt``. (Note: ``pytest-parallel`` is used instead of ``pytest-xdist`` since it has much less overhead for running fast tests.)
+To install test dependencies, use ``pip install -r requirements.txt``.
 
 Usage
 -----
 
-Recommended usage is ``./check_coverage`` or ``./run_tests``. You can also use ``pytest`` to run all the tests in the folder.
+Recommended usage is ``./check_coverage`` or ``./run_tests``. You can also use ``pytest`` to run all the tests in the folder. To run tests in parallel using ``pytest-parallel`` (which is faster but less robust than the default ``pytest-xdist``), use ``./run_tests fast``.
