### 2023-05-23 337a4e9
Refactoring loading approach
