+++ sciris/sc_fileio.py
@@ -168,7 +168,7 @@ def load(filename=None, folder=None, verbose=False, die=None, remapping=None,
 
     # Unpickle it
     try:
-        kw = dict(filename=filename, verbose=verbose, die=die, remapping=remapping, method=method, auto_remap=auto_remap)
+        kw = dict(verbose=verbose, die=die, remapping=remapping, method=method, auto_remap=auto_remap)
         obj = _unpickler(filestr, **kw, **kwargs) # Unpickle the data
     except Exception as E: # pragma: no cover
         exc = type(E) # Figure out what kind of error it is
@@ -2129,7 +2129,7 @@ class UniversalFailed(Failed): # pragma: no cover
         return
 
     def __len__(self):
-        return len(self.dict)
+        return len(self.dict) if hasattr(self, 'dict') else 0
 
     def __getitem__(self, key):
         return self.dict[key]
@@ -2172,7 +2172,14 @@ def _remap_module(remapping, module_name, name):
     return obj
 
 
-class _RobustUnpickler(dill.Unpickler):
+class _LoadsInterface():
+    # Add a .loads method to unpicklers, with support for Sciris remapping
+    @classmethod
+    def loads(cls, string, remapping, **kwargs):
+        unpickler = cls(io.BytesIO(string), remapping=remapping, **kwargs)
+        return unpickler.load()
+
+class _RobustUnpickler(dill.Unpickler, _LoadsInterface):
     ''' Try to import an object, and if that fails, return a Failed object rather than crashing '''
 
     def __init__(self, bytesio, fix_imports=True, encoding="latin1", errors="ignore", remapping=None):
@@ -2185,7 +2192,34 @@ class _RobustUnpickler(dill.Unpickler):
         return obj
 
 
-class _UltraRobustUnpickler(dill.Unpickler): # pragma: no cover
+class _RenamingUnpickler(_RobustUnpickler, _LoadsInterface):
+    ''' Like RobustUnpickler but with automatic population of 'remapping' with known fixes '''
+
+    known_fixes = {
+        ('pandas.core.indexes.numeric', 'Int64Index'): {'pandas.core.indexes.numeric.Int64Index':'pandas.core.indexes.api.Index'},
+        ('pandas.core.indexes.numeric', 'Float64Index'): {'pandas.core.indexes.numeric.Float64Index':'pandas.core.indexes.api.Index'},
+    }
+
+    def find_class(self, module_name, name, last_error = None, verbose=True):
+
+        try:
+            return super().find_class(module_name, name, verbose)
+        except ModuleNotFoundError as E:
+            if last_error == E:
+                raise E # We already tried to fix the error and it didn't work
+
+            if (module_name, name) in self.known_fixes:
+                self.remapping.update(self.known_fixes[(module_name, name)])
+            elif module_name in self.known_fixes:
+                self.remapping.update(self.known_fixes[module_name])
+            else:
+                raise E
+
+            warnmsg = f'Fixing known unpickling deprecation "{str(E)}"'
+            warnings.warn(warnmsg, category=UserWarning, stacklevel=2)
+            return self.find_class(module_name, name, E, verbose)
+
+class _UltraRobustUnpickler(dill.Unpickler, _LoadsInterface): # pragma: no cover
     ''' If all else fails, just make a default object '''
 
     def __init__(self, bytesio, *args, fix_imports=True, encoding="latin1", errors="ignore", 
@@ -2199,16 +2233,14 @@ class _UltraRobustUnpickler(dill.Unpickler): # pragma: no cover
         ''' Ignore all attempts to use the actual class and always make a UniversalFailed class '''
         try:
             obj = _remap_module(self.remapping, module_name, name)
-        except:
[…]
