### 2023-05-27 7cf118d
Replace draw_n with draw so both classes have the same itnerface for benchmarking
