### 2023-05-27 b6d00b2
Add script to benchmark alias methods
