### 2023-06-26 #5076 @danielskatz
@cliffckerr - just checking on your progress as we start the new week...
