+++ paper.bib
@@ -321,3 +321,13 @@
   year={2018},
   publisher={Public Library of Science San Francisco, CA USA}
 }
+
+
+@InProceedings{rocklin2015dask,
+  author    = { Matthew Rocklin },
+  title     = { Dask: Parallel Computation with Blocked algorithms and Task Scheduling },
+  booktitle = { Proceedings of the 14th Python in Science Conference },
+  pages     = { 130 - 136 },
+  year      = { 2015 },
+  editor    = { Kathryn Huff and James Bergstra }
+}
