+++ paper.md
@@ -5,8 +5,8 @@ tags:
   - python
   - scientific software development
   - computational science
-  - high-performance containers
   - numerical utilities
+  - containers
   - plotting
 
 authors:
@@ -63,44 +63,44 @@ affiliations:
  - name: Google, Zürich, Switzerland
    index: 8
 
-date: 5 December 2022
+date: 30 March 2023
 bibliography: paper.bib
 
 ---
 
 # Summary 
 
-[Sciris](http://sciris.org) aims to streamline the development of scientific software by making it easier to perform common tasks. Sciris provides classes and functions that simplify access to frequently used low-level functionality in the core libraries of the scientific Python ecosystem (such as `numpy` for math and `matplotlib` for plotting), as well as in libraries of broader scope (such as `multiprocess` for parallelization and `pickle` for saving and loading objects). While low-level functionality is valuable for developing robust software applications, it can divert focus from the scientific problems being solved. Some of Sciris' key features include: ensuring consistent dictionary, list, and array types (e.g., enabling users to provide inputs as either lists or arrays); enabling ordered dictionary elements to be referenced by index; simplifying datetime arithmetic by allowing date input in multiple formats, including strings; simplifying the saving and loading of files and complex objects; and simplifying the parallel execution of code. Sciris makes writing scientific code in Python faster, more pleasant, and more accessible, especially for people without extensive training in software development. With Sciris, users can achieve the same functionality with fewer lines of code, avoid reinventing the wheel, and spend less time looking up recipes on Stack Overflow. Sciris also forms the basis of ScirisWeb, an additional set of tools for building Flask-based Python webapps.
+[Sciris](http://sciris.org) aims to streamline the development of scientific software by making it easier to perform common tasks. Sciris provides classes and functions that simplify access to frequently used low-level functionality in the core libraries of the scientific Python ecosystem (such as `numpy` for math and `matplotlib` for plotting), as well as in libraries of broader scope (such as `multiprocess` for parallelization and `pickle` for saving and loading objects). While low-level functionality is valuable for developing robust software applications, it can divert focus from the scientific problems being solved. Some of Sciris' key features include: ensuring consistent dictionary, list, and array types (e.g., enabling users to provide inputs as either lists or arrays); enabling ordered dictionary elements to be referenced by index; simplifying datetime arithmetic by allowing date input in multiple formats, including strings; simplifying the saving and loading of files and complex objects; and simplifying the parallel execution of code. With Sciris, users can often achieve the same functionality with fewer lines of code, avoid reinventing the wheel, and spend less time looking up recipes on Stack Overflow. This can make writing scientific code in Python faster, more pleasant, and more accessible, especially for people without extensive training in software development.
 
 
 # Statement of need
 
 ## The landscape of scientific software
-With the increasing availability of large volumes of data and computing resources, scientists across multiple fields of research have been able to tackle increasingly complex problems. But to harness these resources, the need for domain-specific software has become much greater. As the complexity of the questions being tackled has increased, so too has the software used to answer them, creating a steep learning curve and an increasing burden of code review [@burden-codereview].
[…]
