### 2023-06-28 #5076 @aflaxman
Confirmed!  Ready to publish.

From: Daniel S. Katz ***@***.***>
Sent: Wednesday, June 28, 2023 2:21 AM
Cc: Abraham Flaxman ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: Sciris: Simplifying scientific software in Python (Issue #5076)

👋 @tacaswell<https://urldefense.com/v3/__https:/github.com/tacaswell__;!!K-Hz7m0Vt54!jIpNmxX61Q68wFZqC-F_mGXCZgZmNqE6lN8yccwO4HZ5IkfmmDTWGvezMuEr06wgJmG1gXaXJI9VUc6dyljp$> & @aflaxman<https://urldefense.com/v3/__https:/github.com/aflaxman__;!!K-Hz7m0Vt54!jIpNmxX61Q68wFZqC-F_mGXCZgZmNqE6lN8yccwO4HZ5IkfmmDTWGvezMuEr06wgJmG1gXaXJI9VURKoWU4g$> - you both have checked off all items in your lists, so I assume you are ready for this to be accepted and published, but I would appreciate if you could confirm this here before I proceed.
