### 2023-07-26 b4e4f5c
update test parallel docstring
