### 2023-07-26 da4aae8
update globaldict
