### 2023-07-26 ea652ec
about to remove manager
