### 2023-07-26 fbd3c07
working better but too complicated
