### 2023-07-27 5bb960b
add parallel to benchmark
