### 2023-07-27 cc97111
update settings
