+++ sciris/sc_datetime.py
@@ -124,7 +124,8 @@ def getdate(obj=None, astype='str', dateformat=None):
 def readdate(datestr=None, *args, dateformat=None, return_defaults=False, verbose=False):
     '''
     Convenience function for loading a date from a string. If dateformat is None,
-    this function tries a list of standard date types.
+    this function tries a list of standard date types. Note: in most cases :func:`sc.date() <date>`
+    should be used instead.
 
     By default, a numeric date is treated as a POSIX (Unix) timestamp. This can be changed
     with the ``dateformat`` argument, specifically:
@@ -249,13 +250,14 @@ def date(obj=None, *args, start_date=None, readformat=None, to='date', as_date=N
     '''
     Convert any reasonable object -- a string, integer, or datetime object, or
     list/array of any of those -- to a date object (or string, pandas, or numpy
-    date). To convert an integer to a date, you must supply a start date.
-
-    Caution: while this function and :func:`sc.readdate() <readdate>` are similar, and indeed this function
-    calls :func:`sc.readdate() <readdate>` if the input is a string, in this function an integer is treated
-    as a number of days from start_date, while for :func:`sc.readdate() <readdate>` it is treated as a
-    timestamp in seconds.
-
+    date). 
+    
+    If the object is an integer, this is interpreted as follows:
+    
+    - With readformat='posix': treat as a POSIX timestamp, in seconds from 1970
+    - With readformat='ordinal'/'matplotlib': treat as an ordinal number of days from 1970 (Matplotlib default)
+    - With start_date provided: treat as a number of days from this date
+    
     Note: in this and other date functions, arguments work either with or without
     underscores (e.g. ``start_date`` or ``startdate``)
 
@@ -264,7 +266,7 @@ def date(obj=None, *args, start_date=None, readformat=None, to='date', as_date=N
         args (str/int/date/datetime): additional objects to convert
         start_date (str/date/datetime): the starting date, if an integer is supplied
         readformat (str/list): the format to read the date in; passed to :func:`sc.readdate() <readdate>` (NB: can also use "format" instead of "readformat")
-        to (str): the output format: 'date' (default), 'str' (or 'string'), 'pandas', or 'numpy'
+        to (str): the output format: 'date' (default), 'datetime', 'str' (or 'string'), 'pandas', or 'numpy'
         as_date (bool): alternate method of choosing between  output format of 'date' (True) or 'str' (False); if None, use "to" instead
         outformat (str): the format to output the date in, if returning a string
         kwargs (dict): only used for deprecated argument aliases
@@ -282,6 +284,7 @@ def date(obj=None, *args, start_date=None, readformat=None, to='date', as_date=N
     | *New in version 1.2.2:* "readformat" argument; renamed "dateformat" to "outformat"
     | *New in version 2.0.0:* support for :obj:`np.datetime64 <numpy.datetime64>` objects
     | *New in version 3.0.0:* added "to" argument, and support for :obj:`pd.Timestamp <pandas.Timestamp>` and :obj:`np.datetime64 <numpy.datetime64>` output; allow None
+    | *New in version 3.0.1:* allow "datetime" output
     '''
 
     # Handle deprecation
@@ -295,10 +298,17 @@ def date(obj=None, *args, start_date=None, readformat=None, to='date', as_date=N
         warnings.warn(warnmsg, category=FutureWarning, stacklevel=2)
     if as_date is not None: # pragma: no cover
         to = 'date' if as_date else 'str' # Legacy support for as_date boolean
+    
+    def dateify(obj):
+        ''' Handle dates vs datetimes '''
+        if to == 'date' and hasattr(obj, 'date'):
+            return obj.date()
+        else:
+            return obj
 
     # Convert to list and handle other inputs
[…]
