### 2023-08-02 #5076 @cliffckerr
@danielskatz Working on it! Halfway through, hope the second half goes faster :)
https://github.com/sciris/sciris/milestone/21
