+++ sciris/sc_plotting.py
@@ -60,7 +60,7 @@ def ax3d(nrows=None, ncols=None, index=None, fig=None, ax=None, returnfig=False,
     '''
     Create a 3D axis to plot in.
 
-    Usually not invoked directly; kwags are passed to add_subplot()
+    Usually not invoked directly; kwargs are passed to ``fig.add_subplot()``
     
     *New in version 3.0.0:* nrows, ncols, and index arguments first
     '''
@@ -98,14 +98,15 @@ def ax3d(nrows=None, ncols=None, index=None, fig=None, ax=None, returnfig=False,
     if ax is None:
         if fig.axes and index is None:
             ax = pl.gca()
-            if not isinstance(ax, Axes3D):
-                errormsg = f'''Cannot create 3D plot into axes {ax}: ensure "projection='3d'" was used when making it'''
-                raise ValueError(errormsg)
         else:
             if nrows is None: nrows = 1
             if ncols is None: ncols = 1
             if index is None: index = 1
             ax = fig.add_subplot(nrows, ncols, index, projection='3d', **axkwargs)
+    
+    if not isinstance(ax, Axes3D):
+        errormsg = f'''Cannot create 3D plot into axes {ax}: ensure "projection='3d'" was used when making it'''
+        raise ValueError(errormsg)
     
     if elev is not None or azim is not None: # pragma: no cover
         ax.view_init(elev=elev, azim=azim)
@@ -117,7 +118,8 @@ def ax3d(nrows=None, ncols=None, index=None, fig=None, ax=None, returnfig=False,
         return ax
 
 
-def plot3d(x, y, z, c=None, fig=None, ax=None, returnfig=False, figkwargs=None, axkwargs=None, plotkwargs=None, **kwargs):
+def plot3d(x, y, z, c='index', fig=None, ax=None, returnfig=False, figkwargs=None, 
+           axkwargs=None, plotkwargs=None, **kwargs):
     '''
     Plot 3D data as a line
 
@@ -125,28 +127,52 @@ def plot3d(x, y, z, c=None, fig=None, ax=None, returnfig=False, figkwargs=None, 
         x (arr): x coordinate data
         y (arr): y coordinate data
         z (arr): z coordinate data
-        c (str/tuple): color, can be any of the types accepted by matplotlib's plot()
+        c (str/tuple): color, can be an array or any of the types accepted by :func:`pl.plot() <matplotlib.pyplot.plot>`; if "index", color by index
         fig (fig): an existing figure to draw the plot in (or set to True to create a new figure)
         ax (axes): an existing axes to draw the plot in
         returnfig (bool): whether to return the figure, or just the axes
-        figkwargs (dict): passed to figure()
-        axkwargs (dict): passed to axes()
-        plotkwargs (dict): passed to plot()
-        kwargs (dict): also passed to plot()
+        figkwargs (dict): :func:`pl.figure() <matplotlib.pyplot.figure>`
+        axkwargs (dict): :func:`pl.axes() <matplotlib.pyplot.axes>`
+        plotkwargs (dict): passed to :func:`pl.plot() <matplotlib.pyplot.plot>`
+        kwargs (dict): also passed to :func:`pl.plot() <matplotlib.pyplot.plot>`
 
-    **Example**::
+    **Examples**::
 
         x,y,z = pl.rand(3,10)
         sc.plot3d(x, y, z)
+        
+        fig = pl.figure()
+        n = 100
+        x = np.array(sorted(np.random.rand(n)))
[…]
