+++ sciris/sc_plotting.py
@@ -55,19 +55,32 @@ def fig3d(num=None, nrows=1, ncols=1, index=1, returnax=False, figkwargs=None, a
         return fig
 
 
-def ax3d(nrows=None, ncols=None, index=None, fig=None, ax=None, returnfig=False, silent=False, 
-         elev=None, azim=None, figkwargs=None, axkwargs=None, **kwargs):
+def ax3d(nrows=None, ncols=None, index=None, fig=None, ax=None, returnfig=False,
+         elev=None, azim=None, figkwargs=None, **kwargs):
     '''
     Create a 3D axis to plot in.
 
     Usually not invoked directly; kwargs are passed to ``fig.add_subplot()``
     
-    *New in version 3.0.0:* nrows, ncols, and index arguments first
+    Args:
+        nrows (int): number of rows of axes in plot
+        ncols (int): number of columns of axes in plot
+        index (int): index of current plot
+        fig (Figure): if provided, use existing figure
+        ax (Axes): if provided, validate and use these axes
+        returnfig (bool): whether to return the figure (else just the axes)
+        elev (float): the elevation of the 3D viewpoint
+        azim (float): the azimuth of the 3D viewpoint
+        figkwargs (dict): passed to :func:`pl.figure() <matplotlib.pyplot.figure>`
+        kwargs (dict): passed to :func:`pl.axes() <matplotlib.pyplot.axes>`
+        
+    | *New in version 3.0.0:* nrows, ncols, and index arguments first
+    | *New in version 3.0.1:* improved validation; 'silent' and 'axkwargs' argument removed
     '''
-    from mpl_toolkits.mplot3d import Axes3D # analysis:ignore
+    from mpl_toolkits.mplot3d import Axes3D
 
     figkwargs = scu.mergedicts(figkwargs)
-    axkwargs = scu.mergedicts(axkwargs, kwargs, nrows=nrows, ncols=ncols, index=index)
+    axkwargs = scu.mergedicts(dict(nrows=nrows, ncols=ncols, index=index), kwargs)
     nrows = axkwargs.pop('nrows', nrows) # Since fig.add_subplot() can't handle kwargs...
     ncols = axkwargs.pop('ncols', ncols)
     index = axkwargs.pop('index', index)
@@ -80,7 +93,7 @@ def ax3d(nrows=None, ncols=None, index=None, fig=None, ax=None, returnfig=False,
         pass # This is fine, just a different format
 
     # Handle the figure
-    if fig in [True, False] or (fig is None and figkwargs): # Any of these things indicate that we want a new figure
+    if fig in [True, False] or (fig is None and figkwargs): # Confusingly, any of these things indicate that we want a new figure
         fig = pl.figure(**figkwargs)
     elif fig is None:
         if ax is None:
@@ -90,9 +103,6 @@ def ax3d(nrows=None, ncols=None, index=None, fig=None, ax=None, returnfig=False,
                 fig = pl.gcf()
         else: # pragma: no cover
             fig = ax.figure
-            silent = False
-    else:
-        silent = False # Never close an already open figure
 
     # Create and initialize the axis
     if ax is None:
@@ -104,14 +114,16 @@ def ax3d(nrows=None, ncols=None, index=None, fig=None, ax=None, returnfig=False,
             if index is None: index = 1
             ax = fig.add_subplot(nrows, ncols, index, projection='3d', **axkwargs)
     
+    # Validate the axes
     if not isinstance(ax, Axes3D):
         errormsg = f'''Cannot create 3D plot into axes {ax}: ensure "projection='3d'" was used when making it'''
[…]
