### 2023-08-05 f6704f8
working on getting colors for bar3d
