+++ sciris/sc_plotting.py
@@ -130,6 +130,65 @@ def ax3d(nrows=None, ncols=None, index=None, fig=None, ax=None, returnfig=False,
         return ax
 
 
+
+def _process_2d_data(x, y, z, c, flatten=False):
+    ''' Helper function to handle data transformations -- not for the user '''
+
+    # Swap variables so z always exists
+    if z is None and x is not None:
+        z,x = x,z
+    z = np.array(z)
+    
+    if z.ndim == 2:
+        ny,nx = z.shape
+        x = np.arange(nx) if x is None else np.array(x)
+        y = np.arange(ny) if y is None else np.array(y)
+        assert x.ndim == y.ndim, 'Cannot handle x and y axes with different array shapes'
+        if x.ndim == 1 or y.ndim == 1:
+            x,y = np.meshgrid(x, y)
+        if flatten:
+            x,y,z = x.flatten(), y.flatten(), z.flatten() # Flatten everything to 1D
+    elif flatten == False:
+        raise ValueError('Must provide z values as a 2D array')
+    
+    if x is None or y is None:
+        raise ValueError('Must provide x and y values if z is 1D')
+        
+    print('hi2', z.shape)
+        
+    # Handle automatic color scaling
+    if isinstance(c, str):
+        if c == 'z':
+            c = z
+        elif c == 'index':
+            c = np.arange(len(z))
+            
+    print('hi3', c.shape)
+    
+    return x, y, z, c
+
+
+def _process_colors(c, z=None, cmap=None, to2d=False):
+    ''' Helper function to get color data in the right format -- not for the user '''
+    
+    from . import sc_colors as scc # To avoid circular import
+    
+    # Handle colors
+    if c.ndim == 1: # Used by scatter3d and bar3d
+        assert len(c) == len(z), 'Number of colors does not match length of data'
+        c = scc.vectocolor(c, cmap=cmap)
+    elif c.ndim == 2: # Used by surf3d
+        assert c.shape == z.shape, 'Shape of colors does not match shape of data'
+        c = scc.arraycolors(c, cmap=cmap)
+    
+    # Used by bar3d -- flatten from 3D to 2D
+    if to2d and c.ndim == 3:
+        c = c.reshape((-1, c.shape[2]))
+    
+    return c
+    
+
+
 def plot3d(x, y, z, c='index', fig=None, ax=None, returnfig=False, figkwargs=None, 
            axkwargs=None, **kwargs):
     '''
@@ -174,12 +233,12 @@ def plot3d(x, y, z, c='index', fig=None, ax=None, returnfig=False, figkwargs=Non
     fig,ax = ax3d(returnfig=True, fig=fig, ax=ax, figkwargs=figkwargs, **axkwargs)
     
     # Handle different-colored line segments
+    
     if c == 'index':
         c = np.arange(n) # Assign automatically based on index
     if scu.isarray(c) and len(c) in [n, n-1]: # Technically don't use the last color
         if c.ndim == 1:
-            from . import sc_colors as scc # To avoid circular import
-            c = scc.vectocolor(c)
+            c = _process_colors(c, z=z)
         for i in range(n-1):
             ax.plot(x[i:i+2], y[i:i+2], z[i:i+2], c=c[i], **plotkwargs)
             
@@ -191,39 +250,6 @@ def plot3d(x, y, z, c='index', fig=None, ax=None, returnfig=False, figkwargs=Non
         return fig,ax
     else:
         return ax
-
-
-def _process_2d_data(x, y, z, c, flatten=False):
-    ''' Used internally to handle data transformations '''
-
-    # Swap variables so z always exists
-    if z is None and x is not None:
-        z,x = x,z
-    z = np.array(z)
[…]
