+++ sciris/sc_colors.py
@@ -15,7 +15,7 @@ Highlights:
 import struct
 import pylab as pl
 import numpy as np
-from matplotlib import colors as mplc, colormaps as mplcm
+import matplotlib as mpl
 from . import sc_utils as scu
 from . import sc_math as scm
 
@@ -62,7 +62,7 @@ def sanitizecolor(color, asarray=False, alpha=None, normalize=True):
     '''
     if isinstance(color, str):
         try:
-            color = mplc.to_rgb(color)
+            color = mpl.colors.to_rgb(color)
         except ValueError as E:
             errormsg = f'Could not understand "{color}" as a valid color: must be a standard Matplotlib color string'
             raise ValueError(errormsg) from E
@@ -113,9 +113,9 @@ def shifthue(colors=None, hueshift=0.0):
     '''
     colors, origndim = _listify_colors(colors)
     for c,color in enumerate(colors):
-        hsvcolor = mplc.rgb_to_hsv(color)
+        hsvcolor = mpl.colors.rgb_to_hsv(color)
         hsvcolor[0] = (hsvcolor[0]+hueshift) % 1.0 # Calculate new hue and return the modulus
-        rgbcolor = mplc.hsv_to_rgb(hsvcolor)
+        rgbcolor = mpl.colors.hsv_to_rgb(hsvcolor)
         colors[c] = rgbcolor
     colors = _listify_colors(colors, origndim)
     return colors
@@ -169,7 +169,7 @@ def rgb2hsv(colors=None):
     '''
     colors, origndim = _listify_colors(colors)
     for c,color in enumerate(colors):
-        hsvcolor = mplc.rgb_to_hsv(color)
+        hsvcolor = mpl.colors.rgb_to_hsv(color)
         colors[c] = hsvcolor
     colors = _listify_colors(colors, origndim)
     return colors
@@ -186,7 +186,7 @@ def hsv2rgb(colors=None):
     '''
     colors, origndim = _listify_colors(colors)
     for c,color in enumerate(colors):
-        hsvcolor = mplc.hsv_to_rgb(color)
+        hsvcolor = mpl.colors.hsv_to_rgb(color)
         colors[c] = hsvcolor
     colors = _listify_colors(colors, origndim)
     return colors
@@ -197,7 +197,7 @@ def hsv2rgb(colors=None):
 #%% Colormap-related functions
 ##############################################################################
 
-__all__ += ['vectocolor', 'arraycolors', 'gridcolors', 'midpointnorm', 'colormapdemo']
+__all__ += ['vectocolor', 'arraycolors', 'gridcolors', 'midpointnorm', 'manualcolorbar', 'colormapdemo']
 
 
 def vectocolor(vector, cmap=None, asarray=True, reverse=False, minval=None, maxval=None, midpoint=None, nancolor=None):
@@ -463,28 +463,143 @@ def midpointnorm(vcenter=0, vmin=None, vmax=None):
 
     *New in version 1.2.0.*
     '''
-    from matplotlib.colors import TwoSlopeNorm
-    norm = TwoSlopeNorm(vcenter=vcenter, vmin=vmin, vmax=vmax)
+    norm = mpl.colors.TwoSlopeNorm(vcenter=vcenter, vmin=vmin, vmax=vmax)
     return norm
 
 
 
-def manualcolorbar():
+def manualcolorbar(data=None, vmin=0, vmax=1, vcenter=None, colors=None, values=None,
+                   cmap=None, norm=None, label=None, labelkwargs=None, ticks=None, 
+                   ticklabels=None, fig=None, ax=None, cax=None, axkwargs=None, **kwargs):
     '''
-    Add a colorbar to plot that does not support one by default.
+    Add a colorbar to a plot that does not support one by default.
+    
+    There are three main use cases, from least to most manual:
+    
[…]
