+++ sciris/sc_printing.py
@@ -46,20 +46,23 @@ __all__ = ['createcollist', 'objectid', 'objatt', 'objmeth', 'objprop', 'objrepr
 
 def createcollist(items, title=None, strlen=18, ncol=3):
     ''' Creates a string for a nice columnated list (e.g. to use in __repr__ method) '''
-    nrow = int(np.ceil(float(len(items))/ncol))
-    newkeys = []
-    for x in range(nrow):
-        newkeys += items[x::nrow]
-
-    attstring = title + ':' if title else ''
-    c = 0
-    for x in newkeys:
-        if c%ncol == 0: attstring += '\n  '
-        if len(x) > strlen: x = x[:strlen-3] + '...'
-        attstring += '%-*s  ' % (strlen,x)
-        c += 1
-    attstring += '\n'
-    return attstring
+    if len(items):
+        nrow = int(np.ceil(float(len(items))/ncol))
+        newkeys = []
+        for x in range(nrow):
+            newkeys += items[x::nrow]
+    
+        string = title + ':' if title else ''
+        c = 0
+        for x in newkeys:
+            if c%ncol == 0: string += '\n  '
+            if len(x) > strlen: x = x[:strlen-3] + '...'
+            string += '%-*s  ' % (strlen,x)
+            c += 1
+        string += '\n'
+    else:
+        string = ''
+    return string
 
 
 def objectid(obj):
@@ -69,68 +72,104 @@ def objectid(obj):
     return output
 
 
-def objatt(obj, strlen=18, ncol=3):
+def _get_obj_keys(obj, private=False, sort=True, use_dir=False):
+    ''' Helper method to get the keys of an object '''
+    if use_dir:
+        keys = obj.__dir__() # This is the unsorted version of dir()
+    else:
+        if   hasattr(obj, '__dict__'):  keys = sorted(obj.__dict__.keys())
+        elif hasattr(obj, '__slots__'): keys = sorted(obj.__slots__)
+        else:                           keys = [] # pragma: no cover
+    if not private:
+        keys = [k for k in keys if not k.startswith('__')]
+    if sort:
+        keys = sorted(keys)
+    return keys
+
+
+def _is_meth(obj, attr):
+    ''' Helper function to check if an attribute is a method; do not distinguish between bound and unbound '''
+    return callable(getattr(obj, attr))
+
+
+def _is_prop(obj, attr):
+    ''' Helper function to check if an attribute is a property '''
+    return isinstance(getattr(type(obj), attr, None), property)
+
+
+def objatt(obj, strlen=18, ncol=3, private=False, sort=True, _keys=None):
     ''' Return a sorted string of object attributes for the Python __repr__ method '''
-    if   hasattr(obj, '__dict__'):  oldkeys = sorted(obj.__dict__.keys())
-    elif hasattr(obj, '__slots__'): oldkeys = sorted(obj.__slots__)
-    else:                           oldkeys = [] # pragma: no cover
-    if len(oldkeys): output = createcollist(oldkeys, 'Attributes', strlen = 18, ncol = 3)
-    else:            output = '' # pragma: no cover
+    keys = _get_obj_keys(obj, private=private, sort=sort) if _keys is None else _keys
+    output = createcollist(keys, 'Attributes', strlen=strlen, ncol=ncol)
     return output
 
 
-def objmeth(obj, strlen=18, ncol=3):
+def classatt(obj, strlen=18, ncol=3, private=False, sort=True, _objkeys=None, _dirkeys=None):
+    ''' Return a sorted string of class attributes for the Python __repr__ method '''
[…]
