+++ sciris/sc_printing.py
@@ -16,6 +16,7 @@ import sys
 import time
 import tqdm
 import pprint
+import warnings
 import numpy as np
 import collections as co
 from textwrap import fill
@@ -65,10 +66,16 @@ def createcollist(items, title=None, strlen=18, ncol=3):
     return string
 
 
-def objectid(obj):
-    ''' Return the object ID as per the default Python __repr__ method '''
+def objectid(obj, showclasses=False):
+    '''
+    Return the object ID as per the default Python ``__repr__`` method
+    
+    *New in version 3.0.1:* "showclasses" argument
+    '''
     c = obj.__class__
     output = f'<{c.__module__}.{c.__name__} at {hex(id(obj))}>\n'
+    if showclasses:
+        output += f'{c.mro()}\n'
     return output
 
 
@@ -77,8 +84,8 @@ def _get_obj_keys(obj, private=False, sort=True, use_dir=False):
     if use_dir:
         keys = obj.__dir__() # This is the unsorted version of dir()
     else:
-        if   hasattr(obj, '__dict__'):  keys = sorted(obj.__dict__.keys())
-        elif hasattr(obj, '__slots__'): keys = sorted(obj.__slots__)
+        if   hasattr(obj, '__dict__'):  keys = obj.__dict__.keys()
+        elif hasattr(obj, '__slots__'): keys = obj.__slots__
         else:                           keys = [] # pragma: no cover
     if not private:
         keys = [k for k in keys if not k.startswith('__')]
@@ -98,25 +105,28 @@ def _is_prop(obj, attr):
 
 
 def objatt(obj, strlen=18, ncol=3, private=False, sort=True, _keys=None):
-    ''' Return a sorted string of object attributes for the Python __repr__ method '''
+    ''' Return a sorted string of object attributes for the Python __repr__ method; see :func:`sc.prepr() <prepr>` for options '''
     keys = _get_obj_keys(obj, private=private, sort=sort) if _keys is None else _keys
     output = createcollist(keys, 'Attributes', strlen=strlen, ncol=ncol)
     return output
 
 
-def classatt(obj, strlen=18, ncol=3, private=False, sort=True, _objkeys=None, _dirkeys=None):
-    ''' Return a sorted string of class attributes for the Python __repr__ method '''
+def classatt(obj, strlen=18, ncol=3, private=False, sort=True, _objkeys=None, _dirkeys=None, return_keys=False):
+    ''' Return a sorted string of class attributes for the Python __repr__ method; see :func:`sc.prepr() <prepr>` for options '''
     objkeys = _get_obj_keys(obj, private=private, sort=sort, use_dir=False) if _objkeys is None else _objkeys
     dirkeys = _get_obj_keys(obj, private=private, sort=sort, use_dir=True)  if _dirkeys is None else _dirkeys
     keys = set(dirkeys) - set(objkeys) # Find attributes in dir() that are not in __dict__
     keys = filter(keys.__contains__, dirkeys) # Maintain original ordering
     keys = [k for k  in keys if not (_is_meth(obj, k) or _is_prop(obj, k))] # Exclude methods and properties; these are covered elsewhere
-    output = createcollist(keys, 'Class attributes', strlen=strlen, ncol=ncol)
-    return output
+    if return_keys:
+        return keys
+    else:
+        output = createcollist(keys, 'Class attributes', strlen=strlen, ncol=ncol)
+        return output
 
 
 def objmeth(obj, strlen=18, ncol=3, private=False, sort=True, _keys=None):
-    ''' Return a sorted string of object methods for the Python __repr__ method '''
+    ''' Return a sorted string of object methods for the Python __repr__ method; see :func:`sc.prepr() <prepr>` for options '''
[…]
