+++ sciris/sc_settings.py
@@ -24,21 +24,19 @@ from . import sc_odict as sco
 from . import sc_printing as scp
 
 # Matplotlib regression support
-try:
-    get_backend = pl.rcParams._get_backend_or_none
-except:
-    get_backend = pl.get_backend
+try:    get_backend = pl.rcParams._get_backend_or_none # Matplotlib >= 3.7
+except: get_backend = pl.get_backend # Matplotlib <= 3.6
 
 
-__all__ = ['ScirisOptions', 'options', 'parse_env', 'help']
+__all__ = ['style_simple', 'style_fancy', 'ScirisOptions', 'options', 'parse_env', 'help']
 
 #%% Define plotting options
 # See also simple.mplstyle and fancy.mplstyle files, which can be generated by:
-#   sc.saveyaml('simple.mplstyle', sc.sc_settings.rc_simple)
-#   sc.saveyaml('fancy.mplstyle',  sc.sc_settings.rc_fancy)
+#   sc.saveyaml('simple.mplstyle', sc.sc_settings.style_simple)
+#   sc.saveyaml('fancy.mplstyle',  sc.sc_settings.style_fancy)
 
 # Define simple plotting options -- similar to Matplotlib default
-rc_simple = {
+style_simple = {
     'axes.axisbelow':    True, # So grids show up behind
     'axes.spines.right': False,
     'axes.spines.top':   False,
@@ -48,8 +46,8 @@ rc_simple = {
 }
 
 # Define default plotting options -- loosely inspired by Seaborn
-rc_fancy = scu.dcp(rc_simple)
-rc_fancy.update({
+style_fancy = scu.dcp(style_simple)
+style_fancy.update({
     'axes.facecolor': '#f2f2ff',
     'axes.grid':      True,
     'grid.color':     'white',
@@ -506,9 +504,9 @@ class ScirisOptions(sco.objdict):
                 pl.style.use('default') # Need special handling here since not in pl.style.library...ugh
                 rc = {}
             elif stylestr in ['simple', 'sciris']:
-                rc = scu.dcp(rc_simple)
+                rc = scu.dcp(style_simple)
             elif stylestr in ['fancy', 'covasim']:
-                rc = scu.dcp(rc_fancy)
+                rc = scu.dcp(style_fancy)
             elif style in pl.style.library:
                 rc = scu.dcp(pl.style.library[style])
             else: # pragma: no cover
