+++ sciris/sc_fileio.py
@@ -78,7 +78,7 @@ https://stackoverflow.com/questions/41554738/how-to-load-an-old-pickle-file
 '''
 
 
-def _load_filestr(filename, folder, verbose=False):
+def _load_filestr(filename, folder=None, verbose=False):
     ''' Try different options for loading a file on disk into a string -- not for external use '''
     
     # Handle loading of either filename or file object
@@ -108,7 +108,7 @@ def _load_filestr(filename, folder, verbose=False):
         exc = type(E) # Figure out what kind of error it is
         if exc == FileNotFoundError: # This is simple, just raise directly
             raise E
-        elif exc == gz.BadGzipFile:
+        elif exc in gz.BadGzipFile:
             try: # If the gzip file failed, first try as a zstd compressed object
                 if verbose: print('  Reading as zstandard file...')
                 with open(filename, 'rb') as fh:
@@ -2070,28 +2070,20 @@ def savespreadsheet(filename=None, data=None, folder=None, sheetnames=None, clos
 #%% Pickling support methods
 ##############################################################################
 
-__all__ += ['Empty', 'Failed']
+__all__ += ['Failed']
+
+unpickling_errors = sco.objdict()
 
 # Pandas provides a class compatibility map, so use that by default
 try:
     known_fixes = pd.compat.pickle_compat._class_locations_map
 except (NameError or AttributeError):
-    warnmsg = 'Could not load full pandas compatibility dictionary; using manual subset'
+    warnmsg = 'Could not load full pandas pickle compatibility map; using manual subset of known regressions'
     warnings.warn(warnmsg, category=UserWarning, stacklevel=2)
     known_fixes = {
         ('pandas.core.indexes.numeric', 'Int64Index'): {'pandas.core.indexes.numeric.Int64Index':'pandas.core.indexes.api.Index'},
         ('pandas.core.indexes.numeric', 'Float64Index'): {'pandas.core.indexes.numeric.Float64Index':'pandas.core.indexes.api.Index'},
     }
-
-
-class Empty(object): # pragma: no cover
-    ''' Another empty class to represent a failed object loading, but do not proceed with setstate '''
-
-    def __init__(self, *args, **kwargs):
-        pass
-
-    def __setstate__(self, state):
-        pass
 
 
 class Failed:
@@ -2130,6 +2122,9 @@ class Failed:
     def __setitem__(self, key, value):
         self.dict[key] = value
         return
+    
+    def __call__(self, *args, **kwargs):
+        return self
 
     def disp(self, *args, **kwargs):
         return scp.pr(self, *args, **kwargs)
@@ -2158,7 +2153,7 @@ class Failed:
             return
 
 
-def makefailed(module_name=None, name=None, error=None, exception=None, tb=None):
+def _makefailed(module_name=None, name=None, error=None, exception=None, tb=None):
     '''
     Update a class -- not an object! -- that contains the failure info for a pickle 
     that failed to load. It needs to be a class rather than class instance due to
@@ -2195,7 +2190,7 @@ def _remap_module(remapping, module_name, name):
     return obj
 
 
-class _LoadsInterface():
+class _LoadsInterface:
     ''' Add a .loads method to unpicklers, with support for Sciris remapping '''
     @classmethod
     def loads(cls, string, remapping, **kwargs):
@@ -2203,98 +2198,98 @@ class _LoadsInterface():
         return unpickler.load()
 
 
-class _RobustUnpickler(dill.Unpickler, _LoadsInterface):
+class _RobustUnpickler(dill.Unpickler):
     ''' Try to import an object, and if that fails, return a Failed object rather than crashing '''
 
-    def __init__(self, bytesio, fix_imports=True, encoding="latin1", errors="ignore", remapping=None):
[…]
