+++ sciris/sc_fileio.py
@@ -2095,6 +2095,8 @@ class Failed:
     _module  = None # These must be class variables since find_class returns a class, not an instance
     _name    = None
     _failure = None
+    _fixes   = None # Used when the top-level loads() method fails
+    _errors  = None
 
     def __init__(self, *args, **kwargs):
         if args:
@@ -2155,62 +2157,78 @@ class Failed:
             return
 
 
-def _makefailed(module_name=None, name=None, error=None, exception=None, tb=None):
+def _makefailed(module_name=None, name=None, exc=None, fixes=None, errors=None):
     '''
-    Update a class -- not an object! -- that contains the failure info for a pickle 
+    Handle unpickling failures -- for internal use only
+    
+    Create a class -- not an object! -- that contains the failure info for a pickle 
     that failed to load. It needs to be a class rather than class instance due to
     the way pickles are loaded via the ``find_class`` method.
     
-    *New in version 3.1.0:* "tb" argument; removed "universal" argument
+    *New in version 3.1.0:* arguments simplified
     '''
+    
+    # Process exception
     key = (module_name, name)
     fail = sco.objdict()
     fail.module    = module_name
     fail.name      = name
-    fail.error     = error
-    fail.exception = exception
-    fail.traceback = tb
+    fail.error     = str(exc)
+    fail.exception = exc
+    fail.traceback = scu.traceback(exc)
     
     # Add the failure to the global list
     unpickling_errors[key] = fail
     
+    # Attributes -- this is why this needs to be dynamically created!
+    attrs = dict( 
+         _module = module_name,
+         _name = name,
+         _failure = fail
+    )
+    
+    # If the top-level loads() fails, including additional information on the errors
+    if fixes is not None or errors is not None:
+        attrs.update(dict(_unpickling_fixes=fixes, _unpickling_errors=errors))
+    
     # Dynamically define a new class that can be used by find_class() to create an object
     F = type(
-        "FailedLoad", # Name of the class -- can all be the same
+        "NamedFailed", # Name of the class -- can all be the same, but distinct from Failed that it has populated names
         (Failed,), # Inheritance
-        dict( # Attributes
-             _module = module_name,
-             _name = name,
-             _failure = fail
-        )
+        attrs, # Attributes
     )
-    return F
+    return F # Return the newly created class (not class instance)
 
 
 def _remap_module(remapping, module_name, name):
-    ''' Use a remapping dictionary to try to load a module from a different location ''' 
+    ''' Use a remapping dictionary to try to load a module from a different location -- for internal use ''' 
     key1 = f'{module_name}.{name}' # Key provided as a single string
     key2 = (module_name, name) # Key provided as a tuple
-    key3 = module_name
+    key3 = module_name # Just the module, no class
     remapped = remapping.get(key1) or remapping.get(key2) or remapping.get(key3) # If the user has supplied the module directly
-    if remapped is None:
-        new_module = module_name
-        new_name = name
-    else:
+    if remapped is not None:
+        
+        # Check that we have a string or tuple
         if isinstance(remapped, str): # Split a string into a tuple, e.g. 'foo.bar.Cat' to ('foo.bar', 'Cat')
-            remapped = tuple(remapped.rsplit('.', 1))
-            len_remap = len(remapped)
-            if len_remap == 2: # Usual case
-                new_module = remapped[0]
-                new_name   = remapped[1]
-            elif len_remap == 1:
-                new_module = remapped[0]
-                new_name   = name
-            else:
-                errormsg = f'Was expecting 1 or 2 strings, got {len_remap}'
[…]
