+++ sciris/sc_fileio.py
@@ -135,7 +135,7 @@ def _load_filestr(filename, folder=None, verbose=False):
     return filestr
 
 
-def load(filename=None, folder=None, verbose=False, die=None, remapping=None, 
+def load(filename=None, folder=None, verbose=None, die=None, remapping=None, 
          method=None, auto_remap=True, **kwargs):
     '''
     Load a file that has been saved as a gzipped pickle file, e.g. by :func:`sc.save() <save>`.
@@ -155,11 +155,11 @@ def load(filename=None, folder=None, verbose=False, die=None, remapping=None,
 
     Args:
         filename   (str/Path): the filename (or full path) to load
-        folder     (str/Path): the folder
-        verbose    (bool):     print details
+        folder     (str/Path): the folder (not needed if the filename includes it)
+        verbose    (bool):     print nothing (False), critical warnings (None), or full detail (True)
         die        (bool):     whether to raise an exception if errors are encountered (otherwise, load as much as possible)
-        remapping  (dict):     way of mapping old/unavailable module names to new
-        method     (str):      method for loading ('pickle' or 'dill'; if None, try both)
+        remapping  (dict):     way of mapping old/unavailable module names to new (see below for example)
+        method     (str):      method for loading ('pickle', 'dill', 'pandas', or 'robust'; if None, try all)
         auto_remap (bool):     whether to use known deprecations to load failed pickles
         kwargs     (dict):     passed to ``pickle.loads()``/``dill.loads()``
 
@@ -168,7 +168,7 @@ def load(filename=None, folder=None, verbose=False, die=None, remapping=None,
         obj = sc.load('myfile.obj') # Standard usage
         old = sc.load('my-old-file.obj', method='dill', ignore=True) # Load classes from saved files
         old = sc.load('my-old-file.obj', remapping={'foo.Bar': cat.Mat}) # If loading a saved object containing a reference to foo.Bar that is now cat.Mat
-        old = sc.load('my-old-file.obj', remapping={('foo', 'Bar'): ('cat', 'Mat')}) # Equivalent to the above
+        old = sc.load('my-old-file.obj', remapping={('foo', 'Bar'): ('cat', 'Mat')}, die=False) # Equivalent to the above but don't fail
 
     | *New in version 1.1.0:* "remapping" argument
     | *New in version 1.2.2:* ability to load non-gzipped pickles; support for dill; arguments passed to loader
@@ -2070,7 +2070,8 @@ def savespreadsheet(filename=None, data=None, folder=None, sheetnames=None, clos
 #%% Pickling support methods
 ##############################################################################
 
-__all__ += ['Failed']
+__all__ += ['UnpicklingWarning', 'Failed']
+
 
 # Pandas provides a class compatibility map, so use that by default
 try:
@@ -2085,6 +2086,16 @@ except (NameError or AttributeError):
 
 # Keep a temporary global variable of unpickling errors, and a permanent one of failed classes
 unpickling_errors = sco.objdict()
+
+
+class UnpicklingWarning(UserWarning):
+    '''
+    A warning raised when unpickling an object fails
+    
+    *New in version 3.1.0.*    
+    '''
+    pass
+
 
 class Failed:
     '''
@@ -2208,18 +2219,16 @@ def _remap_module(remapping, module_name, name):
     key3 = module_name # Just the module, no class
     remapped = remapping.get(key1) or remapping.get(key2) or remapping.get(key3) # If the user has supplied the module directly
     
-    do type stuff here
-    if isinstance(remapped, type) # THIS WONT WORK
-    if remapped is not None:
+    # Remapping failed, just use original names
+    if remapped is None: 
+        remapped = (module_name, name)
         
[…]
