### 2023-08-09 003231b
replace pkg_resources with importlib where possible
