### 2023-08-09 7f3ce03
restore pkg_resources for python compatibility
