### 2023-08-10 df57e93
unable to delete class
