+++ sciris/sc_nested.py
@@ -95,7 +95,14 @@ def check_iter_type(obj):
     
 
 def get_from_obj(ndict, key, safe=False):
-    ''' Get an item from a dict, list, or object by key '''
+    '''
+    Get an item from a dict, list, or object by key
+    
+    Args:
+        ndict (dict/list/obj): the object to get from
+        key (any): the key to get
+        safe (bool): whether to return None if the key is not found (default False)
+    '''
     itertype = check_iter_type(ndict)
     if itertype == 'dict':
         if safe:
@@ -111,54 +118,74 @@ def get_from_obj(ndict, key, safe=False):
     return out
 
 
-def getnested(nesteddict, keylist, safe=False):
+def getnested(nested, keylist, safe=False):
     '''
     Get the value for the given list of keys
+    
+    Args:
+        nested (any): the nested object (dict, list, or object) to get from
+        keylist (list): the list of keys
+        safe (bool): whether to return None if the key is not found
+    
+    **Example**::
 
-    >>> sc.getnested(foo, ['a','b'])
+        sc.getnested(foo, ['a','b']) # Gets foo['a']['b']
 
-    See sc.makenested() for full documentation.
+    See :func:`sc.makenested() <makenested>` for full documentation.
     '''
     get = partial(get_from_obj, safe=safe)
-    output = reduce(get, keylist, nesteddict)
-    return output
+    nested = reduce(get, keylist, nested)
+    return nested
 
 
-def setnested(nesteddict, keylist, value, force=True):
+def setnested(nested, keylist, value, force=True):
     '''
     Set the value for the given list of keys
+    
+    Args:
+        nested (any): the nested object (dict, list, or object) to modify
+        keylist (list): the list of keys to use
+        value (any): the value to set
+        force (bool): whether to create the keys if they don't exist (NB: only works for dictionaries)
+    
+    **Example**::
 
-    >>> sc.setnested(foo, ['a','b'], 3)
+        sc.setnested(foo, ['a','b'], 3) # Sets foo['a']['b'] = 3
 
-    See sc.makenested() for full documentation.
+    See :func:`sc.makenested() <makenested>` for full documentation.
     '''
     if force:
-        makenested(nesteddict, keylist, overwrite=False)
-    currentlevel = getnested(nesteddict, keylist[:-1])
+        makenested(nested, keylist, overwrite=False)
+    currentlevel = getnested(nested, keylist[:-1])
     if not isinstance(currentlevel, dict): # pragma: no cover
         errormsg = f'Cannot set {keylist} since parent is a {type(currentlevel)}, not a dict'
         raise TypeError(errormsg)
     else:
         currentlevel[keylist[-1]] = value
-    return # Modify nesteddict in place
+    return nested # Return object, but note that it's modified in place
 
 
-def iternested(nesteddict, previous=None):
+def iternested(nesteddict, _previous=None):
     '''
     Return a list of all the twigs in the current dictionary
+    
+    Args:
+        nesteddict (dict): the dictionary
+    
+    **Example**::
 
-    >>> twigs = sc.iternested(foo)
+        twigs = sc.iternested(foo)
 
-    See sc.makenested() for full documentation.
+    See :func:`sc.makenested() <makenested>` for full documentation.
     '''
-    if previous is None:
-        previous = []
+    if _previous is None:
+        _previous = []
     output = []
     for k in nesteddict.items():
-        if isinstance(k[1],dict):
-            output += iternested(k[1], previous+[k[0]]) # Need to add these at the first level
+        if isinstance(k[1], dict):
+            output += iternested(k[1], _previous+[k[0]]) # Need to add these at the first level
         else:
[…]
