### 2023-08-11 4384c57
working on equal
