+++ sciris/sc_nested.py
@@ -5,11 +5,13 @@ Nested dictionary functions.
 import re
 import itertools
 from functools import reduce, partial
+import numpy as np
+import pandas as pd
 from . import sc_utils as scu
 
 
 __all__ = ['getnested', 'setnested', 'makenested', 'iternested', 'iterobj',
-           'mergenested', 'flattendict', 'search', 'nestedloop']
+           'mergenested', 'flattendict', 'search', 'equal', 'nestedloop']
 
 
 def makenested(nesteddict, keylist=None, value=None, overwrite=False, generator=None):
@@ -81,7 +83,7 @@ def makenested(nesteddict, keylist=None, value=None, overwrite=False, generator=
     return
 
 
-def check_iter_type(obj):
+def check_iter_type(obj, check_array=False):
     ''' Helper function to determine if an object is a dict, list, or neither '''
     if isinstance(obj, dict):
         out = 'dict'
@@ -89,6 +91,8 @@ def check_iter_type(obj):
         out = 'list'
     elif hasattr(obj, '__dict__'):
         out = 'object'
+    elif check_array and isinstance(obj, np.ndarray):
+        out = 'array'
     else:
         out = '' # Evaluates to false
     return out
@@ -189,7 +193,32 @@ def iternested(nesteddict, _previous=None):
     return output
 
 
-def iterobj(obj, func, inplace=False, copy=True, twigs_only=False, verbose=False, 
+def iteritems(obj, itertype):
+    ''' Return an iterator over items in this object -- for internal use '''
+    if itertype == 'dict':
+        return obj.items()
+    elif itertype == 'list':
+        return enumerate(obj)
+    elif itertype == 'object':
+        return obj.__dict__.items()
+
+def getitem(obj, key, itertype):
+    ''' Get the value for the item -- for internal use '''
+    if itertype in ['dict', 'list']:
+        return obj[key]
+    elif itertype == 'object':
+        return obj.__dict__[key]
+
+def setitem(obj, key, value, itertype):
+    ''' Set the value for the item -- for internal use '''
+    if itertype in ['dict', 'list']:
+        obj[key] = value
+    elif itertype == 'object':
+        obj.__dict__[key] = value
+    return
+
+
+def iterobj(obj, func=None, inplace=False, copy=True, twigs_only=False, verbose=False, 
             _trace=None, _output=None, *args, **kwargs):
     '''
     Iterate over an object and apply a function to each twig.
@@ -199,7 +228,7 @@ def iterobj(obj, func, inplace=False, copy=True, twigs_only=False, verbose=False
     
     Args:
         obj (any): the object to iterate over
-        func (function): the function to apply
+        func (function): the function to apply; if None, return a dictionary of all twigs in the object
         inplace (bool): whether to modify the object in place (else, collate the output of the functions)
         copy (bool): if modifying an object in place, whether to make a copy first
         twigs_only (bool): whether to apply the function only to twigs of the object
@@ -233,8 +262,12 @@ def iterobj(obj, func, inplace=False, copy=True, twigs_only=False, verbose=False
         data = sc.iterobj(data, collapse, inplace=True)
         sc.printjson(data)
         
-    *New in version 3.0.0.*
+    | *New in version 3.0.0.*
+    | *New in version 3.1.0:* default ``func``, improved ``twigs_only``
     '''
+    if func is None:
+        func = lambda obj: obj
+        
     # Set the trace and output if needed
     if _trace is None:
         _trace = []
@@ -247,42 +280,21 @@ def iterobj(obj, func, inplace=False, copy=True, twigs_only=False, verbose=False
     
     itertype = check_iter_type(obj)
     
-    def iteritems(obj):
-        ''' Return an iterator over items in this object '''
-        if itertype == 'dict':
-            return obj.items()
[…]
