### 2023-08-11 5d4e476
working on equality class
