+++ sciris/sc_nested.py
@@ -389,7 +389,7 @@ def flattendict(nesteddict, sep=None, _prefix=None):
 
 
 # Define a custom "None" value to allow searching for actual None values
-_None = 'sc.search() placeholder'
+_None = '<sc_nested_custom_None>' # This should not be equal to any other value the user could supply
 def search(obj, query=_None, key=_None, value=_None, aslist=True, method='exact', 
            return_values=False, verbose=False, _trace=None):
     """
@@ -556,7 +556,7 @@ class Equality(scu.prettyobj):
     special_cases = [np.ndarray, pd.DataFrame]
     
     def __init__(self, obj, obj2, *args, method='json', detailed=False, 
-                 verbose=False, die=True, _trace=None):
+                 equal_nan=True, verbose=None, die=True):
         '''
         Compare equality between two arbitrary objects -- see :func:`sc.equal() <equal>` for full documentation.
 
@@ -574,16 +574,23 @@ class Equality(scu.prettyobj):
         self.others = [obj2] + list(args)
         self.b = None # Base object after conversion
         self.o = [] # Other objects after conversion
+        self.bd = None # Base dictionary
+        self.od = [] # Other dictionaries
         self.method = method
         self.detailed = detailed
+        self.equal_nan = equal_nan
         self.verbose = verbose
         self.die = die
         self.eq = None # Final value to be populated
-        self.eqdict = sco.objdict() # Detailed output
+        self.results = sco.objdict() # Detailed output, 1D dict
+        self.fullresults = sco.objdict() # Detailed output, 2D dict
+        self.exceptions = sco.objdict()
 
         # Perform the comparison
         self.convert() # Convert the objects
         self.compare() # Do the comparison
+        if self.verbose is not False:
+            self.check_exceptions() # Check if any exceptions were encountered
         return
     
     @property
@@ -606,73 +613,100 @@ class Equality(scu.prettyobj):
         
         # Do the mapping
         self.b = func(self.base)
+        self.bd = iterobj(self.b)
         for other in self.others:
-            self.o.append(func(other))
+            o = func(other)
+            self.o.append(o)
+            self.od.append(iterobj(o))
                 
         return
-            
+    
+
+    def compare_special(self, obj, obj2):
+        ''' Do special comparisons for known objects where == doesn't work '''
+        from . import sc_dataframe as scd # To avoid circular import
+        
+        # For numpy arrays, must use np.array_equals()
+        if isinstance(obj, np.ndarray):
+            eq = np.array_equal(obj, obj2, equal_nan=self.equal_nan)
+    
+        # For dataframes, convert to Sciris to leverage additional comparators
+        elif isinstance(obj, pd.DataFrame):
+            eq = scd.dataframe(obj) == scd.dataframe(obj2)
+        
+        else:
+            errormsg = f'Not able to handle object of {type(obj)}'
+            raise TypeError(errormsg)
+        
+        return eq
+    
+    
+    @staticmethod
+    def keytostr(k):
+        ''' Helper method to convert a key to a "trace" for printing '''
+        return scu.strjoin(k, sep='.')
+        
     
     def compare(self):
         ''' Perform the comparison '''
         
-        for k,v in base.items():
+        for i,k,v in self.b.enumitems():
+            if self.verbose:
+                print(f'Working on item {i+1}/{len(self.b)}: {k}')
             eqs = []
-            for o,other in enumerate(others):
-                try:
-                    ov = getnested(other, k)
-                    eq = ov == v
-                except Exception as E:
-                    if die:
-                        raise E
-                    else:
+            for j,od in enumerate(self.od):
+                
+                # If not present, false by default
+                if k not in od:
+                    eq = False
+                else:
+                    ov = od[k]
+                
+                # Check equality
+                eq = None
[…]
