### 2023-08-11 283f119
working on equal
