+++ sciris/sc_nested.py
@@ -1,5 +1,10 @@
 '''
-Nested dictionary functions.
+Functions for working on nested (multi-level) dictionaries and objects.
+
+Highlights:
+    - :func:`sc.getnested() <getnested>`: get a value from a highly nested dictionary
+    - :func:`sc.search() <search>`: find a value in a nested object
+    - :func:`sc.equal() <equal>`: check complex objects for equality
 '''
 
 import re
@@ -11,8 +16,13 @@ import pandas as pd
 from . import sc_utils as scu
 
 
+##############################################################################
+#%% Nested dict and object functions
+##############################################################################
+
+
 __all__ = ['getnested', 'setnested', 'makenested', 'iternested', 'iterobj',
-           'mergenested', 'flattendict', 'search', 'equal', 'nestedloop']
+           'mergenested', 'flattendict', 'nestedloop']
 
 
 def makenested(nesteddict, keylist=None, value=None, overwrite=False, generator=None):
@@ -266,6 +276,8 @@ def iterobj(obj, func=None, inplace=False, copy=True, twigs_only=False, verbose=
     | *New in version 3.0.0.*
     | *New in version 3.1.0:* default ``func``, improved ``twigs_only``
     '''
+    from . import sc_odict as sco # To avoid circular import
+    
     if func is None:
         func = lambda obj: obj
         
@@ -275,7 +287,7 @@ def iterobj(obj, func=None, inplace=False, copy=True, twigs_only=False, verbose=
         if inplace and copy: # Only need to copy once
             obj = scu.dcp(obj)
     if _output is None:
-        _output = {}
+        _output = sco.objdict()
         if not inplace and not twigs_only:
             _output['root'] = func(obj, *args, **kwargs)
     
@@ -387,6 +399,59 @@ def flattendict(nesteddict, sep=None, _prefix=None):
 
     return output_dict
 
+
+
+def nestedloop(inputs, loop_order):
+    """
+    Zip list of lists in order
+
+    This function takes in a list of lists to iterate over, and their nesting order.
+    It then yields tuples of items in the given order. Only tested for two levels
+    but in theory supports an arbitrary number of items.
+
+    Args:
+        inputs (list): List of lists. All lists should have the same length
+        loop_order (list): Nesting order for the lists
+
+    Returns:
+        Generator yielding tuples of items, one for each list
+
+    Example usage:
+
+    >>> list(sc.nestedloop([['a','b'],[1,2]],[0,1]))
+    [['a', 1], ['a', 2], ['b', 1], ['b', 2]]
+
+    Notice how the first two items have the same value for the first list
+    while the items from the second list vary. If the `loop_order` is
+    reversed, then:
+
+    >>> list(sc.nestedloop([['a','b'],[1,2]],[1,0]))
+    [['a', 1], ['b', 1], ['a', 2], ['b', 2]]
+
+    Notice now how now the first two items have different values from the
+    first list but the same items from the second list.
+
+    From Atomica by Romesh Abeysuriya.
+
+    *New in version 1.0.0.*
+    """
+    loop_order = list(loop_order)  # Convert to list, in case loop order was passed in as a generator e.g. from map()
+    inputs = [inputs[i] for i in loop_order]
+    iterator = itertools.product(*inputs)  # This is in the loop order
+    for item in iterator:
+        out = [None] * len(loop_order)
+        for i in range(len(item)):
+            out[loop_order[i]] = item[i]
+        yield out
+
+
+
+##############################################################################
+#%% Search and equality operators
+##############################################################################
+
+
+__all__ += ['search', 'Equality', 'equal']
 
 # Define a custom "None" value to allow searching for actual None values
[…]
