### 2023-08-11 5280cbf
updating docs on equal
