### 2023-08-12 bf9e346
working on dataframe and nan equality
