### 2023-08-14 #5076 @cliffckerr
@danielskatz At long last!!
https://pypi.org/project/sciris/3.1.0/
https://zenodo.org/record/8245253

I assume I should update the manuscript as well? (Since it refers to v3.0.0 instead of v3.1.0.)
