### 2023-08-14 #5076 @danielskatz
@cliffckerr - Please add the other authors from the paper to the zenodo metadata for the archive, and preferably, add ORCIDs for everyone.

I don't see the version in the body of the manuscript, so I'll generate a proof copy now, which I will then proofread and get back to you with any needed changes.
