### 2023-08-14 #5076 @danielskatz
Congratulations to @cliffckerr (Cliff Kerr) and co-authors on your publication!!

And thanks to @tacaswell and @aflaxman for reviewing!
JOSS is fully volunteer-run and only works because of your willingness to participate
