### 2023-08-23 #5076 @cliffckerr
Hi @danielskatz, I'm wondering if it's possible to make a correction. I missed that @mariadelmarq's name had been parsed incorrectly in the XML and "jats" files. It should be
```
            <given_name>Maria del Mar</given_name>
            <surname>Quiroga</surname>
```
not
```
            <given_name>Maria</given_name>
            <surname>del Mar Quiroga</surname>
```

Is there any way to make this change to the metadata? The PDF itself is fine. I know the metadata has already propagated to Google Scholar etc, so understand it might be hard to put the genie back in the bottle!
