### 2023-08-23 #5076 @danielskatz
@xuanxu - I wonder if we should attempt to make the bibtex work better as well, by changing the author string to be "last, first" for each author rather than the current "first last". Otherwise in cases like this, the bibtex loses some information.
