### 2025-06-04 #5076 @cliffckerr
Hi @xyluo25 -- I would love to and I agree your paper looks very similar to Sciris, but unfortunately I am going on an extended period of leave soon and will not have time before that -- good luck! If you are still looking for reviewers in 2026 then I can help out 😁
